### 2023-11-09 #6045 @editorialbot
[PRE REVIEW]: The MOOSE Thermal Hydraulics Module
**Submitting author:** @joshuahansel (<a href="http://orcid.org/0000-0001-6782-5275">Joshua Hansel</a>)
**Repository:** https://github.com/idaholab/moose
**Branch with paper.md** (empty if default branch): thm-joss-paper-review
**Version:** v1.0.0
