### 2023-11-09 #6045 @kyleniemeyer
@openjournals/dev how can we handle a paper source that is in a specific subdirectory (https://github.com/idaholab/moose/tree/next/modules/thermal_hydraulics), where other module subdirectories may have their own file? Would it be easiest if the paper was on a dedicated branch?
