### 2023-11-09 #6045 @kyleniemeyer
Hi @joshuahansel, I'm going to edit your submission. It looks like we need to work out how to handle the paper source/compilation.

Also, just to document here, the module in question for review is located at https://github.com/idaholab/moose/tree/next/modules/thermal_hydraulics .
