### 2023-11-09 a105ae5
Address test failures: - initialize pointer - modify bad test that ends up enforcing the list of systems - add doco for action
