### 2023-11-09 058a85d
Use std::optional for _bx_norm_name
