### 2023-11-09 e993985
Remove what I believe should be unnecessary PetscObjectCompose calls
