+++ framework/src/linearfvkernels/LinearFVSourceKernel.C
@@ -11,6 +11,8 @@
 #include "Assembly.h"
 #include "SubProblem.h"
 
+registerMooseObject("MooseApp", LinearFVSourceKernel);
+
 InputParameters
 LinearFVSourceKernel::validParams()
 {
