### 2023-11-09 1d9fb87
Added 'use_shell' param to Tester
This parameter controls whether subprocess.Popen launches each test using the
shell. The default value is False. For tests where the shell is needed, the
shell may be turned on for each specific test.
