### 2023-11-09 2c1c69a
Remove bibliography duplicates and warnings for combined and navier stokes
Refs #26015
