+++ modules/combined/doc/content/bib/combined.bib
@@ -1,52 +1,0 @@
-@article{sigmund200199,
-title={A 99 line topology optimization code written in Matlab},
-author={Sigmund, Ole},
-journal={Structural and multidisciplinary optimization},
-volume={21},
-pages={120--127},
-year={2001},
-publisher={Springer}
-}
-@book{bendsoe2003topology,
-  title={Topology optimization: theory, methods, and applications},
-  author={Bendsoe, Martin Philip and Sigmund, Ole},
-  year={2003},
-  publisher={Springer Science \& Business Media}
-}
-@article{tavakoli2014alternating,
-  title={Alternating active-phase algorithm for multimaterial topology optimization problems: a 115-line MATLAB implementation},
-  author={Tavakoli, Rouhollah and Mohseni, Seyyed Mohammad},
-  journal={Structural and Multidisciplinary Optimization},
-  volume={49},
-  pages={621--642},
-  year={2014},
-  publisher={Springer}
-}
-@article{sigmund2007,
-author = {Sigmund, Ole},
-year = {2007},
-month = {04},
-pages = {401-424},
-title = {Morphology-based black and white filters for topology optimization},
-volume = {33},
-journal = {Structural and Multidisciplinary Optimization},
-doi = {10.1007/s00158-006-0087-x}
-}
-@Article{pde_filter,
-author = {Wallin, Mathias and Ivarsson, Niklas and Amir, Oded and Tortorelli, Daniel},
-title = {Consistent Boundary Conditions for PDE Filter Regularization in Topology Optimization},
-year = {2020},
-issue_date = {Sep 2020},
-publisher = {Springer-Verlag},
-address = {Berlin, Heidelberg},
-volume = {62},
-number = {3},
-issn = {1615-147X},
-url = {https://doi.org/10.1007/s00158-020-02556-w},
-doi = {10.1007/s00158-020-02556-w},
-journal = {Struct. Multidiscip. Optim.},
-month = {sep},
-pages = {1299–1311},
-numpages = {13},
-keywords = {PDE filter, Boundary effects, Topology optimization}
-}
