+++ modules/tensor_mechanics/doc/content/bib/tensor_mechanics.bib
@@ -940,17 +940,6 @@
   publisher = {Elsevier}
 }
 
-@article{walters_interaction_2005,
-  author  = {Walters, Matthew C. and Paulino, Glaucio H. and Dodds, Robert H.},
-  journal = {Engineering Fracture Mechanics},
-  month   = {July},
-  number  = {11},
-  pages   = {1635--1663},
-  title   = {Interaction integral procedures for 3-{D} curved cracks including surface tractions},
-  volume  = {72},
-  year    = {2005}
-}
-
 @article{Jeong2004functionallygraded,
     author = {Kim, Jeong-Ho and Paulino, Glaucio H.},
     title = {Consistent Formulations of the Interaction Integral Method for Fracture of Functionally Graded Materials},
