### 2023-11-09 2320a1a
Tests for weights and values are properly multiplied for duplicates points.  closes #26014
