+++ modules/ray_tracing/src/raytracing/TraceRay.C
@@ -1602,6 +1602,9 @@ TraceRay::trace(const std::shared_ptr<Ray> & ray)
 void
 TraceRay::onCompleteTrace(const std::shared_ptr<Ray> & ray)
 {
+  for (RayKernelBase * rk : _study.currentRayKernels(_tid))
+    rk->postTrace();
+
   debugRay("Called onCompleteTrace()\n", (*_current_ray)->getInfo());
   if (_intersection_distance > 0)
     possiblyAddDebugRayMeshPoint(_incoming_point, _intersection_point);
