### 2023-11-10 5ca111b
Restore MUMPS_ARM_STR to configure_petsc.sh for manual Apple Silicon builds
Removing this has been causing failures in CI. Also, increment conda packages.

Refs #25287
