+++ conda/moose-dev/conda_build_config.yaml
@@ -1,5 +1,5 @@
 moose_libmesh:
-  - moose-libmesh 2023.11.06 build_0
+  - moose-libmesh 2023.11.06 build_1
 
 moose_wasp:
   - moose-wasp 2023.11.06
