+++ conda/libmesh/conda_build_config.yaml
@@ -1,5 +1,5 @@
 moose_petsc:
-  - moose-petsc 3.20.0 build_1
+  - moose-petsc 3.20.0 build_2
 
 moose_libmesh_vtk:
   - moose-libmesh-vtk 9.2.6 build_6
