+++ scripts/tests/versioner_hashes.yaml
@@ -159,3 +159,9 @@ d5e3e09267e1f2b47684a089150bec587178e1cf: #25784
   libmesh: 9ed7d66
   moose: c223a73
   wasp: 3740200
+5ca111bb7d90ee3952ee1328ab679104fed41b67: #26023
+  mpich: e018192
+  petsc: 51cfee0
+  libmesh: 9615609
+  moose: 1ad71e8
+  wasp: 3740200
