### 2023-11-10 0228a74
Used 3D flux computation in volume junction
Refs #25958
