+++ modules/thermal_hydraulics/src/userobjects/ADVolumeJunction1PhaseUserObject.C
@@ -13,6 +13,7 @@
 #include "VolumeJunction1Phase.h"
 #include "ADNumericalFlux3EqnBase.h"
 #include "Numerics.h"
+#include "THMUtils.h"
 #include "metaphysicl/parallel_numberarray.h"
 #include "metaphysicl/parallel_dualnumber.h"
 #include "metaphysicl/parallel_semidynamicsparsenumberarray.h"
@@ -89,15 +90,18 @@ void
 ADVolumeJunction1PhaseUserObject::computeFluxesAndResiduals(const unsigned int & c)
 {
   const Real din = _normal[c];
-  const Point di = _dir[0];
-  const Point ni = di * din;
+  const RealVectorValue di = _dir[0];
+  const RealVectorValue ni = di * din;
+  const RealVectorValue nJi = -ni;
   const Real nJi_dot_di = -din;
 
-  std::vector<ADReal> Ui(THMVACE1D::N_FLUX_INPUTS, 0.);
-  Ui[THMVACE1D::RHOA] = _rhoA[0];
-  Ui[THMVACE1D::RHOUA] = _rhouA[0];
-  Ui[THMVACE1D::RHOEA] = _rhoEA[0];
-  Ui[THMVACE1D::AREA] = _A[0];
+  std::vector<ADReal> Ui(THMVACE3D::N_FLUX_INPUTS, 0.);
+  Ui[THMVACE3D::RHOA] = _rhoA[0];
+  Ui[THMVACE3D::RHOUA] = _rhouA[0] * di(0);
+  Ui[THMVACE3D::RHOVA] = _rhouA[0] * di(1);
+  Ui[THMVACE3D::RHOWA] = _rhouA[0] * di(2);
+  Ui[THMVACE3D::RHOEA] = _rhoEA[0];
+  Ui[THMVACE3D::AREA] = _A[0];
 
   const auto & rhoV = _cached_junction_var_values[VolumeJunction1Phase::RHOV_INDEX];
   const auto & rhouV = _cached_junction_var_values[VolumeJunction1Phase::RHOUV_INDEX];
@@ -105,17 +109,27 @@ ADVolumeJunction1PhaseUserObject::computeFluxesAndResiduals(const unsigned int &
   const auto & rhowV = _cached_junction_var_values[VolumeJunction1Phase::RHOWV_INDEX];
   const auto & rhoEV = _cached_junction_var_values[VolumeJunction1Phase::RHOEV_INDEX];
 
-  std::vector<ADReal> UJi(THMVACE1D::N_FLUX_INPUTS, 0.);
-  const ADRealVectorValue rhouV_vec(rhouV, rhovV, rhowV);
-  UJi[THMVACE1D::RHOA] = rhoV / _volume * _A[0];
-  UJi[THMVACE1D::RHOUA] = rhouV_vec * di / _volume * _A[0];
-  UJi[THMVACE1D::RHOEA] = rhoEV / _volume * _A[0];
-  UJi[THMVACE1D::AREA] = _A[0];
+  std::vector<ADReal> UJi(THMVACE3D::N_FLUX_INPUTS, 0.);
+  UJi[THMVACE3D::RHOA] = rhoV / _volume * _A[0];
+  UJi[THMVACE3D::RHOUA] = rhouV / _volume * _A[0];
+  UJi[THMVACE3D::RHOVA] = rhovV / _volume * _A[0];
+  UJi[THMVACE3D::RHOWA] = rhowV / _volume * _A[0];
+  UJi[THMVACE3D::RHOEA] = rhoEV / _volume * _A[0];
+  UJi[THMVACE3D::AREA] = _A[0];
 
-  _flux[c] =
[…]
