### 2023-11-10 427900a
Add base class for linear flux kernels. (#25722)
