### 2023-11-10 47dfe7b
Neighbor material in node-face explicit dynamics constraint
