### 2023-11-10 8695b3f
Add first version of linear diffusion kernel. (#25722)
