+++ framework/include/linearfvkernel/LinearFVDiffusionKernel.h
@@ -1,0 +1,30 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "FluxLinearFVKernel.h"
+
+class LinearFVDiffusionKernel : public FluxLinearFVKernel
+{
+public:
+  static InputParameters validParams();
+  LinearFVDiffusionKernel(const InputParameters & params);
+
+  virtual Real computeElemMatrixContribution() override;
+
+  virtual Real computeNeighborMatrixContribution() override;
+
+  virtual Real computeElemRightHandSideContribution() override;
+
+  virtual Real computeNeighborRightHandSideContribution() override;
+
+protected:
+  const Real _diffusion_coeff;
+};
