### 2023-11-10 f760e32
Fix memory leak for lower-D vector FEs
