### 2023-11-10 9dac522
Disable p-refinement of continuous non-hierarchic bases by default
And Nedelec as well, and document this.
