+++ framework/src/loops/ComputeLinearFVFaceThread.C
@@ -7,50 +7,50 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "ComputeLinearFVElementalThread.h"
+#include "ComputeLinearFVFaceThread.h"
 #include "LinearSystem.h"
 #include "LinearFVKernel.h"
-#include "ElementalLinearFVKernel.h"
+#include "FluxLinearFVKernel.h"
 
-ComputeLinearFVElementalThread::ComputeLinearFVElementalThread(FEProblemBase & fe_problem,
-                                                               const unsigned int linear_system_num,
-                                                               const std::set<TagID> & tags)
+ComputeLinearFVFaceThread::ComputeLinearFVFaceThread(FEProblemBase & fe_problem,
+                                                     const unsigned int linear_system_num,
+                                                     const std::set<TagID> & tags)
   : _fe_problem(fe_problem), _linear_system_number(linear_system_num), _tags(tags)
 {
 }
 
 // Splitting Constructor
-ComputeLinearFVElementalThread::ComputeLinearFVElementalThread(ComputeLinearFVElementalThread & x,
-                                                               Threads::split /*split*/)
+ComputeLinearFVFaceThread::ComputeLinearFVFaceThread(ComputeLinearFVFaceThread & x,
+                                                     Threads::split /*split*/)
   : _fe_problem(x._fe_problem), _linear_system_number(x._linear_system_number), _tags(x._tags)
 
 {
 }
 
 void
-ComputeLinearFVElementalThread::operator()(const ElemInfoRange & range)
+ComputeLinearFVFaceThread::operator()(const FaceInfoRange & range)
 {
   ParallelUniqueId puid;
   _tid = puid.id;
 
-  std::vector<ElementalLinearFVKernel *> kernels;
+  std::vector<FluxLinearFVKernel *> kernels;
   _fe_problem.theWarehouse()
       .query()
       .template condition<AttribSysNum>(_fe_problem.getLinearSystem(_linear_system_number).number())
-      .template condition<AttribSystem>("ElementalLinearFVKernel")
+      .template condition<AttribSystem>("FluxLinearFVKernel")
       .queryInto(kernels);
 
   // Iterate over all the elements in the range
-  for (const auto & elem_info : range)
+  for (const auto & face_info : range)
     for (auto kernel : kernels)
     {
-      kernel->setCurrentElemInfo(elem_info);
+      kernel->setCurrentFaceInfo(face_info);
       kernel->addMatrixContribution();
       kernel->addRightHandSideContribution();
     }
 }
 
 void
-ComputeLinearFVElementalThread::join(const ComputeLinearFVElementalThread & /*y*/)
+ComputeLinearFVFaceThread::join(const ComputeLinearFVFaceThread & /*y*/)
 {
 }
