+++ modules/tensor_mechanics/test/src/materials/ProjectedStatefulPropertiesTestMaterial.C
@@ -1,0 +1,102 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "ProjectedStatefulPropertiesTestMaterial.h"
+
+registerMooseObject("TensorMechanicsTestApp", ProjectedStatefulPropertiesTestMaterial);
+
+template <bool is_ad>
+InputParameters
+ProjectedStatefulPropertiesTestMaterialTempl<is_ad>::validParams()
+{
+  InputParameters params = Material::validParams();
+  params.addClassDescription("");
+  return params;
+}
+
+template <bool is_ad>
+ProjectedStatefulPropertiesTestMaterialTempl<is_ad>::ProjectedStatefulPropertiesTestMaterialTempl(
+    const InputParameters & parameters)
+  : Material(parameters),
+    InterpolatedStatefulMaterialPropertyInterface(this),
+    // Real
+    _prop_real(declareGenericProperty<Real, is_ad>("test_real")),
+    _prop_real_old(getMaterialPropertyOldByName<Real>("test_real")),
+    _prop_real_old_interpolated(getInterpolatedMaterialPropertyOldByName<Real>("test_real")),
+    // RealVectorValue
+    _prop_realvectorvalue(declareGenericProperty<RealVectorValue, is_ad>("test_realvectorvalue")),
+    _prop_realvectorvalue_old(
+        getMaterialPropertyOldByName<RealVectorValue>("test_realvectorvalue")),
+       _prop_realvectorvalue_old_interpolated(
+         getInterpolatedMaterialPropertyOldByName<RealVectorValue>("test_realvectorvalue")),
+    // RankTwoTensor
+    _prop_ranktwotensor(declareGenericProperty<RankTwoTensor, is_ad>("test_ranktwotensor")),
+    _prop_ranktwotensor_old(getMaterialPropertyOldByName<RankTwoTensor>("test_ranktwotensor")),
+    _prop_ranktwotensor_old_interpolated(
+        getInterpolatedMaterialPropertyOldByName<RankTwoTensor>("test_ranktwotensor")),
+    // RankFourTensor
+    _prop_rankfourtensor(declareGenericProperty<RankFourTensor, is_ad>("test_rankfourtensor")),
+    _prop_rankfourtensor_old(getMaterialPropertyOldByName<RankFourTensor>("test_rankfourtensor")),
+    _prop_rankfourtensor_old_interpolated(
+        getInterpolatedMaterialPropertyOldByName<RankFourTensor>("test_rankfourtensor")),
+    // diagnostic outputs
+    _diff_norm(declareProperty<Real>("diff_norm"))
+{
+}
+
+template <bool is_ad>
+void
+ProjectedStatefulPropertiesTestMaterialTempl<is_ad>::initQpStatefulProperties()
+{
+  _prop_real[_qp] = 77;
+  _prop_realvectorvalue[_qp] = {2.4, 5.6, 7.8};
+  _prop_ranktwotensor[_qp] = {1.1, 1.2, 1.3, 2.1, 2.2, 2.3, 3.1, 3.2, 3.3};
+}
+
+template <bool is_ad>
+void
+ProjectedStatefulPropertiesTestMaterialTempl<is_ad>::computeQpProperties()
+{
+  const auto & p = _q_point[_qp];
+
+  // update current properties
[…]
