### 2023-11-12 cf406b8
allow applications to programmatically set bx postprocessor and switch the sign off when bx postprocessor is used #20095
