### 2023-11-13 4c4b7b6
Added a velocity scaling option
Refs #25958
