+++ modules/thermal_hydraulics/src/userobjects/ADVolumeJunction1PhaseUserObject.C
@@ -42,6 +42,11 @@ ADVolumeJunction1PhaseUserObject::validParams()
   params.addRequiredParam<Real>("K", "Form loss coefficient [-]");
   params.addRequiredParam<Real>("A_ref", "Reference area [m^2]");
 
+  params.addParam<bool>("apply_velocity_scaling",
+                        false,
+                        "Set to true to apply the scaling to the normal velocity. See "
+                        "documentation for more information.");
+
   params.addClassDescription(
       "Computes and caches flux and residual vectors for a 1-phase volume junction");
 
@@ -59,6 +64,8 @@ ADVolumeJunction1PhaseUserObject::ADVolumeJunction1PhaseUserObject(const InputPa
 
     _K(getParam<Real>("K")),
     _A_ref(getParam<Real>("A_ref")),
+
+    _apply_velocity_scaling(getParam<bool>("apply_velocity_scaling")),
 
     _fp(getUserObject<SinglePhaseFluidProperties>("fp"))
 {
@@ -95,6 +102,9 @@ ADVolumeJunction1PhaseUserObject::computeFluxesAndResiduals(const unsigned int &
   const RealVectorValue nJi = -ni;
   const Real nJi_dot_di = -din;
 
+  RealVectorValue t1, t2;
+  THM::computeOrthogonalDirections(nJi, t1, t2);
+
   std::vector<ADReal> Ui(THMVACE3D::N_FLUX_INPUTS, 0.);
   Ui[THMVACE3D::RHOA] = _rhoA[0];
   Ui[THMVACE3D::RHOUA] = _rhouA[0] * di(0);
@@ -109,16 +119,48 @@ ADVolumeJunction1PhaseUserObject::computeFluxesAndResiduals(const unsigned int &
   const auto & rhowV = _cached_junction_var_values[VolumeJunction1Phase::RHOWV_INDEX];
   const auto & rhoEV = _cached_junction_var_values[VolumeJunction1Phase::RHOEV_INDEX];
 
+  const ADReal rhoJ = rhoV / _volume;
+  const ADReal vJ = 1.0 / rhoJ;
+  const ADRealVectorValue uvecJ(rhouV / rhoV, rhovV / rhoV, rhowV / rhoV);
+  const ADReal eJ = rhoEV / rhoV - 0.5 * uvecJ * uvecJ;
+  const ADReal pJ = _fp.p_from_v_e(vJ, eJ);
+
   std::vector<ADReal> UJi(THMVACE3D::N_FLUX_INPUTS, 0.);
   UJi[THMVACE3D::RHOA] = rhoV / _volume * _A[0];
-  UJi[THMVACE3D::RHOUA] = rhouV / _volume * _A[0];
-  UJi[THMVACE3D::RHOVA] = rhovV / _volume * _A[0];
-  UJi[THMVACE3D::RHOWA] = rhowV / _volume * _A[0];
-  UJi[THMVACE3D::RHOEA] = rhoEV / _volume * _A[0];
-  UJi[THMVACE3D::AREA] = _A[0];
+  if (_apply_velocity_scaling)
+  {
+    const ADReal unJ = uvecJ * nJi;
+    const ADReal ut1J = uvecJ * t1;
+    const ADReal ut2J = uvecJ * t2;
+    const ADReal uni = _rhouA[0] / _rhoA[0] * nJi_dot_di;
 
-  RealVectorValue t1, t2;
[…]
