+++ modules/thermal_hydraulics/src/components/JunctionParallelChannels1Phase.C
@@ -58,6 +58,7 @@ JunctionParallelChannels1Phase::buildVolumeJunctionUserObject()
     params.set<Real>("K") = _K;
     params.set<Real>("A_ref") = _A_ref;
     params.set<UserObjectName>("fp") = _fp_name;
+    params.set<bool>("apply_velocity_scaling") = getParam<bool>("apply_velocity_scaling");
     params.set<ExecFlagEnum>("execute_on") = execute_on;
     getTHMProblem().addUserObject(class_name, _junction_uo_name, params);
     connectObject(params, _junction_uo_name, "K");
