### 2023-11-13 2d23e9b
First pass at bypassing mesh generation in RGMB
