+++ modules/reactor/include/meshgenerators/ReactorMeshParams.h
@@ -34,6 +34,9 @@ protected:
   /// The the flat-to-flat size of assemblies in the reactor.
   const Real _assembly_pitch;
 
+  /// Whether or not to bypass mesh generation step in RGMB calls
+  const bool _bypass_meshgen;
+
   ///The heights of the axial regions.
   std::vector<Real> _axial_regions;
 
