+++ modules/reactor/src/meshgenerators/AssemblyMeshGenerator.C
@@ -106,8 +106,6 @@ AssemblyMeshGenerator::AssemblyMeshGenerator(const InputParameters & parameters)
                          : std::vector<std::vector<subdomain_id_type>>()),
     _extrude(getParam<bool>("extrude"))
 {
-  declareMeshesForSub("inputs");
-
   MeshGeneratorName reactor_params =
       MeshGeneratorName(getMeshProperty<std::string>(RGMB::reactor_params_name, _inputs[0]));
   // Check that MG name for reactor params is consistent across all assemblies
@@ -118,6 +116,13 @@ AssemblyMeshGenerator::AssemblyMeshGenerator(const InputParameters & parameters)
 
   // Initialize ReactorMeshParams object stored in pin input
   initializeReactorMeshParams(reactor_params);
+
+  // Declare dependency of inputs to sub generator calls. If mesh generation
+  // should be bypassed, then store the input meshes to free later
+  if (!getReactorParam<bool>(RGMB::bypass_meshgen))
+    declareMeshesForSub("inputs");
+  else
+    _mesh_ptrs = getMeshes("inputs");
 
   _geom_type = getReactorParam<std::string>(RGMB::mesh_geometry);
   _mesh_dimensions = getReactorParam<int>(RGMB::mesh_dimensions);
@@ -311,16 +316,20 @@ AssemblyMeshGenerator::AssemblyMeshGenerator(const InputParameters & parameters)
     params.set<boundary_id_type>("external_boundary_id") = _assembly_boundary_id;
     params.set<std::string>("external_boundary_name") = _assembly_boundary_name;
 
-    addMeshSubgenerator(patterned_mg_name, name() + "_pattern", params);
+    callMeshSubgenerator(patterned_mg_name, name() + "_pattern", params);
 
     // Pass mesh meta-data defined in subgenerator constructor to this MeshGenerator
-    copyMeshProperty<bool>("is_control_drum_meta", name() + "_pattern");
-    copyMeshProperty<std::vector<Point>>("control_drum_positions", name() + "_pattern");
-    copyMeshProperty<std::vector<Real>>("control_drum_angles", name() + "_pattern");
-    copyMeshProperty<std::vector<std::vector<Real>>>("control_drums_azimuthal_meta",
-                                                     name() + "_pattern");
-    copyMeshProperty<std::string>("position_file_name", name() + "_pattern");
-    copyMeshProperty<Real>("pattern_pitch_meta", name() + "_pattern");
+    // Only relevant if we are generating RGMB mesh and not bypassing mesh generation
+    if (!getReactorParam<bool>(RGMB::bypass_meshgen))
+    {
+      copyMeshProperty<bool>("is_control_drum_meta", name() + "_pattern");
+      copyMeshProperty<std::vector<Point>>("control_drum_positions", name() + "_pattern");
+      copyMeshProperty<std::vector<Real>>("control_drum_angles", name() + "_pattern");
+      copyMeshProperty<std::vector<std::vector<Real>>>("control_drums_azimuthal_meta",
+                                                       name() + "_pattern");
+      copyMeshProperty<std::string>("position_file_name", name() + "_pattern");
+      copyMeshProperty<Real>("pattern_pitch_meta", name() + "_pattern");
+    }
   }
 
   std::string build_mesh_name = name() + "_delbds";
@@ -345,7 +354,7 @@ AssemblyMeshGenerator::AssemblyMeshGenerator(const InputParameters & parameters)
     params.set<MeshGeneratorName>("input") = name() + "_pattern";
[…]
