### 2023-11-13 0b9a667
The method "writableVariable" is now callable from a VectorAuxKernel and an ArrayAuxKernel.
closes #26039
