### 2023-11-13 835e404
Add necessary material properties
