+++ modules/navier_stokes/src/fvbcs/INSFVMixingLengthTKEDBC.C
@@ -36,6 +36,7 @@ INSFVMixingLengthTKEDBC::boundaryValue(const FaceInfo & fi) const
   const auto boundary_face = singleSidedFaceArg(&fi);
   const auto state = determineState();
 
-  return std::pow(_C_mu(boundary_face, state), 0.75) * std::pow(_k(boundary_face, state), 1.5) /
+  return std::pow(_C_mu(boundary_face, state), 0.75) *
+         std::pow(_k(boundary_face, state), 1.5) /
          (0.07 * _characteristic_length(boundary_face, state));
 }
