+++ modules/navier_stokes/src/fvkernels/INSFVTKESourceSink.C
@@ -32,38 +32,20 @@ INSFVTKESourceSink::validParams()
   params.addParam<std::vector<BoundaryName>>("walls", "Boundaries that correspond to solid walls.");
 
   params.addParam<Real>("max_mixing_length",
-                        10.0,
+                        1e10,
                         "Maximum mixing legth allowed for the domain - adjust for realizable "
                         "k-epsilon to work properly.");
   params.addParam<bool>(
-      "linearized_model", false, "Boolean to determine if the problem is linearized.");
-  params.addParam<MooseFunctorName>(
-      "linear_variable", 1.0, "Linearization coefficient in case the problem has been linearized.");
-  params.addParam<bool>(
-      "realizable_constraint",
+      "linearized_model",
       true,
-      "Boolean to determine if the kEpsilon mixing length realizability constrints are applied.");
-  params.set<unsigned short>("ghost_layers") = 2;
-  params.addParam<Real>("rf", 1.0, "Relaxation factor.");
+      "Boolean to determine if the problem should be use in a linear or nonlinear solve.");
   params.addParam<bool>(
       "non_equilibrium_treatement",
-      true,
+      false,
       "Use non-equilibrium wall treatement (faster than standard wall treatement)");
   params.addParam<Real>("C_mu", 0.09, "Coupled turbulent kinetic energy closure.");
-  MooseEnum relaxation_method("time nl", "nl");
-  params.addParam<MooseEnum>(
-      "relaxation_method",
-      relaxation_method,
-      "The method used for relaxing the turbulent kinetic energy production. "
-      "'nl' = previous nonlinear iteration and 'time' = previous timestep.");
-  params.addParam<unsigned int>(
-      "iters_to_activate",
-      5,
-      "number of iterations needed to activate the source in the turbulent kinetic energy.");
-  params.addParam<Real>(
-      "top_production_bound", 100.0, "Top scale bound for turbulent kinetic energy production.");
-  params.addParam<Real>(
-      "top_destruction_bound", 100.0, "Top scale bound for turbulent kinetic energy destruction.");
+
+  params.set<unsigned short>("ghost_layers") = 2;
 
   return params;
 }
@@ -85,15 +67,8 @@ INSFVTKESourceSink::INSFVTKESourceSink(const InputParameters & params)
     _wall_boundary_names(getParam<std::vector<BoundaryName>>("walls")),
     _max_mixing_length(getParam<Real>("max_mixing_length")),
     _linearized_model(getParam<bool>("linearized_model")),
-    _linear_variable(getFunctor<ADReal>("linear_variable")),
-    _realizable_constraint(getParam<bool>("realizable_constraint")),
-    _rf(getParam<Real>("rf")),
     _non_equilibrium_treatement(getParam<bool>("non_equilibrium_treatement")),
-    _C_mu(getParam<Real>("C_mu")),
-    _relaxation_method(getParam<MooseEnum>("relaxation_method")),
-    _iters_to_activate(getParam<unsigned int>("iters_to_activate")),
-    _top_production_bound(getParam<Real>("top_production_bound")),
-    _top_destruction_bound(getParam<Real>("top_destruction_bound"))
+    _C_mu(getParam<Real>("C_mu"))
 {
 
   if (!_u_var)
@@ -109,18 +84,10 @@ INSFVTKESourceSink::INSFVTKESourceSink(const InputParameters & params)
[…]
