### 2023-11-13 8abc020
Add BC map to the linear variable. (#25722)
