### 2023-11-14 cb98b48
Mods to avoid OOB access
