+++ modules/tensor_mechanics/doc/content/source/userobjects/NodalPatchRecoveryMaterialProperty.md
@@ -1,4 +1,6 @@
-# Nodal Patch Recovery RankTwoAux
+# NodalPatchRecoveryMaterialProperty
+
+This user object performs the calculations and parallel communication to perform the Zienkiewicz-Zhu nodal patch recovery for a scalar component of a provided material property. Supported material property types are listed in[IndexableProperty](IndexableProperty.md).
 
 !syntax description /UserObjects/NodalPatchRecoveryMaterialProperty
 
