+++ modules/tensor_mechanics/doc/content/source/materials/InterpolatedStatefulMaterial.md
@@ -1,0 +1,17 @@
+# InterpolatedStatefulMaterial
+
+!syntax description /Materials/InterpolatedStatefulMaterial
+
+## Description
+
+`InterpolatedStatefulMaterial` reconstitutes the old state of a material property from the old state of the AuxVariables holding the projected components of the material property.
+
+This object is set up by the [ProjectedStatefulMaterialStorageAction](ProjectedStatefulMaterialStorageAction.md).
+
+!syntax parameters /Materials/InterpolatedStatefulMaterial
+
+!syntax inputs /Materials/InterpolatedStatefulMaterial
+
+!syntax children /Materials/InterpolatedStatefulMaterial
+
+!bibtex bibliography
