+++ modules/tensor_mechanics/doc/content/source/actions/ProjectedStatefulMaterialStorageAction.md
@@ -1,0 +1,7 @@
+# ProjectedStatefulMaterialStorageAction
+
+## Description
+
+The `ProjectedStatefulMaterialStorageAction` is the main action in the [ProjectedStatefulMaterialStorage](/ProjectedStatefulMaterialStorage/index.md) system which sets up all necessary objects to project material property components onto nodal or elemental basis functions.
+
+To be expanded.
