+++ conda/peacock/meta.yaml
@@ -3,7 +3,7 @@
 #   moose-dev/*
 #
 # As well as any directions pertaining to modifying those files.
-{% set version = "2023.11.06" %}
+{% set version = "2023.11.10" %}
 
 package:
   name: moose-peacock
