+++ conda/moose-dev/conda_build_config.yaml
@@ -8,4 +8,4 @@ moose_tools:
   - moose-tools 2023.10.27
 
 moose_peacock:
-  - moose-peacock 2023.11.06
+  - moose-peacock 2023.11.10
