### 2023-11-14 12d6008
Fix unity build, docs, and paralleleization (#26053)
and move everything into framework
