+++ framework/src/base/Moose.C
@@ -579,6 +579,8 @@ associateSyntaxInner(Syntax & syntax, ActionFactory & /*action_factory*/)
   registerSyntaxTask("EmptyAction", "Debug/MaterialDerivativeTest", "no_action"); // placeholder
   registerSyntax("MaterialDerivativeTestAction", "Debug/MaterialDerivativeTest/*");
 
+  registerSyntax("ProjectedStatefulMaterialStorageAction", "ProjectedStatefulMaterialStorage/*");
+
   addActionTypes(syntax);
 }
 
