### 2023-11-14 91bd84f
Add ability to control whether we identify variable groups
