### 2023-11-15 #6045 @joshuahansel
@kyleniemeyer Did you ever work out what to do for the paper compilation? Are you waiting on some input from me. Thanks!
