### 2023-11-15 726da1d
Use WASP token type in Field kind for faster data access by pyhit
refs #26068
