+++ modules/navier_stokes/src/kernels/INSMassRZ.C
@@ -38,27 +38,13 @@ INSMassRZ::dStrongViscDUCompLaplace(const unsigned int comp)
 RealVectorValue
 INSMassRZ::strongViscousTermTraction()
 {
-  const Real & r = _q_point[_qp](0);
-  return INSBase::strongViscousTermTraction() +
-         RealVectorValue(2. * _mu[_qp] * (_u_vel[_qp] / (r * r) - _grad_u_vel[_qp](0) / r),
-                         -_mu[_qp] / r * (_grad_v_vel[_qp](0) + _grad_u_vel[_qp](1)),
-                         0);
+  return INSBase::strongViscousTermTraction() + strongViscousTermTractionRZ();
 }
 
 RealVectorValue
 INSMassRZ::dStrongViscDUCompTraction(const unsigned int comp)
 {
-  const Real & r = _q_point[_qp](0);
-  RealVectorValue add_jac(0, 0, 0);
-  if (comp == 0)
-  {
-    add_jac(0) = 2. * _mu[_qp] * (_phi[_j][_qp] / (r * r) - _grad_phi[_j][_qp](0) / r);
-    add_jac(1) = -_mu[_qp] / r * _grad_phi[_j][_qp](1);
-  }
-  else if (comp == 1)
-    add_jac(1) = -_mu[_qp] * _grad_phi[_j][_qp](0) / r;
-
-  return INSBase::dStrongViscDUCompTraction(comp) + add_jac;
+  return INSBase::dStrongViscDUCompTraction(comp) + dStrongViscDUCompTractionRZ(comp);
 }
 
 Real
