+++ framework/doc/content/source/postprocessors/NumMeshDivisions.md
@@ -1,0 +1,9 @@
+# NumMeshDivisions
+
+!syntax description /Postprocessors/NumMeshDivisions
+
+!syntax parameters /Postprocessors/NumMeshDivisions
+
+!syntax inputs /Postprocessors/NumMeshDivisions
+
+!syntax children /Postprocessors/NumMeshDivisions
