### 2023-11-15 1c8a1bf
Make mesh divisions be replicated per thread: - there is no guarantee that the objects calling mesh divsions are NON-threaded - if the callers are threaded, any object called by the mesh division should be thread-safe, no guarantee about that either unless we have a copy per thread
Address Alex's review

- add missing class description
- more constness
- explicit virtual when overriding
- use set_union instead of copying into a vector

Co-authored-by: Alex Lindsay <alexander.lindsay@inl.gov>
