+++ framework/include/meshdivisions/NestedDivision.h
@@ -21,9 +21,9 @@ public:
 
   NestedDivision(const InputParameters & parameters);
 
-  void initialize() override;
-  unsigned int divisionIndex(const Point & pt) const override;
-  unsigned int divisionIndex(const Elem & elem) const override;
+  virtual void initialize() override;
+  virtual unsigned int divisionIndex(const Point & pt) const override;
+  virtual unsigned int divisionIndex(const Elem & elem) const override;
 
 protected:
   /// Vector of nested divisions. Indexing is more and more 'inner' as we progress in the vector
