+++ framework/include/meshdivisions/ExtraElementIntegerDivision.h
@@ -21,9 +21,9 @@ public:
 
   ExtraElementIntegerDivision(const InputParameters & parameters);
 
-  void initialize() override;
-  unsigned int divisionIndex(const Point & pt) const override;
-  unsigned int divisionIndex(const Elem & elem) const override;
+  virtual void initialize() override;
+  virtual unsigned int divisionIndex(const Point & pt) const override;
+  virtual unsigned int divisionIndex(const Elem & elem) const override;
 
 private:
   /// Map from extra element ids to division index. Created on calls to initialize()
