+++ framework/src/problems/FEProblemBase.C
@@ -323,7 +323,7 @@ FEProblemBase::FEProblemBase(const InputParameters & parameters)
     _current_nl_sys(nullptr),
     _aux(nullptr),
     _coupling(Moose::COUPLING_DIAG),
-    _mesh_divisions(/*threaded=*/false),
+    _mesh_divisions(/*threaded=*/true),
     _material_props(declareRestartableDataWithContext<MaterialPropertyStorage>(
         "material_props", &_mesh, _material_prop_registry)),
     _bnd_material_props(declareRestartableDataWithContext<MaterialPropertyStorage>(
@@ -2289,14 +2289,17 @@ FEProblemBase::addMeshDivision(const std::string & type,
   parallel_object_only();
   parameters.set<FEProblemBase *>("_fe_problem_base") = this;
   parameters.set<SubProblem *>("_subproblem") = this;
-  std::shared_ptr<MeshDivision> func = _factory.create<MeshDivision>(type, name, parameters, 0);
-  _mesh_divisions.addObject(func);
+  for (THREAD_ID tid = 0; tid < libMesh::n_threads(); tid++)
+  {
+    std::shared_ptr<MeshDivision> func = _factory.create<MeshDivision>(type, name, parameters, tid);
+    _mesh_divisions.addObject(func, tid);
+  }
 }
 
 MeshDivision &
-FEProblemBase::getMeshDivision(const std::string & name) const
+FEProblemBase::getMeshDivision(const std::string & name, const THREAD_ID tid) const
 {
-  auto * const ret = dynamic_cast<MeshDivision *>(_mesh_divisions.getActiveObject(name, 0).get());
+  auto * const ret = dynamic_cast<MeshDivision *>(_mesh_divisions.getActiveObject(name, tid).get());
   if (!ret)
     mooseError("No MeshDivision object named ", name, " of appropriate type");
   return *ret;
