### 2023-11-15 df49c39
Documented theory for volume junction
Refs #25958
