+++ modules/thermal_hydraulics/doc/content/modules/thermal_hydraulics/theory_manual/vace_model/volume_junction.md
@@ -1,0 +1,241 @@
+## Volume Junction
+
+A general conservation law equation can be expressed as
+
+!equation
+\pd{a}{t} + \nabla\cdot\mathbf{f} = 0 \eqc
+
+where $a$ is the conserved quantity. Integrating this equation over a volume
+$\Omega$ gives
+
+!equation
+\int\limits_{\Omega} \pd{a}{t} d\Omega + \int\limits_{\Omega} \nabla\cdot\mathbf{f} d\Omega = 0 \eqp
+
+The time derivative can be expressed as follows:
+
+!equation
+\int\limits_{\Omega} \pd{a}{t} d\Omega = \ddt{\bar{a} V} \eqc
+
+where $\bar{a}$ is the average of $a$ over $\Omega$:
+
+!equation
+\bar{a} \equiv \frac{1}{V}\int\limits_{\Omega} a d\Omega \eqc
+
+and $V \equiv \int\limits_{\Omega} d\Omega$.
+
+The flux term can be expressed as follows:
+
+!equation
+\int\limits_{\Omega} \nabla\cdot\mathbf{f} d\Omega
+  = \int\limits_\Gamma \mathbf{f}\cdot\mathbf{n} d\Gamma \eqc
+
+where $\Gamma$ is the surface of $\Omega$, and $\mathbf{n}$ is the outward
+normal vector at a position on the surface.
+
+Now consider that $\Gamma$ is partitioned into a number of regions:
+$\Gamma = \Gamma_{\text{wall}} \cup \Gamma_{\text{flow}}$, where
+$\Gamma_{\text{flow}} = \bigcup\limits_{i=1}^{N} \Gamma_{\text{flow},i}$.
+The surface $\Gamma_{\text{wall}}$ corresponds to a wall, and the surface
+$\Gamma_{\text{flow},i}$ corresponds to a flow surface, across which a fluid
+may pass. The number of flow surfaces on the volume is denoted by $N$. See
+Figure [fig:volume_junction] for an illustration.
+
+!media thermal_hydraulics/tikz_diagrams/volume_junction.png
+       id=fig:volume_junction
+       caption=Volume junction illustration.
+       style=width:50%;display:block;margin-left:auto;margin-right:auto;text-align:center;
+
+Now the surface integral can be expressed as follows:
+
+!equation
+\int\limits_\Gamma \mathbf{f}\cdot\mathbf{n} d\Gamma =
+    \int\limits_{\Gamma_{\text{wall}}} \mathbf{f}\cdot\mathbf{n} d\Gamma
+    + \sum\limits_{i=1}^N \int\limits_{\Gamma_{\text{flow},i}} \mathbf{f}\cdot\mathbf{n} d\Gamma \eqp
+
+Now consider that the flux $\mathbf{f}$ is constant over each flow surface, and that
+each flow surface is flat (thus making $\mathbf{n}$ constant over the surface).
+Then,
+
+!equation
+\int\limits_{\Gamma_{\text{flow},i}} \mathbf{f}\cdot\mathbf{n} d\Gamma
+    = \mathbf{f}_i \cdot \mathbf{n}_{J,i} A_i \eqc
+
+where $\mathbf{n}_{J,i} = -\mathbf{n}_i$, and $\mathbf{n}_i$ is the outward-facing
[…]
