+++ modules/thermal_hydraulics/doc/content/modules/thermal_hydraulics/theory_manual/vace_model/index.md
@@ -1,0 +1,10 @@
+# Variable-Area Compressible Euler Flow Model
+
+The Variable-Area Compressible Euler (VACE) flow model is used to model one-dimensional,
+single-phase flow in ducts of variable cross-section, using the Euler equations
+of gas dynamics.
+
+!include modules/thermal_hydraulics/theory_manual/vace_model/flow_equations.md
+
+!include modules/thermal_hydraulics/theory_manual/vace_model/volume_junction.md
+
