### 2023-11-15 7436d6a
Enable (hopefully) neighbor materials in explicit contact constraint
