+++ framework/src/systems/NonlinearSystemBase.C
@@ -1143,16 +1143,10 @@ NonlinearSystemBase::setConstraintSecondaryValues(NumericVector<Number> & soluti
 
             // reinit variables on the primary element's face at the contact point
             _fe_problem.setNeighborSubdomainID(primary_elem, 0);
+            subproblem.reinitNeighborPhys(primary_elem, primary_side, points, 0);
             _fe_problem.reinitNeighborPhys(
                 _mesh.elemPtr(primary_elem->id()), primary_side, points, 0);
-            // _fe_problem.reinitNeighbor(primary_elem, primary_side, 0);
-            // _fe_problem.reinitNeighborFaceRef(
-            //     primary_elem, primary_side, primary_boundary, TOLERANCE, &points, nullptr, 0);
-
-            // Two material interface relies on reiniting materials boundary
-            // _fe_problem.reinitMaterialsBoundary(primary_boundary, 0);
             _fe_problem.reinitMaterialsNeighbor(primary_elem->subdomain_id(), 0);
-            // _fe_problem.reinitMaterialsInterface(primary_boundary, 0);
 
             for (const auto & nfc : constraints)
             {
@@ -1303,14 +1297,9 @@ NonlinearSystemBase::constraintResiduals(NumericVector<Number> & residual, bool 
             // reinit variables on the primary element's face at the contact point
             _fe_problem.setNeighborSubdomainID(primary_elem, 0);
             subproblem.reinitNeighborPhys(primary_elem, primary_side, points, 0);
-            // _fe_problem.reinitNeighbor(primary_elem, primary_side, 0);
-            _fe_problem.reinitNeighborFaceRef(
-                primary_elem, primary_side, primary_boundary, TOLERANCE, &points, nullptr, 0);
-
-            // Two material interface relies on reiniting materials boundary
-            _fe_problem.reinitMaterialsBoundary(primary_boundary, 0);
+            _fe_problem.reinitNeighborPhys(
+                _mesh.elemPtr(primary_elem->id()), primary_side, points, 0);
             _fe_problem.reinitMaterialsNeighbor(primary_elem->subdomain_id(), 0);
-            // _fe_problem.reinitMaterialsInterface(primary_boundary, 0);
 
             for (const auto & nfc : constraints)
             {
@@ -2205,14 +2194,9 @@ NonlinearSystemBase::constraintJacobians(bool displaced)
             // reinit variables on the primary element's face at the contact point
             _fe_problem.setNeighborSubdomainID(primary_elem, 0);
             subproblem.reinitNeighborPhys(primary_elem, primary_side, points, 0);
-            // _fe_problem.reinitNeighbor(primary_elem, primary_side, 0);
-            _fe_problem.reinitNeighborFaceRef(
-                primary_elem, primary_side, primary_boundary, TOLERANCE, &points, nullptr, 0);
-
-            // Two material interface relies on reiniting materials boundary
-            _fe_problem.reinitMaterialsBoundary(primary_boundary, 0);
+            _fe_problem.reinitNeighborPhys(
+                _mesh.elemPtr(primary_elem->id()), primary_side, points, 0);
             _fe_problem.reinitMaterialsNeighbor(primary_elem->subdomain_id(), 0);
-            // _fe_problem.reinitMaterialsInterface(primary_boundary, 0);
 
             for (const auto & nfc : constraints)
             {
