+++ framework/src/constraints/NodeFaceConstraint.C
@@ -284,3 +284,11 @@ NodeFaceConstraint::overwriteSecondaryResidual()
 {
   return _overwrite_secondary_residual;
 }
+
+const std::set<BoundaryID> &
+NodeFaceConstraint::getBoundaryIDs()
+{
+  _boundary_ids.insert(_primary);
+  _boundary_ids.insert(_secondary);
+  return _boundary_ids;
+}
