### 2023-11-15 244aa75
Remove deprecated elasticity tensor class, Refs #23451, Closes #26074
