+++ modules/reactor/src/meshgenerators/AssemblyMeshGenerator.C
@@ -117,13 +117,6 @@ AssemblyMeshGenerator::AssemblyMeshGenerator(const InputParameters & parameters)
   // Initialize ReactorMeshParams object stored in pin input
   initializeReactorMeshParams(reactor_params);
 
-  // Declare dependency of inputs to sub generator calls. If mesh generation
-  // should be bypassed, then store the input meshes to free later
-  if (!getReactorParam<bool>(RGMB::bypass_meshgen))
-    declareMeshesForSub("inputs");
-  else
-    _mesh_ptrs = getMeshes("inputs");
-
   _geom_type = getReactorParam<std::string>(RGMB::mesh_geometry);
   _mesh_dimensions = getReactorParam<int>(RGMB::mesh_dimensions);
 
@@ -250,78 +243,81 @@ AssemblyMeshGenerator::AssemblyMeshGenerator(const InputParameters & parameters)
   else
     _has_duct_block_names = false;
 
-  _assembly_boundary_id = 2000 + _assembly_type;
-  _assembly_boundary_name = "outer_assembly_" + std::to_string(_assembly_type);
-
-  // Call PatternedHexMeshGenerator or PatternedCartesianMeshGenerator to stitch assembly
+  if (!getReactorParam<bool>(RGMB::bypass_meshgen))
   {
-    const auto patterned_mg_name =
-        _geom_type == "Hex" ? "PatternedHexMeshGenerator" : "PatternedCartesianMeshGenerator";
-    auto params = _app.getFactory().getValidParams(patterned_mg_name);
+    // Declare dependency of inputs to sub generator calls. If mesh generation
+    declareMeshesForSub("inputs");
 
-    if (_geom_type == "Hex")
+    _assembly_boundary_id = 2000 + _assembly_type;
+    _assembly_boundary_name = "outer_assembly_" + std::to_string(_assembly_type);
+
+    // Call PatternedHexMeshGenerator or PatternedCartesianMeshGenerator to stitch assembly
     {
-      params.set<Real>("hexagon_size") = getReactorParam<Real>(RGMB::assembly_pitch) / 2.0;
-      params.set<MooseEnum>("hexagon_size_style") = "apothem";
-    }
-    else
-    {
-      if (_background_region_id.size() == 0)
-        params.set<MooseEnum>("pattern_boundary") = "none";
+      const auto patterned_mg_name =
+          _geom_type == "Hex" ? "PatternedHexMeshGenerator" : "PatternedCartesianMeshGenerator";
+      auto params = _app.getFactory().getValidParams(patterned_mg_name);
+
+      if (_geom_type == "Hex")
+      {
+        params.set<Real>("hexagon_size") = getReactorParam<Real>(RGMB::assembly_pitch) / 2.0;
+        params.set<MooseEnum>("hexagon_size_style") = "apothem";
+      }
       else
       {
-        params.set<MooseEnum>("pattern_boundary") = "expanded";
-        params.set<Real>("square_size") = getReactorParam<Real>(RGMB::assembly_pitch);
-        params.set<bool>("uniform_mesh_on_sides") = true;
+        if (_background_region_id.size() == 0)
+          params.set<MooseEnum>("pattern_boundary") = "none";
+        else
+        {
+          params.set<MooseEnum>("pattern_boundary") = "expanded";
+          params.set<Real>("square_size") = getReactorParam<Real>(RGMB::assembly_pitch);
+          params.set<bool>("uniform_mesh_on_sides") = true;
+        }
       }
-    }
-    params.set<std::vector<std::string>>("id_name") = {"pin_id"};
-    params.set<std::vector<MooseEnum>>("assign_type") = {
-        MooseEnum("cell", "cell")}; // give elems IDs relative to position in assembly
-    params.set<std::vector<MeshGeneratorName>>("inputs") = _inputs;
[…]
