+++ modules/reactor/src/meshgenerators/CoreMeshGenerator.C
@@ -151,7 +151,7 @@ CoreMeshGenerator::CoreMeshGenerator(const InputParameters & parameters)
       }
     }
     // using PRMG-specific options with PTMG
-    if (_periphery_meshgenerator == "triangle")
+    else if (_periphery_meshgenerator == "triangle")
     {
       if (parameters.isParamSetByUser("periphery_num_layers"))
       {
@@ -159,6 +159,9 @@ CoreMeshGenerator::CoreMeshGenerator(const InputParameters & parameters)
                    "periphery_num_layers cannot be used with PTMG periphery mesher.");
       }
     }
+    else
+      paramError("periphery_generator",
+                 "Provided periphery meshgenerator has not been implemented.");
   }
 
   MeshGeneratorName first_nondummy_assembly = "";
@@ -228,14 +231,6 @@ CoreMeshGenerator::CoreMeshGenerator(const InputParameters & parameters)
   // Initialize ReactorMeshParams object stored in pin input
   initializeReactorMeshParams(reactor_params);
 
-  // Declare that all of the meshes in the "inputs" parameter are to be used by
-  // a sub mesh generator. If mesh generation should be bypassed, then store
-  // the input meshes to free later
-  if (!getReactorParam<bool>(RGMB::bypass_meshgen))
-    declareMeshesForSub("inputs");
-  else
-    _mesh_ptrs = getMeshes("inputs");
-
   _geom_type = getReactorParam<std::string>(RGMB::mesh_geometry);
   _mesh_dimensions = getReactorParam<int>(RGMB::mesh_dimensions);
 
@@ -276,272 +271,280 @@ CoreMeshGenerator::CoreMeshGenerator(const InputParameters & parameters)
     }
   }
 
-  // Stitch assemblies into a hexagonal / Cartesian core lattice
+  if (!getReactorParam<bool>(RGMB::bypass_meshgen))
   {
-    // create a dummy assembly that is a renamed version of one of the inputs
-    if (make_empty)
+    // Declare that all of the meshes in the "inputs" parameter are to be used by
+    // a sub mesh generator.
+    declareMeshesForSub("inputs");
+
+    // Stitch assemblies into a hexagonal / Cartesian core lattice
     {
+      // create a dummy assembly that is a renamed version of one of the inputs
+      if (make_empty)
       {
-        if (assembly_homogenization)
         {
-          auto params = _app.getFactory().getValidParams("SimpleHexagonGenerator");
-
-          params.set<Real>("hexagon_size") = getReactorParam<Real>(RGMB::assembly_pitch) / 2.0;
-          params.set<std::vector<subdomain_id_type>>("block_id") = {(UINT16_MAX / 2) - 1};
-
-          callMeshSubgenerator("SimpleHexagonGenerator", std::string(_empty_key), params);
-        }
-        else
-        {
-          const auto adaptive_mg_name =
-              _geom_type == "Hex" ? "HexagonConcentricCircleAdaptiveBoundaryMeshGenerator"
-                                  : "CartesianConcentricCircleAdaptiveBoundaryMeshGenerator";
-          auto params = _app.getFactory().getValidParams(adaptive_mg_name);
-
-          const auto assembly_pitch = getReactorParam<Real>(RGMB::assembly_pitch);
-          if (_geom_type == "Hex")
+          if (assembly_homogenization)
           {
-            params.set<Real>("hexagon_size") = assembly_pitch / 2.0;
-            params.set<std::vector<unsigned int>>("num_sectors_per_side") =
-                std::vector<unsigned int>(6, 2);
+            auto params = _app.getFactory().getValidParams("SimpleHexagonGenerator");
+
[…]
