+++ modules/reactor/include/meshgenerators/ReactorGeometryMeshBuilderBase.h
@@ -190,16 +190,6 @@ protected:
                                 std::string elem_block_name,
                                 SubdomainID & next_free_id);
 
-  /**
-   * Calls addMeshSubGenerator to add mesh sub generator defined by RGMB
-   * NOTE: addMeshSubGenerator only gets called if bypass_mesh_generation has not been set to true
-   * in ReactorMeshParams
-   * @param type of mesh generator
-   * @param name of mesh generator
-   * @param input parameters associated with mesh generator
-   */
-  void callMeshSubgenerator(std::string mg_type, std::string mg_name, InputParameters & params);
-
   ///The ReactorMeshParams object that is storing the reactor global information for this reactor geometry mesh
   MeshGeneratorName _reactor_params;
   /// specify the depletion id is generated at which reactor generation level

[…]
