### 2023-11-16 8285bf9
Refactored numerical flux unit tests
Refs #25958
