+++ modules/combined/doc/content/modules/combined/tutorials/index.md
@@ -3,7 +3,7 @@
 The following tutorial and example problems demonstrate how capabilities from multiple
 physics modules can be used together:
 
-- [Intoduction to Coupled Thermo-Mechanical Modeling](combined/tutorials/introduction/index.md)
+- [Introduction to Coupled Thermo-Mechanical Modeling](combined/tutorials/introduction/index.md)
 - [Using Stochastic Tools with Multiphysics Models](combined/examples/stm_thermomechanics.md)
 
 # Solid Isotropic Material Penalization (SIMP) Topology Optimization
