### 2023-11-16 #6045 @kyleniemeyer
Hi @joshuahansel, I just heard back. Since editorialbot compiles the first paper it finds in a repo, we need to make sure that there aren't multiple JOSS papers in a repo. So, I think it makes most sense to have the paper for this module in its own branch (JOSS papers can just live on a separate branch, and don't need to be merged into the main branch with the production code).
