### 2023-11-16 e0642e7
Allow MultiAppCopyTransfer to transfer vector variables
Closes #26045
