### 2023-11-16 5548d3e
Duplicate physics source page as a syntax page refs #25642
