+++ modules/heat_transfer/doc/content/syntax/Physics/HeatConductionFE/index.md
@@ -1,0 +1,1 @@
+../../../source/physics/HeatConductionFE.md
