### 2023-11-16 dae0467
Minor cleanup in ADPump1PhaseUserObject
