+++ modules/thermal_hydraulics/src/userobjects/ADPump1PhaseUserObject.C
@@ -49,11 +49,13 @@ ADPump1PhaseUserObject::computeFluxesAndResiduals(const unsigned int & c)
   const auto & rhowV = _cached_junction_var_values[VolumeJunction1Phase::RHOWV_INDEX];
 
   const ADRealVectorValue di = _dir[0];
-  const ADRealVectorValue rhouV_vec(rhouV, rhovV, rhowV);
+  const ADReal rho = rhoV / _volume;
+  const ADRealVectorValue uvec(rhouV / rhoV, rhovV / rhoV, rhowV / rhoV);
 
   // compute momentum and energy source terms
-  const ADRealVectorValue S_momentum = 0.5 * (rhoV / _volume) * _g * _head * _A_ref * di;
-  const ADReal S_energy = 0.5 * ((rhouV_vec * di) / _volume) * _g * _head * _A_ref;
+  const ADRealVectorValue S_momentum = 0.5 * rho * _g * _head * _A_ref * di;
+  const ADReal S_energy = S_momentum * uvec;
+
   _residual[VolumeJunction1Phase::RHOUV_INDEX] -= S_momentum(0);
   _residual[VolumeJunction1Phase::RHOVV_INDEX] -= S_momentum(1);
   _residual[VolumeJunction1Phase::RHOWV_INDEX] -= S_momentum(2);
