+++ test/src/main.C
@@ -10,8 +10,7 @@
 #include "MooseInit.h"
 #include "MooseTestApp.h"
 #include "Moose.h"
-#include "AppFactory.h"
-#include "Parser.h"
+#include "MooseCreate.h"
 
 #include <iostream>
 
@@ -27,13 +26,11 @@ main(int argc, char * argv[])
   // Register this application's MooseApp and any it depends on
   MooseTestApp::registerApps();
 
-  // Construct front parser
-  Parser front_parser;
+  // Instanitiate Moose App
+  MooseCreate create(argc, argv);
 
-  auto * p = &front_parser;
-
-  // Create an instance of the application and store it in a smart pointer for easy cleanup
-  std::shared_ptr<MooseApp> app = AppFactory::createAppShared("MooseTestApp", argc, argv, p);
+  // Get the created application;
+  std::shared_ptr<MooseApp> app = create.getApp();
 
   // Execute the application
   app->run();
