+++ framework/src/base/AppFactory.C
@@ -38,7 +38,7 @@ MooseAppPtr
 AppFactory::createAppShared(const std::string & default_app_type,
                             int argc,
                             char ** argv,
-                            Parser * parser,
+                            std::shared_ptr<Parser> parser,
                             MPI_Comm comm_world_in)
 {
   auto command_line = std::make_shared<CommandLine>(argc, argv);
@@ -54,15 +54,15 @@ AppFactory::createAppShared(const std::string & default_app_type,
   app_params.set<int>("_argc") = argc;
   app_params.set<char **>("_argv") = argv;
   app_params.set<std::shared_ptr<CommandLine>>("_command_line") = command_line;
-  app_params.set<Parser *>("_parser") = parser;
 
-  return AppFactory::instance().createShared(app_type, "main", app_params, comm_world_in);
+  return AppFactory::instance().createShared(app_type, "main", app_params, parser, comm_world_in);
 }
 
 MooseAppPtr
 AppFactory::createShared(const std::string & app_type,
                          const std::string & name,
                          InputParameters parameters,
+                         std::shared_ptr<Parser> parser,
                          MPI_Comm comm_world_in)
 {
   // Error if the application type is not located
@@ -73,6 +73,10 @@ AppFactory::createShared(const std::string & app_type,
 
   // Take the app_type and add it to the parameters so that it can be retrieved in the Application
   parameters.set<std::string>("_type") = app_type;
+
+  // Take the front parser and add it to the parameters so that it can be retrieved in the
+  // Application
+  parameters.set<std::shared_ptr<Parser>>("_parser") = parser;
 
   // Check to make sure that all required parameters are supplied
   parameters.checkParams("");
