### 2023-11-16 #6045 @joshuahansel
Thanks! I've moved the files to the root of the repo on my branch (renamed `joss_paper.md` to `paper.md` just in case). It's on the branch `thm-joss-paper-review` on my fork: https://github.com/joshuahansel/moose/tree/thm-joss-paper-review. Is this what you need?
