### 2023-11-17 ebe3af2
Update main.c file after parser change, change parser to parserOther in some calls
