+++ modules/functional_expansion_tools/src/main.C
@@ -11,6 +11,7 @@
 #include "Moose.h"
 #include "MooseApp.h"
 #include "AppFactory.h"
+#include "MooseCreate.h"
 
 #include "FunctionalExpansionToolsTestApp.h"
 
@@ -27,9 +28,11 @@ main(int argc, char * argv[])
   // Register this application's MooseApp and any it depends on
   FunctionalExpansionToolsTestApp::registerApps();
 
-  // Create an instance of the application and store it in a smart pointer for easy cleanup
-  std::shared_ptr<MooseApp> app =
-      AppFactory::createAppShared("FunctionalExpansionToolsTestApp", argc, argv);
+  // Instanitiate Moose App
+  MooseCreate create("FunctionalExpansionToolsTestApp", argc, argv);
+
+  // Get the created application;
+  std::shared_ptr<MooseApp> app = create.getApp();
 
   // Execute the application
   app->run();
