### 2023-11-17 ddcd882
Add MooseVariableInterface to ensure the warehouse query finds the target. (#25722)
