+++ modules/tensor_mechanics/doc/content/source/vectorpostprocessors/SectionDisplacementAverage.md
@@ -1,0 +1,18 @@
+# SectionDisplacementAverage
+
+!syntax description /VectorPostprocessors/SectionDisplacementAverage
+
+## Description
+
+This article is a stub:
+
+\begin{equation}
+K_{A}=\sqrt{K}
+\end{equation}
+
+
+!syntax parameters /VectorPostprocessors/SectionDisplacementAverage
+
+!syntax inputs /VectorPostprocessors/SectionDisplacementAverage
+
+!syntax children /VectorPostprocessors/SectionDisplacementAverage
