### 2023-11-17 a5f1dd0
adding compnent with mass source Refs #26086
