### 2023-11-17 db538dd
Regold unit test (#26053)
