+++ unit/src/MooseServerTest.C
@@ -942,53 +942,53 @@ TEST_F(MooseServerTest, CompletionDocumentRootLevel)
   // expected completions with zero-based lines and columns
 
   std::string completions_expect = R"INPUT(
-label: active               text: active = '__all__'             desc:            If specified only the blocks named w... pos: [42.0]-[42.0]
-label: inactive             text: inactive = 'value'             desc:            If specified blocks matching these i... pos: [42.0]-[42.0]
-label: Adaptivity           text: [Adaptivity]\n  \n[]           desc:            application named block                 pos: [42.0]-[42.0]
-label: AuxKernels           text: [AuxKernels]\n  \n[]           desc:            application named block                 pos: [42.0]-[42.0]
-label: AuxScalarKernels     text: [AuxScalarKernels]\n  \n[]     desc:            application named block                 pos: [42.0]-[42.0]
-label: AuxVariables         text: [AuxVariables]\n  \n[]         desc:            application named block                 pos: [42.0]-[42.0]
-label: BCs                  text: [BCs]\n  \n[]                  desc:            application named block                 pos: [42.0]-[42.0]
-label: Bounds               text: [Bounds]\n  \n[]               desc:            application named block                 pos: [42.0]-[42.0]
-label: Constraints          text: [Constraints]\n  \n[]          desc:            application named block                 pos: [42.0]-[42.0]
-label: Controls             text: [Controls]\n  \n[]             desc:            application named block                 pos: [42.0]-[42.0]
-label: DGKernels            text: [DGKernels]\n  \n[]            desc:            application named block                 pos: [42.0]-[42.0]
-label: Dampers              text: [Dampers]\n  \n[]              desc:            application named block                 pos: [42.0]-[42.0]
-label: Debug                text: [Debug]\n  \n[]                desc:            application named block                 pos: [42.0]-[42.0]
-label: DeprecatedBlock      text: [DeprecatedBlock]\n  \n[]      desc:            application named block                 pos: [42.0]-[42.0]
-label: DiracKernels         text: [DiracKernels]\n  \n[]         desc:            application named block                 pos: [42.0]-[42.0]
-label: Distributions        text: [Distributions]\n  \n[]        desc:            application named block                 pos: [42.0]-[42.0]
-label: Executioner          text: [Executioner]\n  \n[]          desc:            application named block                 pos: [42.0]-[42.0]
-label: Executors            text: [Executors]\n  \n[]            desc:            application named block                 pos: [42.0]-[42.0]
-label: FVBCs                text: [FVBCs]\n  \n[]                desc:            application named block                 pos: [42.0]-[42.0]
-label: FVICs                text: [FVICs]\n  \n[]                desc:            application named block                 pos: [42.0]-[42.0]
-label: FVInterfaceKernels   text: [FVInterfaceKernels]\n  \n[]   desc:            application named block                 pos: [42.0]-[42.0]
-label: FVKernels            text: [FVKernels]\n  \n[]            desc:            application named block                 pos: [42.0]-[42.0]
-label: Functions            text: [Functions]\n  \n[]            desc:            application named block                 pos: [42.0]-[42.0]
-label: FunctorMaterials     text: [FunctorMaterials]\n  \n[]     desc:            application named block                 pos: [42.0]-[42.0]
-label: GlobalParams         text: [GlobalParams]\n  \n[]         desc:            application named block                 pos: [42.0]-[42.0]
-label: ICs                  text: [ICs]\n  \n[]                  desc:            application named block                 pos: [42.0]-[42.0]
-label: InterfaceKernels     text: [InterfaceKernels]\n  \n[]     desc:            application named block                 pos: [42.0]-[42.0]
[…]
