### 2023-11-17 #6045 @joshuahansel
I'm looking into this, asking my colleagues. By the way, there is a precedent for a MOOSE module JOSS paper: (https://github.com/openjournals/joss-reviews/issues/3314). From what I can tell, it looks like they did what I tried on my first try, but in their case, their paper was the first one found in the repo. Ours was living at one level deeper (and then got moved to the root level on the branch in my fork). I'll see what my colleagues say.
