### 2023-11-17 #6045 @kyleniemeyer
@joshuahansel yes, if there is just a single paper in a repo, then it can just be in a `paper` folder or similar (or wherever) - the issue is because there are now multiple submissions from the same repo.
