### 2023-11-17 d20ed43
FEAT: Added an array coupled time derivative kernel
