+++ framework/src/kernels/ArrayCoupledTimeDerivative.C
@@ -15,25 +15,30 @@ InputParameters
 ArrayCoupledTimeDerivative::validParams()
 {
   InputParameters params = ArrayKernel::validParams();
-  params.addClassDescription("Time derivative Array Kernel that acts on a coupled variable. Weak form: "
-                             "$(\\psi_i, \\frac{\\partial v_h}{\\partial t})$. The coupled varialbe and"
-                             "the variable must have the same dimensionality");
+  params.addClassDescription(
+      "Time derivative Array Kernel that acts on a coupled variable. Weak form: "
+      "$(\\psi_i, \\frac{\\partial v_h}{\\partial t})$. The coupled varialbe and"
+      "the variable must have the same dimensionality");
   params.addRequiredCoupledVar("v", "Coupled variable");
   return params;
 }
 
 ArrayCoupledTimeDerivative::ArrayCoupledTimeDerivative(const InputParameters & parameters)
-  : ArrayKernel(parameters), _v_dot(coupledArrayDot("v")), _dv_dot(coupledArrayDotDu("v")), _v_var(coupled("v"))
+  : ArrayKernel(parameters),
+    _v_dot(coupledArrayDot("v")),
+    _dv_dot(coupledArrayDotDu("v")),
+    _v_var(coupled("v"))
 {
 }
 
 void
-ArrayCoupledTimeDerivative::computeQpResidual(RealEigenVector &residual)
+ArrayCoupledTimeDerivative::computeQpResidual(RealEigenVector & residual)
 {
 
-  mooseAssert(_var.count()==_v_dot.size(), "The variable and coupled variable have unequal sizes: "
-                                           "  variable size        : " std::to_string(_var.count())
-                                           "  coupled variable size: " _v_dot.size());
+  mooseAssert(_var.count() == _v_dot.size(),
+              "The variable and coupled variable have unequal sizes: "
+              "  variable size        : " std::to_string(
+                  _var.count()) "  coupled variable size: " _v_dot.size());
 
   residual = _test[_i][_qp] * _v_dot[_qp];
 }
@@ -48,7 +53,8 @@ RealEigenMatrix
 ArrayCoupledTimeDerivative::computeQpOffDiagJacobian(const MooseVariableFEBase & jvar)
 {
   if (jvar.number() == _v_var)
-    return _test[_i][_qp] * _phi[_j][_qp] * _dv_dot[_qp] * RealEigenVector::Ones(jvar.count()).asDiagonal();
+    return _test[_i][_qp] * _phi[_j][_qp] * _dv_dot[_qp] *
+           RealEigenVector::Ones(jvar.count()).asDiagonal();
 
   return RealEigenMatrix::Zero(_var.count(), jvar.count());
 }
