### 2023-11-17 5b65447
Make node patch recovery threadsafe (#26053)
