### 2023-11-17 eaab15c
Add missing parameter to VectorCurlPenaltyDirichletBC (#26092)
Co-authored-by: Guillaume Giudicelli <guillaume.giudicelli@gmail.com>
