### 2023-11-17 6434dfd
Apply clang-format suggestions (#26105)
