### 2023-11-18 dec535f
Add a VPP for performing reductions on functors based on a mesh division refs #25901 #22789
