+++ framework/src/vectorpostprocessors/MeshDivisionFunctorReductionVectorPostprocessor.C
@@ -1,0 +1,158 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "MeshDivisionFunctorReductionVectorPostprocessor.h"
+#include "MeshDivision.h"
+#include "MooseMesh.h"
+#include "MooseMeshUtils.h"
+#include "FEProblemBase.h"
+
+#include "libmesh/mesh_base.h"
+
+registerMooseObject("MooseApp", MeshDivisionFunctorReductionVectorPostprocessor);
+
+InputParameters
+MeshDivisionFunctorReductionVectorPostprocessor::validParams()
+{
+  InputParameters params = ElementVectorPostprocessor::validParams();
+  MooseEnum reduction("integral average min max");
+  params.addRequiredParam<MooseEnum>("reduction", reduction, "Reduction operation to apply");
+  params.addRequiredParam<std::vector<MooseFunctorName>>("functors",
+                                                         "Functors to apply the reduction on");
+  params.addRequiredParam<MeshDivisionName>(
+      "mesh_division",
+      "Mesh division object forming the partition of the elements to perform the reduction within");
+  params.addClassDescription("Perform reductions on functors based on a per-mesh-division basis");
+  return params;
+}
+
+MeshDivisionFunctorReductionVectorPostprocessor::MeshDivisionFunctorReductionVectorPostprocessor(
+    const InputParameters & parameters)
+  : ElementVectorPostprocessor(parameters),
+    NonADFunctorInterface(this),
+    _reduction(getParam<MooseEnum>("reduction")),
+    _nfunctors(getParam<std::vector<MooseFunctorName>>("functors").size()),
+    _mesh_division(_fe_problem.getMeshDivision(getParam<MeshDivisionName>("mesh_division"), _tid))
+{
+  // Gather the functors
+  const auto functor_names = getParam<std::vector<MooseFunctorName>>("functors");
+  for (const auto i : make_range(_nfunctors))
+    _functors.push_back(&getFunctor<Real>(functor_names[i]));
+
+  // Set up reduction vectors
+  const auto ndiv = _mesh_division.getNumDivisions();
+  _volumes.resize(ndiv);
+  for (const auto i : index_range(functor_names))
+  {
+    auto & p = declareVector(functor_names[i]);
+    p.resize(ndiv);
+    _functor_reductions.push_back(&p);
+  }
+}
+
+void
+MeshDivisionFunctorReductionVectorPostprocessor::initialize()
+{
+  for (auto & reduction : _functor_reductions)
+  {
+    if (_reduction == ReductionEnum::INTEGRAL || _reduction == ReductionEnum::AVERAGE)
+      std::fill(reduction->begin(), reduction->end(), 0);
+    else if (_reduction == ReductionEnum::MIN)
+      std::fill(reduction->begin(), reduction->end(), std::numeric_limits<Real>::max());
+    else if (_reduction == ReductionEnum::MAX)
+      std::fill(reduction->begin(), reduction->end(), std::numeric_limits<Real>::min());
+    else
+      mooseAssert(true, "Unknown reduction type");
+  }
+  std::fill(_volumes.begin(), _volumes.end(), 0);
+}
+
+void
+MeshDivisionFunctorReductionVectorPostprocessor::execute()
+{
[…]
