### 2023-11-18 3634421
Add doc and a few tests for mesh division-based reduction refs #22789 #25901
