### 2023-11-18 024c39d
Serial access range (#26053)
