+++ framework/doc/content/source/actions/ProjectedStatefulMaterialStorageAction.md
@@ -2,6 +2,10 @@
 
 ## Description
 
-The `ProjectedStatefulMaterialStorageAction` is the main action in the [ProjectedStatefulMaterialStorage](/ProjectedStatefulMaterialStorage/index.md) system which sets up all necessary objects to project material property components onto nodal or elemental basis functions.
+The `ProjectedStatefulMaterialStorageAction` is the main action in the
+[ProjectedStatefulMaterialStorage](/ProjectedStatefulMaterialStorage/index.md)
+system which sets up all necessary objects to project material property
+components onto nodal or elemental basis functions.
 
-To be expanded.
+Projected and interpolated old state can be enabled using the
+`use_inerpolated_state` parameter available in the `MaterialPropertyInterface`.
