+++ framework/doc/content/source/userobjects/NodalPatchRecoveryMaterialProperty.md
@@ -1,6 +1,9 @@
 # NodalPatchRecoveryMaterialProperty
 
-This user object performs the calculations and parallel communication to perform the Zienkiewicz-Zhu nodal patch recovery for a scalar component of a provided material property. Supported material property types are listed in[IndexableProperty](IndexableProperty.md).
+This user object performs the calculations and parallel communication to perform
+the Zienkiewicz-Zhu nodal patch recovery for a scalar component of a provided
+material property. Supported material property types are listed in
+[IndexableProperty](IndexableProperty.md).
 
 !syntax description /UserObjects/NodalPatchRecoveryMaterialProperty
 
