+++ framework/include/utils_nonunity/RankFourTensor.h
@@ -56,18 +56,6 @@ template <>
 void mooseSetToZero<ADRankFourTensor>(ADRankFourTensor & v);
 }
 
-namespace Moose
-{
-template <typename T>
-struct SerialAccess;
-template <typename T>
-struct SerialAccess<RankFourTensorTempl<T>>
-{
-  static T * data(RankFourTensorTempl<T> & obj) { return &obj(0, 0, 0, 0); }
-  static constexpr std::size_t size(RankFourTensorTempl<T> &) { return RankFourTensorTempl<T>::N4; }
-};
-}
-
 /**
  * RankFourTensorTempl is designed to handle any N-dimensional fourth order tensor, C.
  * Since N is hard-coded to 3, RankFourTensorTempl holds 81 separate C_ijkl entries,
