+++ framework/src/userobjects/ProjectedStatefulMaterialNodalPatchRecovery.C
@@ -8,40 +8,214 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ProjectedStatefulMaterialNodalPatchRecovery.h"
+#include "ElementUserObject.h"
 #include "MaterialBase.h"
 #include "Assembly.h"
 
-registerMooseObject("MooseApp", ProjectedStatefulMaterialNodalPatchRecovery);
+// TIMPI includes
+#include "timpi/communicator.h"
+#include "timpi/parallel_sync.h"
+#include "libmesh/parallel_eigen.h"
+
+registerMooseObject("MooseApp", ProjectedStatefulMaterialNodalPatchRecoveryReal);
+registerMooseObject("MooseApp", ADProjectedStatefulMaterialNodalPatchRecoveryReal);
+registerMooseObject("MooseApp", ProjectedStatefulMaterialNodalPatchRecoveryRealVectorValue);
+registerMooseObject("MooseApp", ADProjectedStatefulMaterialNodalPatchRecoveryRealVectorValue);
+registerMooseObject("MooseApp", ProjectedStatefulMaterialNodalPatchRecoveryRankTwoTensor);
+registerMooseObject("MooseApp", ADProjectedStatefulMaterialNodalPatchRecoveryRankTwoTensor);
+registerMooseObject("MooseApp", ProjectedStatefulMaterialNodalPatchRecoveryRankFourTensor);
+registerMooseObject("MooseApp", ADProjectedStatefulMaterialNodalPatchRecoveryRankFourTensor);
 
 InputParameters
-ProjectedStatefulMaterialNodalPatchRecovery::validParams()
+ProjectedStatefulMaterialNodalPatchRecoveryBase::validParams()
 {
-  InputParameters params = NodalPatchRecoveryMaterialProperty::validParams();
+  InputParameters params = ElementUserObject::validParams();
+
+  params.addRelationshipManager("ElementSideNeighborLayers",
+                                Moose::RelationshipManagerType::ALGEBRAIC,
+                                [](const InputParameters &, InputParameters & rm_params)
+                                { rm_params.set<unsigned short>("layers") = 2; });
+
   return params;
 }
 
-ProjectedStatefulMaterialNodalPatchRecovery::ProjectedStatefulMaterialNodalPatchRecovery(
-    const InputParameters & parameters)
-  : NodalPatchRecoveryMaterialProperty(parameters),
+template <typename T, bool is_ad>
+InputParameters
+ProjectedStatefulMaterialNodalPatchRecoveryTempl<T, is_ad>::validParams()
+{
+  InputParameters params = ProjectedStatefulMaterialNodalPatchRecoveryBase::validParams();
+  MooseEnum orders("CONSTANT FIRST SECOND THIRD FOURTH");
+  params.addRequiredParam<MaterialPropertyName>(
+      "property", "Name of the material property to perform nodal patch recovery on");
+  params.addRequiredParam<MooseEnum>(
+      "patch_polynomial_order",
+      orders,
+      "Polynomial order used in least squares fitting of material property "
+      "over the local patch of elements connected to a given node");
+
+  params.addRelationshipManager("ElementSideNeighborLayers",
+                                Moose::RelationshipManagerType::ALGEBRAIC,
+                                [](const InputParameters &, InputParameters & rm_params)
+                                { rm_params.set<unsigned short>("layers") = 2; });
+
+  params.addParamNamesToGroup("patch_polynomial_order", "Advanced");
+
+  return params;
+}
+
+template <typename T, bool is_ad>
[…]
