### 2023-11-19 0c26b3a
Add capability to sort results for authors.py based on different criteria
