+++ scripts/versioner.py
@@ -31,7 +31,7 @@ MOOSE_DIR = os.environ.get('MOOSE_DIR',
 
 ### Tracking Libraries
 # note: Order is important only for historical lookups; git_ancestor(commit) == True
-TRACKING_LIBRARIES = ['mpich', 'petsc', 'libmesh', 'wasp', 'moose', 'app']
+TRACKING_LIBRARIES = ['mpich', 'petsc', 'libmesh', 'wasp', 'moose-dev', 'app']
 
 ### Beautify the output of jinja2 rendered content that may only exists in conda-build scenarios
 # pylint: disable=unused-argument
@@ -114,7 +114,7 @@ class Versioner:
         """ parses arguments """
         parser = argparse.ArgumentParser(description='Supplies a hash for a given library')
         parser.add_argument('library', nargs='?', metavar='library', choices=entities,
-                            help=f'choose from: {", ".join(entities)}', default='moose')
+                            help=f'choose from: {", ".join(entities)}', default='moose-dev')
         parser.add_argument('commit', nargs='?', metavar='commit', default='HEAD',
                             help='default %(default)s')
         parser.add_argument('-q', '--quiet', action='store_true', default=False,
