+++ scripts/versioner.yaml
@@ -47,14 +47,14 @@ packages:
       - conda/wasp/meta.yaml
       - conda/wasp/conda_build_config.yaml
       - conda/wasp/build.sh
-  moose:
+  moose-dev:
     dependencies:
       - libmesh
       - wasp
     apptainer:
       from: libmesh
     influential:
-      - apptainer/moose.def
+      - apptainer/moose-dev.def
       - apptainer/remove_channels.def
       - conda/peacock/meta.yaml
       - conda/peacock/conda_build_config.yaml
@@ -63,8 +63,8 @@ packages:
       - conda/moose/build.sh
   app:
     dependencies:
-      - moose
+      - moose-dev
     apptainer:
-      from: moose
+      from: moose-dev
     influential:
       - apptainer/app.def
