+++ unit/src/JSONReaderTest.C
@@ -11,12 +11,14 @@
 #include "JSONFileReader.h"
 #include "AppFactory.h"
 #include "Executioner.h"
+#include "MooseCreate.h"
 
 TEST(JSONFileReader, errors)
 {
   // Create a minimal app that can create objects
   const char * argv[2] = {"foo", "\0"};
-  const auto & app = AppFactory::createAppShared("MooseUnitApp", 1, (char **)argv);
+  MooseCreate create("MooseUnitApp", 1, (char **)argv);
+  std::shared_ptr<MooseApp> app = create.getApp();
   const auto & factory = &app->getFactory();
   app->parameters().set<bool>("minimal") = true;
   app->run();
@@ -92,7 +94,8 @@ TEST(JSONFileReader, getters)
 {
   // Create a minimal app that can create objects
   const char * argv[2] = {"foo", "\0"};
-  const auto & app = AppFactory::createAppShared("MooseUnitApp", 1, (char **)argv);
+  MooseCreate create("MooseUnitApp", 1, (char **)argv);
+  const auto & app = create.getApp();
   const auto & factory = &app->getFactory();
   app->parameters().set<bool>("minimal") = true;
   app->run();
