+++ unit/src/BuildMeshTest.C
@@ -3,6 +3,7 @@
 #include "BuildMeshTest.h"
 #include "Registry.h"
 #include "MooseApp.h"
+#include "MooseCreate.h"
 #include "MooseMesh.h"
 #include "MooseUnitApp.h"
 #include "AppFactory.h"
@@ -39,7 +40,8 @@ void
 BuildMeshTest::SetUp()
 {
   const char * argv[2] = {"foo", "\0"};
-  _app = AppFactory::createAppShared("MooseUnitApp", 1, (char **)argv);
+  MooseCreate create("MooseUnitApp", 1, (char **)argv);
+  _app = create.getApp();
   _factory = &_app->getFactory();
   std::string mesh_type = "MeshGeneratorMesh";
   std::string mesh_gen_type = "BuildMeshBaseTypesGenerator";
