+++ unit/src/PositionsTest.C
@@ -12,12 +12,14 @@
 #include "InputPositions.h"
 #include "AppFactory.h"
 #include "Executioner.h"
+#include "MooseCreate.h"
 
 TEST(Positions, getUninitialized)
 {
   // Create a minimal app that can create objects
   const char * argv[2] = {"foo", "\0"};
-  const auto & app = AppFactory::createAppShared("MooseUnitApp", 1, (char **)argv);
+  MooseCreate create("MooseUnitApp", 1, (char **)argv);
+  const auto & app = create.getApp();
   const auto & factory = &app->getFactory();
   app->parameters().set<bool>("minimal") = true;
   app->run();
@@ -84,7 +86,8 @@ TEST(Positions, getters)
 {
   // Create a minimal app that can create objects
   const char * argv[2] = {"foo", "\0"};
-  const auto & app = AppFactory::createAppShared("MooseUnitApp", 1, (char **)argv);
+  MooseCreate create("MooseUnitApp", 1, (char **)argv);
+  const auto & app = create.getApp();
   const auto & factory = &app->getFactory();
   app->parameters().set<bool>("minimal") = true;
   app->run();
