+++ unit/src/LimitersTest.C
@@ -19,6 +19,7 @@
 #include "libmesh/parallel_object.h"
 #include "libmesh/vector_value.h"
 #include "libmesh/tensor_value.h"
+#include "MooseCreate.h"
 
 using namespace libMesh;
 using namespace Moose::FV;
@@ -31,7 +32,8 @@ TEST(LimitersTest, limitVector)
   TensorValue<Real> grad(0);
 
   const char * argv[2] = {"foo", "\0"};
-  auto app = AppFactory::createAppShared("MooseUnitApp", 1, (char **)argv);
+  MooseCreate create("MooseUnitApp", 1, (char **)argv);
+  auto app = create.getApp();
   ReplicatedMesh mesh(app->comm(), /*dim=*/2);
   MeshTools::Generation::build_square(mesh, 2, 2);
   ElemInfo ei(mesh.elem_ptr(0));
