+++ framework/include/bcs/VectorPenaltyDirichletBC.h
@@ -24,8 +24,8 @@ public:
   VectorPenaltyDirichletBC(const InputParameters & parameters);
 
 protected:
-  virtual Real computeQpResidual();
-  virtual Real computeQpJacobian();
+  virtual Real computeQpResidual() override;
+  virtual Real computeQpJacobian() override;
 
 private:
   /// The penalty coefficient
