+++ framework/src/parser/ParserOther.C
@@ -1092,6 +1092,7 @@ ParserOther::extractParams(const std::string & prefix, InputParameters & p)
         setscalar(NonlinearVariableName, string);
         setscalar(AuxVariableName, string);
         setscalar(FunctionName, string);
+        setscalar(MeshDivisionName, string);
         setscalar(UserObjectName, string);
         setscalar(VectorPostprocessorName, string);
         setscalar(IndicatorName, string);
@@ -1159,6 +1160,7 @@ ParserOther::extractParams(const std::string & prefix, InputParameters & p)
         setvector(NonlinearVariableName, string);
         setvector(AuxVariableName, string);
         setvector(FunctionName, string);
+        setvector(MeshDivisionName, string);
         setvector(UserObjectName, string);
         setvector(IndicatorName, string);
         setvector(MarkerName, string);
