### 2023-11-20 6db06c1
MAINT: Adding requests from reviewers
- typo fix
- whitespace fix
- more const
- remove unnecessary virtual
- remove use of `use_displaced_mesh = false`
- update requirement wording
- hit formatting

Co-authored-by: Alex Lindsay <alexander.lindsay@inl.gov>
