### 2023-11-21 62ce2c7
Help the user to choose a vector penalty bc (#26105)
