### 2023-11-21 71a0648
DOC: Added documentation for the new internal variables
