### 2023-11-21 e39ba90
Remove unnecessary switch statement (#26177)
