### 2023-11-21 33960ed
regression tests for strengthening terms, plus src correction with id'ed corner case from tests
