### 2023-11-21 20595ea
Address Alex review
- more range-based loops
- more auto in current range-based loops
- remove threading failure from printing in threaded loop

Co-authored-by: Alex Lindsay <alexander.lindsay@inl.gov>

Update framework/src/vectorpostprocessors/MeshDivisionFunctorReductionVectorPostprocessor.C

Co-authored-by: Alex Lindsay <alexander.lindsay@inl.gov>
