### 2023-11-21 076f8ae
Fix nodal calculations
