### 2023-11-21 7e9143d
Make NodeFace writeable friendly.
Pull variables into constraint.
