+++ framework/src/interfaces/Coupleable.C
@@ -21,6 +21,7 @@
 #include "AuxKernel.h"
 #include "ElementUserObject.h"
 #include "NodalUserObject.h"
+#include "NodeFaceConstraint.h"
 
 Coupleable::Coupleable(const MooseObject * moose_object, bool nodal, bool is_fv)
   : _c_parameters(moose_object->parameters()),
@@ -862,12 +863,13 @@ Coupleable::writableVariable(const std::string & var_name, unsigned int comp)
   const auto * aux = dynamic_cast<const AuxKernel *>(this);
   const auto * euo = dynamic_cast<const ElementUserObject *>(this);
   const auto * nuo = dynamic_cast<const NodalUserObject *>(this);
+  const auto * nfc = dynamic_cast<const NodeFaceConstraint *>(this);
 
-  if (!aux && !euo && !nuo)
-    mooseError("writableVariable() can only be called from AuxKernels, ElementUserObjects, or "
-               "NodalUserObjects. '",
+  if (!aux && !euo && !nuo && !nfc)
+    mooseError("writableVariable() can only be called from AuxKernels, ElementUserObjects, "
+               "NodalUserObjects, or NodeFaceConstraints. '",
                _obj->name(),
-               "' is neither of those.");
+               "' is none of those.");
 
   if (aux && !aux->isNodal() && var->isNodal())
     mooseError("The elemental AuxKernel '",
@@ -931,35 +933,42 @@ Coupleable::checkWritableVar(MooseWritableVariable * var)
 {
   // check block restrictions for compatibility
   const auto * br = dynamic_cast<const BlockRestrictable *>(this);
-  if (!var->hasBlocks(br->blockIDs()))
-    mooseError("The variable '",
-               var->name(),
-               "' must be defined on all blocks '",
-               _obj->name(),
-               "' is defined on");
 
-  // make sure only one object can access a variable
-  for (const auto & ci : _obj->getMooseApp().getInterfaceObjects<Coupleable>())
-    if (ci != this && ci->_writable_coupled_variables[_c_tid].count(var))
-    {
-      // if both this and ci are block restrictable then we check if the block restrictions
-      // are not overlapping. If they don't we permit the call.
-      const auto * br_other = dynamic_cast<const BlockRestrictable *>(ci);
-      if (br && br_other && br->blockRestricted() && br_other->blockRestricted() &&
-          !MooseUtils::setsIntersect(br->blockIDs(), br_other->blockIDs()))
-        continue;
-
-      mooseError("'",
-                 ci->_obj->name(),
-                 "' already obtained a writable reference to '",
+  if (!br)
+    mooseWarning("This object '", _obj->name(), "'is not of the block restrictable type.");
+  else
+  {
+    if (!var->hasBlocks(br->blockIDs()))
+      mooseError("The variable '",
                  var->name(),
-                 "'. Only one object can obtain such a reference per variable and subdomain in a "
-                 "simulation.");
-    }
+                 "' must be defined on all blocks '",
+                 _obj->name(),
+                 "' is defined on");
 
+    // make sure only one object can access a variable
+    for (const auto & ci : _obj->getMooseApp().getInterfaceObjects<Coupleable>())
+      if (ci != this && ci->_writable_coupled_variables[_c_tid].count(var))
+      {
+        // if both this and ci are block restrictable then we check if the block restrictions
+        // are not overlapping. If they don't we permit the call.
+        const auto * br_other = dynamic_cast<const BlockRestrictable *>(ci);
+        if (br && br_other && br->blockRestricted() && br_other->blockRestricted() &&
+            !MooseUtils::setsIntersect(br->blockIDs(), br_other->blockIDs()))
+          continue;
+
+        mooseError("'",
+                   ci->_obj->name(),
[…]
