+++ framework/src/transfers/MultiAppGeneralFieldTransfer.C
@@ -15,6 +15,7 @@
 #include "MooseMesh.h"
 #include "MooseTypes.h"
 #include "MooseVariableFE.h"
+#include "MeshDivision.h"
 #include "Positions.h"
 #include "MultiAppPositions.h" // remove after use_nearest_app deprecation
 #include "MooseAppCoordTransform.h"
@@ -72,6 +73,12 @@ MultiAppGeneralFieldTransfer::validParams()
                              nodes_or_sides,
                              "Whether elemental variable boundary restriction is considered by "
                              "element side or element nodes");
+
+  // Mesh division restriction
+  params.addParam<MeshDivisionName>("from_mesh_division",
+                                    "Mesh division object on the origin application");
+  params.addParam<MeshDivisionName>("to_mesh_division",
+                                    "Mesh division object on the target application");
 
   // Array and vector variables
   params.addParam<std::vector<unsigned int>>("source_variable_components",
@@ -210,6 +217,12 @@ MultiAppGeneralFieldTransfer::initialSetup()
       if (_from_boundaries.size() != boundary_names.size())
         paramError("from_boundaries", "Some boundaries were not found in the mesh");
     }
+
+    if (isParamValid("from_mesh_division"))
+    {
+      const auto & mesh_div_name = getParam<MeshDivisionName>("from_mesh_division");
+      _from_mesh_divisions.push_back(&_from_problems[i_from]->getMeshDivision(mesh_div_name));
+    }
   }
 
   // Loop over all target problems
@@ -232,6 +245,12 @@ MultiAppGeneralFieldTransfer::initialSetup()
       _to_boundaries.insert(boundary_ids.begin(), boundary_ids.end());
       if (_to_boundaries.size() != boundary_names.size())
         paramError("to_boundaries", "Some boundaries were not found in the mesh");
+    }
+
+    if (isParamValid("to_mesh_division"))
+    {
+      const auto & mesh_div_name = getParam<MeshDivisionName>("to_mesh_division");
+      _to_mesh_divisions.push_back(&_from_problems[i_to]->getMeshDivision(mesh_div_name));
     }
   }
 
@@ -492,6 +511,7 @@ MultiAppGeneralFieldTransfer::extractOutgoingPoints(const unsigned int var_index
           request_gather(*to_sys, f, &g, nullsetter, varvec);
 
       // We dont look at boundary restriction, not supported for higher order target variables
+      // Same for mesh divisions
       const auto & to_begin = _to_blocks.empty()
                                   ? to_mesh.active_local_elements_begin()
                                   : to_mesh.active_local_subdomain_set_elements_begin(_to_blocks);
@@ -531,6 +551,10 @@ MultiAppGeneralFieldTransfer::extractOutgoingPoints(const unsigned int var_index
         if (!_to_boundaries.empty() && !onBoundaries(_to_boundaries, to_moose_mesh, node))
           continue;
 
+        // Skip if the node is not indexed within the mesh division
+        if (!_to_mesh_divisions.empty() && !inMeshDivision(_to_mesh_divisions[i_to], *node))
+          continue;
+
         // Cache point information
         // We will use this information later for setting values back to solution vectors
         cacheOutgoingPointInfo(
@@ -551,6 +575,10 @@ MultiAppGeneralFieldTransfer::extractOutgoingPoints(const unsigned int var_index
 
         // Skip if the element does not have a side on the boundary
         if (!_to_boundaries.empty() && !onBoundaries(_to_boundaries, to_moose_mesh, elem))
+          continue;
+
+        // Skip if the element is not indexed within the mesh division
[…]
