### 2023-11-22 8b1658e
Keep track of whether the mesh is fully indexed The concern is for higher order variables being transferred from a division that would cut across an element refs #25883
