### 2023-11-22 8aaed5c
Adds docs, minor clean up for tests and input files, Closes #26120
