+++ modules/tensor_mechanics/doc/content/source/materials/crystal_plasticity/CrystalPlasticityFCCDislocationLinkHuCocksUpdate.md
@@ -1,0 +1,107 @@
+# Crystal Plasticity FCC Dislocation Link Hu-Cocks Update
+
+!syntax description /Materials/CrystalPlasticityFCCDislocationLinkHuCocksUpdate
+
+## Description
+
+This crystal plasticity dislocation glide model was initially developed to model thermal creep in 316 stainless steel, an austenitic face-centered cubic (FCC) material [!citep](hu2016multiscale).
+The Hu-Cocks model uses a per-slip-plane basis as the foundation of the constitutive law formulation.
+Rather than tracking the dislocation densities on individual slip systems (12 systems) as other crystal plasticity formulations have done, see [CrystalPlasticityKalidindiUpdate](/CrystalPlasticityKalidindiUpdate.md), the Hu-Cocks model tracks dislocation pinning points on each of the slip planes (4 planes).
+Barriers to dislocation motion, including forest dislocations, solutes, and solid precipitates are also tracked on a per-slip plane basis.
+
+`CrystalPlasticityFCCDislocationLinkHuCocksUpdate` is designed to be used in conjunction with the [ComputeMultipleCrystalPlasticityStress](/ComputeMultipleCrystalPlasticityStress.md) class to calculate the response of a FCC crystalline solid.
+Details about the algorithm and specific stress and strain measures used in the `CrystalPlasticityUpdate` base class are given on the documentation page for [ComputeMultipleCrystalPlasticityStress](/ComputeMultipleCrystalPlasticityStress.md).
+
+As in other crystal plasticity constitutive models compatible with [ComputeMultipleCrystalPlasticityStress](/ComputeMultipleCrystalPlasticityStress.md), the user must supply the slip plane normal and slip direction information, and the lengths of the unit cell lattice parameters.
+
+## Constitutive Model Definition
+
+The dislocation pinning point density, $N_d$, in the `CrystalPlasticityFCCDislocationLinkHuCocksUpdate` class is tracked as a stateful material property on each of the four slip planes, denoted with $\Omega$, of the FCC crystal.
+However, to interface with the [ComputeMultipleCrystalPlasticityStress](/ComputeMultipleCrystalPlasticityStress.md) class, the dislocation glide (or slip) resistance and the applied resolved shear stress are calculated on a per-slip-system basis; slip systems are notated here with a $\alpha$ superscript.
+
+Recall that for an FCC crystal, there are three slip directions per slip plane, for a total of 12 slip systems.
+The slip resistance is calculated as a function of the pinning point density on each of the slip planes, and all three slip systems within a slip plane are assumed to have the same pinning point density value.
+
+### Slip Rate Calculation
+
+The slip increment on each slip system is defined with a power law relationship as
+\begin{equation}
+    \Delta \gamma^{\alpha} = \Delta \gamma_o \left| \frac{\tau^{\alpha}}{g^{\alpha}} \right| ^p sign \left( \tau^{\alpha} \right)
+\end{equation}
+where $\alpha$ indicates a slip system (a total of 12 for an FCC crystal), $\Delta \gamma_o$ is the reference shear strain increment, $\tau^{\alpha}$ is the applied shear stress on slip system $\alpha$, and $g^{\alpha}$ is the slip
+system resistance value.
+
+The shear strain increment $\Delta \gamma^{\Omega}$ for each slip plane $\Omega$ is defined as the sum of the slip system shear increments on that slip plane:
+\begin{equation}
[…]
