+++ modules/tensor_mechanics/doc/content/source/materials/crystal_plasticity/CrystalPlasticityHCPDislocationSlipBeyerleinUpdate.md
@@ -6,20 +6,20 @@
 
 `CrystalPlasticityHCPDislocationSlipBeyerleinUpdate` is designed to be used in conjunction with the
 [ComputeMultipleCrystalPlasticityStress](/ComputeMultipleCrystalPlasticityStress.md) class to calculate
-the response of a FCC crystalline solid. Details about the algorithm and specific
+the response of a HCP crystalline solid. Details about the algorithm and specific
 stress and strain measures used in the `CrystalPlasticityUpdate` base class are
 given on the documentation page for
 [ComputeMultipleCrystalPlasticityStress](/ComputeMultipleCrystalPlasticityStress.md).
 
-As in other crystal plasticity consitutive models compatible with [ComputeMultipleCrystalPlasticityStress](/ComputeMultipleCrystalPlasticityStress.md), the user must supply the slip plane normal and slip direction information, and the lengths of the unit cell lattice parameters.
-Unlike other crystal plasticity consistutive models, `CrystalPlasticityHCPDislocationSlipBeyerleinUpdate` is set for use with the HCP lattice type only. Because of this lattice type requirement, four entries each are expected for both the slip plane normal and the slip direction vector entries.
+As in other crystal plasticity constitutive models compatible with [ComputeMultipleCrystalPlasticityStress](/ComputeMultipleCrystalPlasticityStress.md), the user must supply the slip plane normal and slip direction information, and the lengths of the unit cell lattice parameters.
+Unlike other crystal plasticity constitutive models, `CrystalPlasticityHCPDislocationSlipBeyerleinUpdate` is set for use with the HCP lattice type only. Because of this lattice type requirement, four entries each are expected for both the slip plane normal and the slip direction vector entries.
 
 ## Constitutive Model Definition
 
 A constitutive model for only the glide and evolution of forest dislocations within a Hexagonal Close-Packed (HCP) crystal lattice is implemented in this class. The constitutive model is taken from the work of [!cite](beyerlein2008dislocation,capolungo2009interaction,beyerlein2010probabilistic), which was developed for a viscoplastic self-consistent application. Here the constitutive model is employed within a crystal plasticity implementation.
 
 `CrystalPlasticityHCPDislocationSlipBeyerleinUpdate` allows for the evaluation of the slip resistance, forest dislocation density, and substructure density on different types of slip systems, such as basal$<a>$, prismatic$<a>$, pyramidal$<c+a>$, and so on. Within this class these different types of slip systems are termed `modes`, following [!cite](beyerlein2008dislocation). Vectors of values for each slip mode dependent parameter may be supplied within the input file. The number of slip systems, value of the burgers vector, and the value of the initial lattice friction are examples of slip mode dependent input file parameters.
-The specification of modes are not explicityly identified within the code, and no particular order is enforced.
+The specification of modes is not explicitly identified within the code, and no particular order is enforced.
 
 !alert warning
 Users are responsible for ensuring that the order of mode-specific parameters is consistent throughout the input file, and that the order of slip system modes corresponds correctly to the order of slip plane normals and slip directions given in the slip system text file, `slip_sys_file_name`.
@@ -31,14 +31,14 @@ The resistance to dislocation glide is represented in this constitutive model as
   \label{eqn:hcpGlideTotalSlipResistance}
[…]
