+++ modules/tensor_mechanics/doc/content/bib/tensor_mechanics.bib
@@ -930,6 +930,30 @@
   author  = {I.J. Beyerlein and C.N. Tomé}
 }
 
+@article{hu2016multiscale,
+   author = {Jianan Hu and Alan C.F. Cocks},
+   doi = {10.1016/j.ijsolstr.2015.09.021},
+   journal = {International Journal of Solids and Structures},
+   month = {1},
+   pages = {21-37},
+   publisher = {Elsevier Ltd},
+   title = {A multi-scale self-consistent model describing the lattice deformation in austenitic stainless steels},
+   volume = {78-79},
+   year = {2016}
+}
+
+@article{hu2016evaluation,
+   author = {Jianan Hu and Bo Chen and David J. Smith and Peter E.J. Flewitt and Alan C.F. Cocks},
+   doi = {10.1016/j.ijplas.2016.05.009},
+   journal = {International Journal of Plasticity},
+   month = {9},
+   pages = {203-223},
+   publisher = {Elsevier Ltd},
+   title = {On the evaluation of the Bauschinger effect in an austenitic stainless steel—The role of multi-scale residual stresses},
+   volume = {84},
+   year = {2016}
+}
+
 @article{borden2016phase,
   title     = {A phase-field formulation for fracture in ductile materials: Finite deformation balance law derivation, plastic degradation, and stress triaxiality effects},
   author    = {Borden, Michael J and Hughes, Thomas JR and Landis, Chad M and Anvari, Amin and Lee, Isaac J},
@@ -973,4 +997,3 @@
     url = {https://www.sciencedirect.com/science/article/pii/002076839390049D},
     author = {R. Nahta and B. Moran}
 }
-
