### 2023-11-22 ded7b49
clang format, Refs #26120
