### 2023-11-22 97cd5f4
Address second reviewer comments, Refs #25932
