### 2023-11-22 b98cd72
Cleanup app definition for includes, add ability to replace dependency
