+++ scripts/apptainer_generator.py
@@ -123,6 +123,8 @@ class ApptainerGenerator:
             add_default_args(parser, remote=True)
             parser.add_argument('--local', action='store_true',
                                 help='Use a local dependency container')
+            parser.add_argument('--dep', type=str,
+                                help='Use this dependency instead')
             parser.add_argument('--alt-dep-tag', type=str,
                                 help='An alternate dependency tag to pull')
             parser.add_argument('--alt-dep-tag-prefix', type=str,
@@ -387,7 +389,10 @@ class ApptainerGenerator:
         header += '# Consumed jinja variables\n'
         header += '#\n'
         for var in sorted(jinja_vars):
-            header += f'#   {var}="{jinja_data.get(var, "")}"\n'
+            contents = f'"{jinja_data.get(var, "")}"'
+            if '\n' in contents:
+                contents = 'Multi-line string'
+            header += f'#   {var}={contents}\n'
         header += '#\n'
         header += '\n'
 
@@ -397,20 +402,20 @@ class ApptainerGenerator:
         """
         Adds common labels to the given definition content
         """
-        definition += '\n\n%labels\n'
+        contents = '%labels\n'
         name = self.name
         if hasattr(self.args, 'modify') and self.args.modify is not None:
             name += '.modified'
-        definition += f'    {name}.host.hostname {socket.gethostname()}\n'
-        definition += f'    {name}.host.user {getpass.getuser()}\n'
-        definition += f'    {name}.moose.sha {self.git_repo_sha(MOOSE_DIR)}\n'
-        definition += f'    {name}.version {self.tag}\n'
+        contents += f'{name}.host.hostname {socket.gethostname()}\n'
+        contents += f'{name}.host.user {getpass.getuser()}\n'
+        contents += f'{name}.moose.sha {self.git_repo_sha(MOOSE_DIR)}\n'
+        contents += f'{name}.version {self.tag}\n'
         # If we have CIVET info, add the url
         if 'CIVET_SERVER' in os.environ and 'CIVET_JOB_ID' in os.environ:
             civet_server = os.environ.get('CIVET_SERVER')
             civet_job_id = os.environ.get('CIVET_JOB_ID')
-            definition += f'    {name}.civet.job {civet_server}/job/{civet_job_id}\n'
-        return definition
+            contents += f'{name}.civet.job {civet_server}/job/{civet_job_id}\n'
+        return definition + '\n\n' + self.add_def_whitespace(contents)
 
     @staticmethod
     def create_filename(app_root, section_key, actions):
@@ -448,35 +453,54 @@ class ApptainerGenerator:
                                       file_path))
         return file_list
 
-    def add_application_additions(self, app_root, jinja_data):
+    @staticmethod
+    def add_def_whitespace(content):
         """
-        Add application specific sections if found to jinja data
-        Example, if found:
-            app_root/apptainer/post_pre_configure.sh
-            app_root/apptainer/post_pre.sh
-            app_root/apptainer/environment.sh
-            etc
-        then the contents therein will be exchanged where appropriate.
+        Adds whitespace to content for adding to a defintion file
+        """
+        if not content:
+            return content
+        return content.replace('\n', '\n    ')
 
-        File naming syntax:
-            <section>_<method>_<action>.sh
+    def add_definition_includes(self, jinja_data):
         """
-        # Add your sections here (be sure to modify apptainer/app.def to match)
-        section_meta = {'environment': [],
-                        'post': ['configure', 'make', 'makeinstall'],
-                        'test': ['test']}
-        for section_key, actions in section_meta.items():
-            file_list = self.create_filename(app_root, section_key, actions)
-            for (a_section, file_path) in file_list:
-                with open(file_path, 'r') as section_file:
[…]
