### 2023-11-22 d36b480
Add entry point for begin and end tests
