### 2023-11-22 570eefb
Use auto consistently when thread-joining UOs use static cast also since it s only called that way define empty functions in header
Equivalent to code formatting but manual so refs #7289
