+++ modules/tensor_mechanics/src/vectorpostprocessors/InteractionIntegral.C
@@ -483,8 +483,7 @@ template <bool is_ad>
 void
 InteractionIntegralTempl<is_ad>::threadJoin(const UserObject & y)
 {
-  const InteractionIntegralTempl<is_ad> & uo =
-      static_cast<const InteractionIntegralTempl<is_ad> &>(y);
+  const auto & uo = static_cast<const InteractionIntegralTempl<is_ad> &>(y);
 
   for (auto i = beginIndex(_interaction_integral); i < _interaction_integral.size(); ++i)
     _interaction_integral[i] += uo._interaction_integral[i];
