+++ modules/phase_field/src/vectorpostprocessors/GrainTextureVectorPostprocessor.C
@@ -69,8 +69,7 @@ GrainTextureVectorPostprocessor::execute()
 void
 GrainTextureVectorPostprocessor::threadJoin(const UserObject & y)
 {
-  const GrainTextureVectorPostprocessor & vpp =
-      static_cast<const GrainTextureVectorPostprocessor &>(y);
+  const auto & vpp = static_cast<const GrainTextureVectorPostprocessor &>(y);
   SamplerBase::threadJoin(vpp);
 }
 
