+++ framework/src/userobjects/ElementIntegralUserObject.C
@@ -52,7 +52,7 @@ ElementIntegralUserObject::getValue() const
 void
 ElementIntegralUserObject::threadJoin(const UserObject & y)
 {
-  const ElementIntegralUserObject & pps = static_cast<const ElementIntegralUserObject &>(y);
+  const auto & pps = static_cast<const ElementIntegralUserObject &>(y);
   _integral_value += pps._integral_value;
 }
 
