+++ modules/phase_field/src/userobjects/ComputeExternalGrainForceAndTorque.C
@@ -186,8 +186,7 @@ ComputeExternalGrainForceAndTorque::finalize()
 void
 ComputeExternalGrainForceAndTorque::threadJoin(const UserObject & y)
 {
-  const ComputeExternalGrainForceAndTorque & pps =
-      static_cast<const ComputeExternalGrainForceAndTorque &>(y);
+  const auto & pps = static_cast<const ComputeExternalGrainForceAndTorque &>(y);
   for (unsigned int i = 0; i < _ncomp; ++i)
     _force_torque_store[i] += pps._force_torque_store[i];
   if (_fe_problem.currentlyComputingJacobian())
