### 2023-11-22 8fd0364
Use auto consistently when thread-joining PPs and VPPs change some dynamic casts to static_casts as cast is guaranteed to pass refs #7829 as code standard pr
