+++ modules/navier_stokes/src/postprocessors/INSExplicitTimestepSelector.C
@@ -108,6 +108,6 @@ INSExplicitTimestepSelector::finalize()
 void
 INSExplicitTimestepSelector::threadJoin(const UserObject & uo)
 {
-  const INSExplicitTimestepSelector & pps = dynamic_cast<const INSExplicitTimestepSelector &>(uo);
+  const auto & pps = static_cast<const INSExplicitTimestepSelector &>(uo);
   _value = std::min(_value, pps._value);
 }
