+++ modules/navier_stokes/include/fvkernels/INSFVTKESourceSink.h
@@ -19,6 +19,8 @@ class INSFVTKESourceSink : public FVElementalKernel
 public:
   static InputParameters validParams();
 
+  virtual void initialSetup() override;
+
   INSFVTKESourceSink(const InputParameters & parameters);
 
 protected:

[…]
