+++ framework/include/kernels/Mass.h
@@ -1,22 +1,0 @@
-//* This file is part of the MOOSE framework
-//* https://www.mooseframework.org
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
-#pragma once
-
-#include "Reaction.h"
-
-class Mass : public Reaction
-{
-public:
-  static InputParameters validParams();
-
-  Mass(const InputParameters & parameters);
-
-  virtual void computeResidual() override {}
-};
