+++ framework/src/transfers/MultiAppGeneralFieldTransfer.C
@@ -360,15 +360,16 @@ MultiAppGeneralFieldTransfer::transferVariable(unsigned int i)
 
   // Fill values and app ids for incoming points
   // We are responsible to compute values for these incoming points
-  auto gather_functor = [this](processor_id_type /*pid*/,
-                               const std::vector<Point> & incoming_points,
-                               std::vector<std::pair<Real, Real>> & outgoing_vals)
+  auto gather_functor =
+      [this](processor_id_type /*pid*/,
+             const std::vector<std::pair<Point, unsigned int>> & incoming_locations,
+             std::vector<std::pair<Real, Real>> & outgoing_vals)
   {
     outgoing_vals.resize(
-        incoming_points.size(),
+        incoming_locations.size(),
         {GeneralFieldTransfer::BetterOutOfMeshValue, GeneralFieldTransfer::BetterOutOfMeshValue});
     // Evaluate interpolation values for these incoming points
-    evaluateInterpValues(incoming_points, outgoing_vals);
+    evaluateInterpValues(incoming_locations, outgoing_vals);
   };
 
   DofobjectToInterpValVec dofobject_to_valsvec(_to_problems.size());
@@ -378,7 +379,7 @@ MultiAppGeneralFieldTransfer::transferVariable(unsigned int i)
   // Copy data out to incoming_vals_ids
   auto action_functor = [this, &i, &dofobject_to_valsvec, &interp_caches, &distance_caches](
                             processor_id_type pid,
-                            const std::vector<Point> & my_outgoing_points,
+                            const std::vector<std::pair<Point, unsigned int>> & my_outgoing_points,
                             const std::vector<std::pair<Real, Real>> & incoming_vals)
   {
     auto & pointInfoVec = _processor_to_pointInfoVec[pid];
@@ -456,7 +457,10 @@ MultiAppGeneralFieldTransfer::cacheOutgoingPointInfo(const Point point,
   // We need to send these data to these processors
   for (auto pid : processors)
   {
-    outgoing_points[pid].push_back(point);
+    unsigned int div_index = 0;
+    if (_from_mesh_divisions.size())
+      _from_mesh_divisions[problem_id]->divisionIndex(point);
+    outgoing_points[pid].push_back(std::pair<Point, unsigned int>(point, div_index));
     // Store point information
     // We can use these information when insert values to solution vector
     PointInfo pointinfo;
@@ -464,7 +468,7 @@ MultiAppGeneralFieldTransfer::cacheOutgoingPointInfo(const Point point,
     pointinfo.dof_object_id = dof_object_id;
     pointinfo.offset = 0;
     if (_from_mesh_divisions.size())
-      pointinfo.spatial_restriction_id = _from_mesh_divisions[problem_id]->divisionIndex(point);
+      pointinfo.spatial_restriction_id = div_index;
     _processor_to_pointInfoVec[pid].push_back(pointinfo);
   }
 }
@@ -613,7 +617,7 @@ MultiAppGeneralFieldTransfer::cacheIncomingInterpVals(
     processor_id_type pid,
     const unsigned int var_index,
     std::vector<PointInfo> & pointInfoVec,
-    const std::vector<Point> & point_requests,
+    const std::vector<std::pair<Point, unsigned int>> & point_requests,
     const std::vector<std::pair<Real, Real>> & incoming_vals,
     DofobjectToInterpValVec & dofobject_to_valsvec,
     InterpCaches & interp_caches,
@@ -653,8 +657,8 @@ MultiAppGeneralFieldTransfer::cacheIncomingInterpVals(
       // Cache solution on target mesh in its local frame of reference
       InterpCache & value_cache = interp_caches[problem_id];
[…]
