### 2023-11-24 9136786
Fix WASP update and rebuild script command line argument handling
closes #25821, refs #25948
