### 2023-11-26 a323f72
Add EM constants to EM namespace, and replace manually-defined constants in module source.
Refs #13744
