+++ modules/electromagnetics/src/bcs/VectorTransientAbsorbingBC.C
@@ -8,6 +8,7 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "VectorTransientAbsorbingBC.h"
+#include "ElectromagneticConstants.h"
 #include "ElectromagneticEnums.h"
 #include "Function.h"
 #include <complex>
@@ -36,7 +37,6 @@ VectorTransientAbsorbingBC::VectorTransientAbsorbingBC(const InputParameters & p
   : VectorIntegratedBC(parameters),
 
     _admittance(getFunction("admittance")),
-    _mu0(4.0 * libMesh::pi * 1.0e-7),
 
     _component(getParam<MooseEnum>("component")),
 
@@ -84,7 +84,7 @@ VectorTransientAbsorbingBC::computeQpResidual()
   VectorValue<std::complex<double>> field_dot(field_dot_0, field_dot_1, field_dot_2);
 
   // Calculate solution field contribution to BC residual
-  std::complex<double> p_dot_test = _mu0 * _admittance.value(_t, _q_point[_qp]) *
+  std::complex<double> p_dot_test = EM::mu_0 * _admittance.value(_t, _q_point[_qp]) *
                                     _test[_i][_qp].cross(_normals[_qp]) *
                                     _normals[_qp].cross(field_dot);
 
@@ -100,7 +100,7 @@ Real
 VectorTransientAbsorbingBC::computeQpJacobian()
 {
   RealVectorValue prefix =
-      _mu0 * _admittance.value(_t, _q_point[_qp]) * _test[_i][_qp].cross(_normals[_qp]);
+      EM::mu_0 * _admittance.value(_t, _q_point[_qp]) * _test[_i][_qp].cross(_normals[_qp]);
 
   return prefix * _normals[_qp].cross(_du_dot_du[_qp] * _phi[_j][_qp]);
 }
@@ -111,7 +111,7 @@ VectorTransientAbsorbingBC::computeQpOffDiagJacobian(unsigned int jvar)
   if (jvar == _coupled_var_num)
   {
     RealVectorValue prefix =
-        _mu0 * _admittance.value(_t, _q_point[_qp]) * _test[_i][_qp].cross(_normals[_qp]);
+        EM::mu_0 * _admittance.value(_t, _q_point[_qp]) * _test[_i][_qp].cross(_normals[_qp]);
 
     return prefix * _normals[_qp].cross(_coupled_dot_du[_qp] * _phi[_j][_qp]);
   }
