### 2023-11-26 8ad5c5c
Switch admittance of free space to new constants
