### 2023-11-26 96c0e1b
Move wasp copy from %setup
