### 2023-11-27 16cb6f3
Add and use PETSC_SRC_DIR for simplification
