### 2023-11-27 a38a927
Add _GIT_SHA and _GIT_REMOTE to petsc and libmesh
