### 2023-11-27 dd9ca22
Remove debugging print
