### 2023-11-27 61b10a9
Add a useful error here
