### 2023-11-27 3fdaeea
Fix the regex to allow for no beginning space
