### 2023-11-27 72dc255
Make November 2023 news public
Add December 2023 template

Refs #11447
