+++ modules/doc/content/newsletter/2023/2023_12.md
@@ -1,0 +1,14 @@
+# MOOSE Newsletter (December 2023)
+
+!alert! construction title=In Progress
+This MOOSE Newsletter edition is in progress. Please check back in January 2024
+for a complete description of all MOOSE changes.
+!alert-end!
+
+## MOOSE Improvements
+
+## libMesh-level Changes
+
+## PETSc-level Changes
+
+## Bug Fixes and Minor Enhancements
