+++ modules/doc/content/newsletter/2023/2023_11.md
@@ -1,10 +1,5 @@
 # MOOSE Newsletter (November 2023)
 
-!alert! construction title=In Progress
-This MOOSE Newsletter edition is in progress. Please check back in December 2023
-for a complete description of all MOOSE changes.
-!alert-end!
-
 ## MOOSE Improvements
 
 ### Coordinate transformations during transfers between applications
