### 2023-11-27 619cc78
Create example of THM-SC coupling Refs #29500
