### 2023-11-27 cff28fb
Create example of THM-SC coupling Refs #29500
