+++ modules/subchannel/Makefile
@@ -33,14 +33,15 @@ CONTACT                     := no
 EXTERNAL_PETSC_SOLVER       := no
 FLUID_PROPERTIES            := yes
 FUNCTIONAL_EXPANSION_TOOLS  := no
-HEAT_CONDUCTION             := yes
+HEAT_TRANSFER               := yes
 LEVEL_SET                   := no
 MISC                        := no
 NAVIER_STOKES               := yes
 PHASE_FIELD                 := no
 POROUS_FLOW                 := no
 RDG                         := no
-REACTOR						          := yes
+REACTOR			                := yes
+THERMAL_HYDRAULICS	        := yes
 RICHARDS                    := no
 SOLID_MECHANICS             := no
 STOCHASTIC_TOOLS            := no
