### 2023-11-27 efe3596
MooseDocs doesn't seem to like lable reference from aligned env
