+++ modules/navier_stokes/doc/content/source/kernels/INSMomentumLaplaceForm.md
@@ -8,7 +8,6 @@ stress. For constant density and viscosity, the following simplification can be
 applied, as described in [!citep](peterson2018overview):
 
 \begin{aligned}
-\label{laplace}
 \nabla\cdot\sigma &= \nabla\cdot\left(-p\bm{I} + \mu\left(\nabla\vec{u} + \left(\nabla\vec{u}\right)^T\right)
                   &= -\nabla p + \mu\left(\nabla\cdot\nabla\vec{u} + \nabla\cdot\left(\nabla\vec{u}\right)^T\right)
                   &= -\nabla p + \mu[left(\nabla\left(\nabla\cdot\vec{u}\right) + \nabla\cdot\left(\nabla\vec{u}\right)^T\right)
@@ -16,7 +15,7 @@ applied, as described in [!citep](peterson2018overview):
                   &= -\nabla p + \mu\nabla^2\vec{u}
 \end{aligned}
 
-Moving from lines 2 to 3 in [laplace] we assumed sufficient smoothness in
+Moving from lines 2 to 3 in the above equation we assumed sufficient smoothness in
 $\vec{u}$ to interchange the divergence and gradient operations. In moving from
 lines 3 to 4 we applied the incompressibility constraint $\nabla\cdot\vec{u} =
 0$. The final form is a vector Laplacian (which can be different in
