### 2023-11-27 1c38135
Add later November work to newsletter
