+++ modules/doc/content/newsletter/2023/2023_11.md
@@ -2,6 +2,16 @@
 
 ## MOOSE Improvements
 
+### MeshDivisions system to split the mesh in arbitrary ways
+
+The [MeshDivisions](syntax/MeshDivisions/index.md) system has been added to MOOSE.
+It was released with a number of simple examples, including [SubdomainDivision.md] for a block ID-based
+division of the mesh and the [CartesianGridDivision.md] for a Cartesian grid-numbering overlaid on top
+of the mesh.
+It is intended to be able to specify arbitrary mappings in transfers between applications.
+For now, the new system can be used for arbitrary reduction operations over the mesh divisions using the
+[MeshDivisionFunctorReductionVectorPostprocessor.md] mostly for postprocessing purposes.
+
 ### Coordinate transformations during transfers between applications
 
 General coordinate transformations between applications have been implemented in the
@@ -87,3 +97,9 @@ when transferring between applications.
 ## PETSc-level Changes
 
 ## Bug Fixes and Minor Enhancements
+
+- A DirichletBC leveraging functors was added for finite element variables: [FunctorDirichletBC.md].
+  This helps create arbitrary couplings between variables at boundary conditions.
+- The ability to evaluate functors defined on the other side of a boundary was added to the
+  [FVFunctorDirichletBC.md]. This helps connect two volumes together where variables connected
+  through Dirichlet boundary conditions are defined on two separate subdomains.
