+++ modules/doc/content/newsletter/2023/2023_11.md
@@ -19,6 +19,35 @@ General coordinate transformations between applications have been implemented in
 from RZ to XYZ geometries, for example, or accounting for arbitrary rotations (defined in the `[Mesh]` blocks)
 when transferring between applications.
 
+### Changes to Darcy/Forchheimer friction
+
+We have changed the way that users should formulate Darcy and Forchheimer
+friction coefficients for simulating friction flows with the finite-volume
+Navier-Stokes classes. Information about the new formulation can be found in the
+relevant friction kernel, [PINSFVMomentumFriction.md]. We have also introduced a
+real example of a class that computes Darcy and Forchheimer friction
+coefficients, [FunctorErgunDragCoefficients.md].
+
+### Selective exclusion of finite element families for p-refinement
+
+Users may now selectively exclude finite element families from
+p-refinement. This allows a use case such as a discontinuous Galerkin
+discretization for the "nonlinear" variables and Lagrange-based auxiliary
+variables for data transfers. In this example the user may want to exclude the
+Lagrange family from p-refinement while p-refining the monomial or L2 Lagrange
+nonlinear variable. Families to exclude are controlled using the
+`disable_p_refinement_for_families` parameter in the `Adaptivity` block. By
+default, all non-hierarchic finite element families are excluded from
+p-refinement.
+
+### Custom eigen normalizations
+
+Users may now provide a custon normalization to their eigen solves using the
+[!param](/Problem/EigenProblem/bx_norm) parameter. This is the last piece that
+the "new" SLEPc-based eigen system needed to emulate the native MOOSE-based
+eigen system. With this addition, the native MOOSE implementation may be removed
+in the future.
+
 ## libMesh-level Changes
 
 ### `2023.11.06` Update
@@ -103,3 +132,4 @@ when transferring between applications.
 - The ability to evaluate functors defined on the other side of a boundary was added to the
   [FVFunctorDirichletBC.md]. This helps connect two volumes together where variables connected
   through Dirichlet boundary conditions are defined on two separate subdomains.
+- The [MultiAppCopyTransfer.md] now supports transfer of vector variables.
