### 2023-11-28 effae2e
Add support for multiple streams (unused)
refs #23352
