+++ framework/src/restart/RestartableDataReader.C
@@ -41,49 +41,68 @@ RestartableDataReader::RestartableDataReader(MooseApp & app, std::vector<Restart
 }
 
 void
-RestartableDataReader::setInput(std::unique_ptr<std::stringstream> header_stream,
+RestartableDataReader::addInput(std::unique_ptr<std::stringstream> header_stream,
                                 std::unique_ptr<std::stringstream> data_stream)
 {
-  mooseAssert(!_streams.header, "Header input already set");
-  mooseAssert(!_streams.data, "Data stream already set");
-  _streams.header = std::make_unique<StringInputStream>(std::move(header_stream));
-  _streams.data = std::make_unique<StringInputStream>(std::move(data_stream));
+  std::unique_ptr<InputStream> header =
+      std::make_unique<StringInputStream>(std::move(header_stream));
+  std::unique_ptr<InputStream> data = std::make_unique<StringInputStream>(std::move(data_stream));
+  addInput(std::move(header), std::move(data));
 }
 
 void
-RestartableDataReader::setInput(const std::filesystem::path & folder_base)
+RestartableDataReader::addInput(const std::filesystem::path & folder_base)
 {
-  mooseAssert(!_streams.header, "Header input already set");
-  mooseAssert(!_streams.data, "Data stream already set");
-  _streams.header = std::make_unique<FileInputStream>(restartableHeaderFile(folder_base));
-  _streams.data = std::make_unique<FileInputStream>(restartableDataFile(folder_base));
+  std::unique_ptr<InputStream> header =
+      std::make_unique<FileInputStream>(restartableHeaderFile(folder_base));
+  std::unique_ptr<InputStream> data =
+      std::make_unique<FileInputStream>(restartableDataFile(folder_base));
+  addInput(std::move(header), std::move(data));
 }
 
-RestartableDataReader::InputStreams
+bool
+RestartableDataReader::isRestoring() const
+{
+  // TODO: not sure if this is right
+  return _is_restoring;
+}
+
+std::vector<RestartableDataReader::InputStreams>
 RestartableDataReader::clear()
 {
   _is_restoring = false;
   _header.clear();
-  InputStreams streams;
+  std::vector<InputStreams> streams;
   std::swap(_streams, streams);
   return streams;
 }
 
-std::vector<std::unordered_map<std::string, RestartableDataReader::HeaderEntry>>
-RestartableDataReader::readHeader(InputStream & header_input) const
+void
+RestartableDataReader::addInput(std::unique_ptr<InputStream> header,
+                                std::unique_ptr<InputStream> data)
 {
-  std::vector<std::unordered_map<std::string, RestartableDataReader::HeaderEntry>> header;
+  InputStreams stream;
+  stream.header = std::move(header);
+  stream.data = std::move(data);
+  _streams.emplace_back(std::move(stream));
+}
 
-  auto stream_ptr = header_input.get();
-  auto & stream = *stream_ptr;
+void
+RestartableDataReader::readHeader(const std::size_t index)
+{
+  mooseAssert(index < _header.size(), "Invalid index");
+
+  InputStream & input_stream = *_streams[index].header;
+  std::shared_ptr<std::istream> stream_ptr = input_stream.get();
+  std::istream & stream = *stream_ptr;
 
   stream.seekg(0);
 
-  const auto error = [&header_input](auto... args)
+  const auto error = [&input_stream](auto... args)
   {
     std::stringstream err_prefix;
     err_prefix << "While reading restartable data in ";
-    const auto filename = header_input.getFilename();
+    const auto filename = input_stream.getFilename();
     if (filename)
       err_prefix << std::filesystem::absolute(filename->parent_path());
     else
@@ -138,7 +157,12 @@ RestartableDataReader::readHeader(InputStream & header_input) const
[…]
