### 2023-11-28 322bf4e
Remove document that no longer exist Refs #26240
