### 2023-11-28 dc37eca
Update documentation for the thm_rehbinder test
Refs #26240
