+++ modules/porous_flow/doc/content/bib/porous_flow.bib
@@ -596,6 +596,7 @@ abstract = {Semi-analytical or closed analytical solutions of the stationary the
   pages={609 - 623},
   url={https://api.semanticscholar.org/CorpusID:123656493}
 }
+<<<<<<< HEAD
 @Article{w14081310,
 AUTHOR = {Karmakar, Shyamal and Tatomir, Alexandru and Oehlmann, Sandra and Giese, Markus and Sauter, Martin},
 TITLE = {Numerical Benchmark Studies on Flow and Solute Transport in Geological Reservoirs},
@@ -628,3 +629,5 @@ volume = {24},
 journal = {Advances in Water Resources},
 doi = {10.1016/S0309-1708(00)00023-3}
 }
+=======
+>>>>>>> abaf588073 (Update documentation for the thm_rehbinder test)
