### 2023-11-28 e0fe9eb
mixture model init commit Refs #22637
