+++ modules/navier_stokes/src/auxkernels/INSFVSlipVelocityAux.C
@@ -1,0 +1,143 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "INSFVSlipVelocityAux.h"
+#include "INSFVVelocityVariable.h"
+#include "Function.h"
+
+registerMooseObject("NavierStokesApp", INSFVSlipVelocityAux);
+
+InputParameters
+INSFVSlipVelocityAux::validParams()
+{
+  InputParameters params = AuxKernel::validParams();
+  params.addClassDescription("Computes the turbulent viscosity for the mixing length model.");
+  params.addRequiredCoupledVar("u", "The velocity in the x direction.");
+  params.addCoupledVar("v", "The velocity in the y direction.");
+  params.addCoupledVar("w", "The velocity in the z direction.");
+  params.addRequiredParam<MooseFunctorName>("rho_c", "Continuous phase density.");
+  params.addRequiredParam<MooseFunctorName>("rho_d", "Dispersed phase density.");
+  params.addParam<RealVectorValue>(
+      "gravity", RealVectorValue(0, 0, 0), "Gravity acceleration vector");
+  params.addClassDescription("Object for advecting momentum with slip velocity, e.g. rho*u");
+  params.addParam<Real>("force_value", 0.0, "Coefficient to multiply by the body force term");
+  params.addParam<FunctionName>("force_function", "1", "A function that describes the body force");
+  params.addParam<PostprocessorName>(
+      "force_postprocessor", 1, "A postprocessor whose value is multiplied by the body force");
+  params.addParam<RealVectorValue>(
+      "force_direction", RealVectorValue(1, 0, 0), "Gravity acceleration vector");
+  params.addParam<MooseFunctorName>(
+      "linear_coef_name", 0.44, "Linear friction coefficient name as a material property");
+  params.addParam<MooseFunctorName>(
+      "particle_diameter", 1.0, "Diameter of particles in the dispersed phase.");
+  params.addParam<MooseFunctorName>("fd", 0.0, "Fraction dispersed phase.");
+  MooseEnum momentum_component("x=0 y=1 z=2");
+  params.addRequiredParam<MooseEnum>(
+      "momentum_component",
+      momentum_component,
+      "The component of the momentum equation that this kernel applies to.");
+  return params;
+}
+
+INSFVSlipVelocityAux::INSFVSlipVelocityAux(const InputParameters & params)
+  : AuxKernel(params),
+    _dim(_subproblem.mesh().dimension()),
+    _u_var(dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("u", 0))),
+    _v_var(params.isParamValid("v")
+               ? dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("v", 0))
+               : nullptr),
+    _w_var(params.isParamValid("w")
+               ? dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("w", 0))
+               : nullptr),
+    _rho_c(getFunctor<ADReal>("rho_c")),
+    _rho_d(getFunctor<ADReal>("rho_d")),
+    _gravity(getParam<RealVectorValue>("gravity")),
+    _force_scale(getParam<Real>("force_value")),
+    _force_function(getFunction("force_function")),
+    _force_postprocessor(getPostprocessorValue("force_postprocessor")),
[…]
