### 2023-11-28 b9bc322
Do not add IC for components if restarting
Closes #26155
