+++ modules/thermal_hydraulics/src/components/HeatTransferBase.C
@@ -139,7 +139,8 @@ HeatTransferBase::addHeatedPerimeter()
       makeFunctionControllableIfConstant(_P_hf_fn_name, "P_hf");
     }
 
-    getTHMProblem().addFunctionIC(_P_hf_name, _P_hf_fn_name, _flow_channel_subdomains);
+    if (!_app.isRestarting())
+      getTHMProblem().addFunctionIC(_P_hf_name, _P_hf_fn_name, _flow_channel_subdomains);
   }
 }
 
