+++ modules/thermal_hydraulics/src/components/HeatTransferFromExternalAppHeatFlux1Phase.C
@@ -34,7 +34,9 @@ HeatTransferFromExternalAppHeatFlux1Phase::addVariables()
 
   getTHMProblem().addSimVariable(
       false, _q_wall_name, FEType(CONSTANT, MONOMIAL), _flow_channel_subdomains);
-  getTHMProblem().addConstantIC(_q_wall_name, 0, _flow_channel_subdomains);
+
+  if (!_app.isRestarting())
+    getTHMProblem().addConstantIC(_q_wall_name, 0, _flow_channel_subdomains);
 }
 
 void
