### 2023-11-28 dac84b9
adding missing includes and docs Refs #9007
