+++ modules/navier_stokes/src/fvkernels/INSFVTKEDSourceSink.C
@@ -28,7 +28,8 @@ INSFVTKEDSourceSink::validParams()
   params.addRequiredParam<MooseFunctorName>(NS::density, "fluid density");
   params.addRequiredParam<MooseFunctorName>(NS::mu, "Dynamic viscosity.");
   params.addRequiredParam<MooseFunctorName>("mu_t", "Turbulent viscosity.");
-  params.addParam<std::vector<BoundaryName>>("walls", "Boundaries that correspond to solid walls.");
+  params.addParam<std::vector<BoundaryName>>(
+      "walls", {}, "Boundaries that correspond to solid walls.");
   params.addParam<Real>("max_mixing_length",
                         1e10,
                         "Maximum mixing legth allowed for the domain - adjust for realizable "
