+++ modules/navier_stokes/include/fvkernels/INSFVTKEDSourceSink.h
@@ -10,6 +10,8 @@
 #pragma once
 
 #include "FVElementalKernel.h"
+#include "MathFVUtils.h"
+#include "INSFVMomentumResidualObject.h"
 
 /**
  * Simple class to demonstrate off diagonal Jacobian contributions.
