+++ modules/navier_stokes/doc/content/source/fvkernels/INSFVTurbulentDiffusion.md
@@ -5,8 +5,10 @@ at certain boundaries.
 
 The particularity of this kernel is that it allows us to skip computing diffusion
 for near-wall elements. The key for this skip are the boundaries identified in
-the `walls` list. For any element that is in contact with a boundary identified
-in the `walls` list, diffusion will be skipped for that element over all faces.
+the [!param](/FVKernels/INSFVTurbulentDiffusion/walls) list.
+For any element that is in contact with a boundary identified
+in the [!param](/FVKernels/INSFVTurbulentDiffusion/walls) list,
+diffusion will be skipped for that element over all faces.
 
 !syntax parameters /FVKernels/INSFVTurbulentDiffusion
 
