+++ modules/navier_stokes/doc/content/source/auxkernels/TurbulentConductivityAux.md
@@ -2,7 +2,9 @@
 
 !syntax description /AuxKernels/TurbulentConductivityAux
 
-This is the axiliary kernel used to compute the thermal effective turbulence conductivity
+## Overview
+
+This is the auxiliary kernel used to compute the thermal effective turbulence conductivity
 
 \begin{equation}
 k_t = \frac{c_p \mu_t}{Pr_t} \,,
@@ -11,7 +13,7 @@ k_t = \frac{c_p \mu_t}{Pr_t} \,,
 where:
 
 - $c_p$ is the specific heat at constant pressure,
-- $\mu_t$ is the dynamic turbulence viscosity,
+- $\mu_t$ is the dynamic turbulent viscosity,
 - $Pr_t$ is the turbulent Prandtl number, which usually ranges between 0.3 and 0.9.
 
 !syntax parameters /AuxKernels/TurbulentConductivityAux
