+++ modules/navier_stokes/doc/content/source/fvkernels/INSFVTKEDSourceSink.md
@@ -2,13 +2,13 @@
 
 !syntax description /FVKernels/INSFVTKEDSourceSink
 
-The object computes the turbulent source and sink term for the turbulent kinetic energy equation.
+The object computes the turbulent source and sink term for the turbulent kinetic energy dissipation rate equation.
 
-Two terms are computed `destruction` and `production` and the term `destruction - production` is
+Two terms are computed: `destruction` and `production` and the term `destruction - production` is
 passed to the residual.
 A different treatment is used for the bulk and the near wall regions.
 
-*Bulk formulation*:
+## Bulk formulation:
 
 The turbulent production $G_k$ is modeled as follows:
 
@@ -24,7 +24,7 @@ where:
 - $S$ is the shear strain tensor internal norm, defined as $S = \sqrt{2\mathbf{S}:\mathbf{S}}$ with the shear strain tensor defined as $\mathbf{S} = \frac{1}{2} [\nabla \vec{u} + (\nabla \vec{u})^T]$,
 - $k$ is the turbulent kinetic energy, which generally comes from a coupled transport equation or can be set by the user for reproducing canonical cases.
 
-The dissipation rate of the turbulent kinetic energy dissipation rate is modeled as follows:
+The destruction of the turbulent kinetic energy dissipation rate is modeled as follows:
 
 \begin{equation}
 \epsilon_{\epsilon} = \frac{C_{2,\epsilon} \rho \epsilon}{t_k} \,,
@@ -33,21 +33,22 @@ The dissipation rate of the turbulent kinetic energy dissipation rate is modeled
 where:
 
 - $C_{2,\epsilon} = 1.92$ is a closure parameter,
-- $\epsilon$ is the solution variable, i.e., the dissipation rate of the turbulent kinetic energy dissipation rate,
-- $t_k = \frac{k}{\epsilon}$ is the turbulent time scale; if the `linearized_model = true`, this timescale is computed from the previous iteration; if `linearized_model = false`, in a nonlinear solve, this timescale is aded to the Jacobian.
+- $\epsilon$ is the solution variable, i.e., the dissipation rate of the turbulent kinetic energy,
+- $t_k = \frac{k}{\epsilon}$ is the turbulent time scale; if the [!param](/FVKernels/INSFVTKEDSourceSink/linearized_model) is `true`, this timescale is computed from the previous iteration; if [!param](/FVKernels/INSFVTKEDSourceSink/linearized_model) is `false`, in a nonlinear solve, this timescale is aded to the Jacobian.
 
-*Wall formulation*:
+## Wall formulation:
 
 All cells in contact with a boundary identified in the `walls` list are applied a different
 treatment in which the equilibrium value for the $\epsilon = \epsilon_{eq}$ is set.
 A separate formulation is used for the `sub-laminar` and `logarithmic` boundary layers.
 The determination of whether the near-wall cell lies in the laminar or logarithmic region
 is performed via the non-dimensional wall distance $y^+$.
-The non-dimensional wall distance is defined as defined differently according to the
-`non_equilibrium_treatement`.
+The non-dimensional wall distance ca be as defined differently according to the
+[!param](/FVKernels/INSFVTKEDSourceSink/non_equilibrium_treatement) parameter.
 
-If `non_equilibrium_treatement = false`, the standard wall function formulations is used in
-which the $y^+$ is found by an incremental fixed point algorithm as follows:
+If [!param](/FVKernels/INSFVTKEDSourceSink/non_equilibrium_treatement) is `false`, the
+standard wall function formulations is used in
+which $y^+$ is found by an incremental fixed point algorithm as follows:
 
 \begin{equation}
 y^+ = \frac{\rho y_p u_{\tau}}{\mu} \,,
@@ -60,18 +61,17 @@ where:
[…]
