+++ modules/navier_stokes/doc/content/source/fvkernels/INSFVTurbulentAdvection.md
@@ -8,8 +8,10 @@ consistency with other INSFV naming conventions, can be of type
 
 The particularity of this kernel is that it allows us to skip computing advection
 for near-wall elements. The key for this skip are the boundaries identified in
-the `walls` list. For any element that is in contact with a boundary identified
-in the `walls` list, advection will be skipped for that element over all faces.
+the [!param](/FVKernels/INSFVTurbulentAdvection/walls) list.
+For any element that is in contact with a boundary identified
+in the [!param](/FVKernels/INSFVTurbulentAdvection/walls) list,
+advection will be skipped for that element over all faces.
 
 !alert note
 This kernel is mainly used for applying wall functions to turbulent problems.

[…]
