+++ modules/tensor_mechanics/src/vectorpostprocessors/SectionDisplacementAverage.C
@@ -26,7 +26,11 @@ SectionDisplacementAverage::validParams()
       "The list of blocks in which to search for cross sectional nodes to compute the displacement "
       "vector average.");
   params.addRequiredParam<Point>("axis_direction", "Direction of the structural component's axis");
-  params.addRequiredParam<Real>("length", "Distance to cross section from the global origin.");
+  params.addRequiredParam<Point>("reference_point",
+                                 "Structural component reference starting point from which the "
+                                 "input parameter 'lengths' applies.");
+  params.addRequiredParam<std::vector<Real>>(
+      "lengths", "Distance(s) to cross section from the global origin.");
   params.addParam<Real>("tolerance",
                         1.0e-6,
                         "Distance tolerance to identify nodes on the user-defined cross section.");
@@ -37,11 +41,14 @@ SectionDisplacementAverage::SectionDisplacementAverage(const InputParameters & p
   : GeneralVectorPostprocessor(parameters),
     _displaced_mesh(_app.actionWarehouse().displacedMesh()),
     _mesh(_app.actionWarehouse().mesh()),
-    _section_displacements(declareVector("section_displacements")),
+    _section_displacements_x(declareVector("section_displacements_x")),
+    _section_displacements_y(declareVector("section_displacements_y")),
+    _section_displacements_z(declareVector("section_displacements_z")),
     _direction(getParam<Point>("axis_direction")),
-    _length(getParam<Real>("length")),
+    _reference_point(getParam<Point>("reference_point")),
+    _lengths(getParam<std::vector<Real>>("lengths")),
     _tolerance(getParam<Real>("tolerance")),
-    _number_of_nodes(0)
+    _number_of_nodes(_lengths.size())
 {
   if (!MooseUtils::absoluteFuzzyEqual(_direction.norm_sq(), 1.0))
     paramError("axis_direction",
@@ -52,19 +59,36 @@ SectionDisplacementAverage::SectionDisplacementAverage(const InputParameters & p
 void
 SectionDisplacementAverage::initialize()
 {
-  _section_displacements.clear();
-  _section_displacements.resize(3);
+  _section_displacements_x.clear();
+  _section_displacements_z.clear();
+  _section_displacements_z.clear();
+  _section_displacements_x.resize(_lengths.size());
+  _section_displacements_y.resize(_lengths.size());
+  _section_displacements_z.resize(_lengths.size());
+  _number_of_nodes.clear();
+  _number_of_nodes.resize(_lengths.size());
 }
 
 void
 SectionDisplacementAverage::finalize()
 {
-  _communicator.sum(_section_displacements);
+  _communicator.sum(_section_displacements_x);
+  _communicator.sum(_section_displacements_y);
+  _communicator.sum(_section_displacements_z);
+
   _communicator.sum(_number_of_nodes);
 
-  _section_displacements[0] /= _number_of_nodes;
-  _section_displacements[1] /= _number_of_nodes;
-  _section_displacements[2] /= _number_of_nodes;
+  for (const auto li : index_range(_lengths))
+    if (_number_of_nodes[li] < 1)
+      mooseError(
+          "No nodes were found in SectionDisplacementAverage postprocessor. Revise your input.");
+
+  for (const auto li : index_range(_lengths))
+  {
+    _section_displacements_x[li] /= _number_of_nodes[li];
+    _section_displacements_y[li] /= _number_of_nodes[li];
+    _section_displacements_z[li] /= _number_of_nodes[li];
+  }
 }
 
 void
@@ -77,62 +101,65 @@ SectionDisplacementAverage::execute()
 
   auto * active_nodes = _mesh->getActiveSemiLocalNodeRange();
 
[…]
