### 2023-11-28 32a6a20
Move VectorFunctionReaction to the framework (#26177)
