### 2023-11-28 aac4232
Factor out acceptance criteria for nearest-position like transfer refs #25883
