+++ framework/src/transfers/MultiAppGeneralFieldUserObjectTransfer.C
@@ -92,8 +92,7 @@ MultiAppGeneralFieldUserObjectTransfer::evaluateInterpValuesWithUserObjects(
   for (auto & [pt, mesh_div] : incoming_points)
   {
     bool point_found = false;
-    if (_nearest_positions_obj)
-      outgoing_vals[i_pt].second = GeneralFieldTransfer::BetterOutOfMeshValue;
+    outgoing_vals[i_pt].second = GeneralFieldTransfer::BetterOutOfMeshValue;
 
     // Loop on all local origin problems until:
     // - we've found the point in an app and the value at that point is valid
@@ -103,43 +102,26 @@ MultiAppGeneralFieldUserObjectTransfer::evaluateInterpValuesWithUserObjects(
          (!point_found || _search_value_conflicts || _nearest_positions_obj);
          ++i_from)
     {
+      Real distance = 1;
       // Check spatial restrictions
-      if (!acceptPointInOriginMesh(i_from, local_bboxes, pt, mesh_div))
+      if (!acceptPointInOriginMesh(i_from, local_bboxes, pt, mesh_div, distance))
         continue;
       else
       {
+        const auto from_global_num = getGlobalSourceAppIndex(i_from);
+
         // Get user object from the local problem
         const UserObject & user_object =
             _from_problems[i_from]->getUserObjectBase(_user_object_name);
 
         // Use spatial value routine to compute the origin value to transfer
-        const auto from_global_num = getGlobalSourceAppIndex(i_from);
         auto val = user_object.spatialValue(_from_transforms[from_global_num]->mapBack(pt));
-
-        // Get nearest position (often a subapp position) for the target point
-        // We want values from the child app that is closest to the same position as the target
-        Point nearest_position_source;
-        if (_nearest_positions_obj)
-        {
-          const bool initial = _fe_problem.getCurrentExecuteOnFlag() == EXEC_INITIAL;
-          const Point nearest_position = _nearest_positions_obj->getNearestPosition(pt, initial);
-          nearest_position_source = _nearest_positions_obj->getNearestPosition(
-              (*_from_transforms[from_global_num])(Point(0, 0, 0)), initial);
-
-          // Source (usually app position) is not closest to the same positions as the target, dont
-          // send values
-          if (nearest_position != nearest_position_source)
-            continue;
-        }
 
         // Look for overlaps. The check is not active outside of overlap search because in that
         // case we accept the first value from the lowest ranked process
         // NOTE: There is no guarantee this will be the final value used among all problems
         //       but we register an overlap as soon as two values are possible from this rank
-        if (detectConflict(val,
-                           outgoing_vals[i_pt].first,
-                           _nearest_positions_obj ? (pt - nearest_position_source).norm() : 1,
-                           outgoing_vals[i_pt].second))
+        if (detectConflict(val, outgoing_vals[i_pt].first, distance, outgoing_vals[i_pt].second))
           registerConflict(i_from, 0, _from_transforms[from_global_num]->mapBack(pt), 1, true);
 
         // No need to consider decision factors if value is invalid
@@ -149,15 +131,10 @@ MultiAppGeneralFieldUserObjectTransfer::evaluateInterpValuesWithUserObjects(
           point_found = true;
 
         // Assign value
-        if (!_nearest_positions_obj)
+        if (distance < outgoing_vals[i_pt].second)
         {
           outgoing_vals[i_pt].first = val;
-          outgoing_vals[i_pt].second = 1;
-        }
-        else if ((pt - nearest_position_source).norm() < outgoing_vals[i_pt].second)
-        {
-          outgoing_vals[i_pt].first = val;
-          outgoing_vals[i_pt].second = (pt - nearest_position_source).norm();
+          outgoing_vals[i_pt].second = distance;
         }
       }
     }
