+++ framework/src/transfers/MultiAppGeneralFieldShapeEvaluationTransfer.C
@@ -108,8 +108,7 @@ MultiAppGeneralFieldShapeEvaluationTransfer::evaluateInterpValuesWithMeshFunctio
   for (auto & [pt, mesh_div] : incoming_points)
   {
     bool point_found = false;
-    if (_nearest_positions_obj)
-      outgoing_vals[i_pt].second = GeneralFieldTransfer::BetterOutOfMeshValue;
+    outgoing_vals[i_pt].second = GeneralFieldTransfer::BetterOutOfMeshValue;
 
     // Loop on all local origin problems until:
     // - we've found the point in an app and the value at that point is valid
@@ -121,7 +120,8 @@ MultiAppGeneralFieldShapeEvaluationTransfer::evaluateInterpValuesWithMeshFunctio
          ++i_from)
     {
       // Check spatial restrictions
-      if (!acceptPointInOriginMesh(i_from, local_bboxes, pt, mesh_div))
+      Real distance = 1;
+      if (!acceptPointInOriginMesh(i_from, local_bboxes, pt, mesh_div, distance))
         continue;
       else
       {
@@ -129,34 +129,11 @@ MultiAppGeneralFieldShapeEvaluationTransfer::evaluateInterpValuesWithMeshFunctio
         const auto from_global_num = getGlobalSourceAppIndex(i_from);
         auto val = (local_meshfuns[i_from])(_from_transforms[from_global_num]->mapBack(pt));
 
-        // Get nearest position (often a subapp position) for the target point
-        // We want values from the child app that is closest to the same position as the target
-        Point nearest_position_source;
-        if (_nearest_positions_obj)
-        {
-          const bool initial = _fe_problem.getCurrentExecuteOnFlag() == EXEC_INITIAL;
-          const Point nearest_position = _nearest_positions_obj->getNearestPosition(pt, initial);
-          nearest_position_source =
-              _nearest_positions_obj->getNearestPosition(_from_positions[i_from], initial);
-          if (!_skip_coordinate_collapsing &&
-              _from_transforms[from_global_num]->hasNonTranslationTransformation())
-            mooseError(
-                "Rotation and scaling currently unsupported with nearest positions transfer.");
-
-          // Source (often app position) is not closest to the same positions as the
-          // target, dont send values
-          if (nearest_position != nearest_position_source)
-            continue;
-        }
-
         // Look for overlaps. The check is not active outside of overlap search because in that
         // case we accept the first value from the lowest ranked process
         // NOTE: There is no guarantee this will be the final value used among all problems
         //       but for shape evaluation we really do expect only one value to even be valid
-        if (detectConflict(val,
-                           outgoing_vals[i_pt].first,
-                           _nearest_positions_obj ? (pt - nearest_position_source).norm() : 1,
-                           outgoing_vals[i_pt].second))
+        if (detectConflict(val, outgoing_vals[i_pt].first, distance, outgoing_vals[i_pt].second))
           registerConflict(i_from, 0, _from_transforms[from_global_num]->mapBack(pt), 1, true);
 
         // No need to consider decision factors if value is invalid
@@ -165,16 +142,11 @@ MultiAppGeneralFieldShapeEvaluationTransfer::evaluateInterpValuesWithMeshFunctio
         else
           point_found = true;
 
-        // Assign value
-        if (!_nearest_positions_obj)
+        // Assign value if it's closer
+        if (distance < outgoing_vals[i_pt].second)
         {
           outgoing_vals[i_pt].first = val;
-          outgoing_vals[i_pt].second = 1;
-        }
-        else if ((pt - nearest_position_source).norm() < outgoing_vals[i_pt].second)
-        {
-          outgoing_vals[i_pt].first = val;
[…]

[…]
