### 2023-11-28 6b1e6e9
Update PETSc, libmesh, VTK, and WASP
Summary of libmesh changes:

- enum_to_string and string_to_enum now initialize their maps via
  static constructors, so they (and functions which depend on them,
  like Elem::get_info() are now thread-safe.
- Added FEInterface::is_hierarchic() shim.
- Added TypeTensor<T>::column() extraction method.
- Added L2_LAGRANGE_VEC finite element family, providing a space
  which locally matches LAGRANGE_VEC but which is discontinuous
  between elements.
- Configure option --with-xdr-include, to make it easier to build
  on systems with XDR support in non-default l
[…]
