+++ conda/libmesh/conda_build_config.yaml
@@ -1,8 +1,8 @@
 moose_petsc:
-  - moose-petsc 3.20.0 build_2
+  - moose-petsc 3.20.1 build_0
 
 moose_libmesh_vtk:
-  - moose-libmesh-vtk 9.2.6 build_7
+  - moose-libmesh-vtk 9.3.0 build_0
 
 #### Darwin SDK SYSROOT
 CONDA_BUILD_SYSROOT:                                        # [osx]
