+++ conda/moose-dev/conda_build_config.yaml
@@ -1,11 +1,11 @@
 moose_libmesh:
-  - moose-libmesh 2023.11.06 build_1
+  - moose-libmesh 2023.11.28 build_0
 
 moose_wasp:
-  - moose-wasp 2023.11.06
+  - moose-wasp 2023.11.28
 
 moose_tools:
-  - moose-tools 2023.10.27
+  - moose-tools 2023.11.28
 
 moose_peacock:
-  - moose-peacock 2023.11.10
+  - moose-peacock 2023.11.28
