+++ conda/wasp/meta.yaml
@@ -5,7 +5,7 @@
 #
 # As well as any directions pertaining to modifying those files.
 # ALSO: Follow the directions in scripts/tests/versioner_hashes.yaml
-{% set version = "2023.11.06" %}
+{% set version = "2023.11.28" %}
 
 package:
   name: moose-wasp
