+++ framework/doc/packages_config.yml
@@ -1,9 +1,9 @@
-gcc: 10.4.0
-clang: 12.0.1
+gcc: 12.2.1
+clang: 16.0.6
 minimum_gcc: 7.5.0
 minimum_clang: 10.0.1
 minimum_python: 3.7
 mpich: 4.0.2
 petsc_alt: 3.11.4
-petsc: 3.20.0
-moose_dev: 2023.11.10
+petsc: 3.20.1
+moose_dev: 2023.11.28
