+++ conda/petsc/meta.yaml
@@ -7,8 +7,8 @@
 #
 # As well as any directions pertaining to modifying those files.
 # ALSO: Follow the directions in scripts/tests/versioner_hashes.yaml
-{% set build = 2 %}
-{% set version = "3.20.0" %}
+{% set build = 0 %}
+{% set version = "3.20.1" %}
 
 # permanent
 {% set strbuild = "build_" + build|string %}
