+++ conda/mpich/meta.yaml
@@ -6,7 +6,7 @@
 #
 # As well as any directions pertaining to modifying those files.
 # ALSO: Follow the directions in scripts/tests/versioner_hashes.yaml
-{% set build = 14 %}
+{% set build = 15 %}
 {% set version = "4.0.2" %}
 
 # permanent
