+++ modules/doc/content/newsletter/2023/2023_11.md
@@ -59,6 +59,36 @@ when transferring between applications.
 - Minor fixes to documentation, comments, and unit tests, including
   VexCL-enabled unit tests of MetaPhysicL
 
+### `2023.11.28` Update
+
+- `enum_to_string` and `string_to_enum` now initialize their maps via
+  static constructors, so they (and functions which depend on them,
+  like `Elem::get_info()` are now thread-safe.
+- Added `FEInterface::is_hierarchic()` shim.
+- Added `TypeTensor<T>::column()` extraction method.
+- Added `L2_LAGRANGE_VEC` finite element family, providing a space
+  which locally matches `LAGRANGE_VEC` but which is discontinuous
+  between elements.
+- Configure option `--with-xdr-include`, to make it easier to build
+  on systems with XDR support in non-default locations.
+- Configure now detects non-default $CC from PETSc when possible.
+- Configure now adds compiler flags to enable safe use of
+  Floating-Point Exceptions, even with newer clang/clang++ and
+  optimization enabled.
+- Support for reads of ExodusII meshes with the `WEDGE6` element string.
+- Support for Metis partitioning of meshes with overlapping nodes or
+  elements even with `unique_id()` disabled.
+- Updated to newer Nanoflann, switched to git submodule
+- PETSc now uses VecGhostGetGhostIS instead of
+  VecGetLocalToGlobalMapping for more efficiency on newer PETSc
+  versions.
+- Bugfix and test coverage for recent bug in reads of Abaqus meshes
+  with element sets.
+- Bugfix and test coverage for XDR writes of refined meshes with extra
+  element integers.
+- Fixes for `reduced_basis` EIM features.
+- Fixes for example codes triggering GCC 13 warnings.
+
 ## PETSc-level Changes
 
 ## Bug Fixes and Minor Enhancements
