### 2023-11-29 78dd9ed
initial working version of reporter that sums vector of vectors into a single vector
