+++ modules/optimization/test/include/reporters/VectorOfVectorTestReporter.h
@@ -1,0 +1,27 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "GeneralReporter.h"
+
+class VectorOfVectorTestReporter : public GeneralReporter
+{
+public:
+  static InputParameters validParams();
+
+  VectorOfVectorTestReporter(const InputParameters & parameters);
+
+  virtual void initialize() override {}
+  virtual void finalize() override {}
+  virtual void execute() override {}
+
+private:
+  std::vector<std::vector<Real>> & _vectors;
+};
