+++ modules/optimization/include/reporters/ReporterVectorSum.h
@@ -1,0 +1,29 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "GeneralReporter.h"
+
+class ReporterVectorSum : public GeneralReporter
+{
+public:
+  static InputParameters validParams();
+
+  ReporterVectorSum(const InputParameters & parameters);
+
+  virtual void initialize() override {}
+  virtual void initialSetup() override;
+  virtual void execute() override;
+  virtual void finalize() override {}
+
+private:
+  std::vector<double> & _summed_data;
+  const std::vector<std::vector<Real>> * _reporter_data;
+};
