### 2023-11-29 241d11b
Remove unneeded quadrature string/enum map (#26204)
