### 2023-11-29 482da56
Update test that references the gcc version
