+++ python/MooseDocs/test/gold/latex/extensions/package.tex
@@ -1,7 +1,7 @@
 
 \chapter{\label{package-extension}Package Extension}
 \begin{verbatim}
-/path/to/gcc-10.4.0
+/path/to/gcc-12.2.1
 \end{verbatim}
 
-\par This is a sentence with gcc-10.4.0.
+\par This is a sentence with gcc-12.2.1.
