+++ modules/navier_stokes/doc/content/source/fvbcs/INSFVTKEDWallFunctionBC.md
@@ -1,14 +1,14 @@
-# INSFVTKEDWallFunction
+# INSFVTKEDWallFunctionBC
 
-!syntax description /FVBCs/INSFVTKEDWallFunction
+!syntax description /FVBCs/INSFVTKEDWallFunctionBC
 
 This boundary condtion should only be used if no wall treatment is added.
 Implements wall boundary conditions for the turbulent kinetic energy dissipation rate.
 A separate treatment is used for the laminar and logarithmic layers.
 A linear blending functions is used to interpolate between both layers.
 
-!syntax parameters /FVBCs/INSFVTKEDWallFunction
+!syntax parameters /FVBCs/INSFVTKEDWallFunctionBC
 
-!syntax inputs /FVBCs/INSFVTKEDWallFunction
+!syntax inputs /FVBCs/INSFVTKEDWallFunctionBC
 
-!syntax children /FVBCs/INSFVTKEDWallFunction
+!syntax children /FVBCs/INSFVTKEDWallFunctionBC
