### 2023-11-29 915c463
Constrain OpenSSL
Constrain OpenSSL below 3.2.x for now.

Add necessary environment variables so we can build packages on our
machines.
