+++ conda/moose-dev/meta.yaml
@@ -20,7 +20,9 @@ build:
   number: 0
   string: build_0
   skip: true  # [win]
-
+  script_env:
+    - REQUESTS_CA_BUNDLE
+    - SSL_CERT_FILE
 requirements:
   run:
     - {{ moose_tools }}
