+++ conda/moose-dev/conda_build_config.yaml
@@ -5,7 +5,7 @@ moose_wasp:
   - moose-wasp 2023.11.29
 
 moose_tools:
-  - moose-tools 2023.10.27
+  - moose-tools 2023.11.28
 
 moose_peacock:
   - moose-peacock 2023.11.29
