### 2023-11-29 0675e22
Implement balance-momentum equations for explicit contact (#25666)
Kinematic constraint works for primary surface being rigid and works
also in parallel.

Requires verification and adding documentation.
