+++ modules/contact/src/constraints/ExplicitDynamicsContactConstraint.C
@@ -26,8 +26,6 @@
 #include "libmesh/string_to_enum.h"
 #include "libmesh/sparse_matrix.h"
 
-#include "UpdateNodalAuxVariable.h"
-
 registerMooseObject("ContactApp", ExplicitDynamicsContactConstraint);
 
 const unsigned int ExplicitDynamicsContactConstraint::_no_iterations = 0;
@@ -62,27 +60,17 @@ ExplicitDynamicsContactConstraint::validParams()
   params.addParam<Real>("friction_coefficient", 0, "The friction coefficient");
   params.addParam<Real>("tangential_tolerance",
                         "Tangential distance to extend edges of contact surfaces");
-  params.addParam<Real>(
-      "capture_tolerance", 0, "Normal distance from surface within which nodes are captured");
-  params.addParam<Real>("tension_release",
-                        0.0,
-                        "Tension release threshold.  A node in contact "
-                        "will not be released if its tensile load is below "
-                        "this value.  No tension release if negative.");
   params.addParam<bool>(
       "normalize_penalty",
       false,
       "Whether to normalize the penalty parameter with the nodal area for penalty contact.");
   params.addParam<bool>(
       "print_contact_nodes", false, "Whether to print the number of nodes in contact.");
-  // params.addParam<std::vector<UserObjectName>>(
-  //     "vel_uos", "List of nodal user objects to update the velocity vector.");
   params.addClassDescription(
       "Apply non-penetration constraints on the mechanical deformation "
       "using a node on face, primary/secondary algorithm, and multiple options "
       "for the physical behavior on the interface and the mathematical "
       "formulation for constraint enforcement");
-
   return params;
 }
 
@@ -91,13 +79,11 @@ Threads::spin_mutex ExplicitDynamicsContactConstraint::_contact_set_mutex;
 ExplicitDynamicsContactConstraint::ExplicitDynamicsContactConstraint(
     const InputParameters & parameters)
   : NodeFaceConstraint(parameters),
-    TwoMaterialPropertyInterface(this, Moose::EMPTY_BLOCK_IDS, getBoundaryIDs()),
+    TwoMaterialPropertyInterface(this, Moose::EMPTY_BLOCK_IDS, buildBoundaryIDs()),
     _displaced_problem(parameters.get<FEProblemBase *>("_fe_problem_base")->getDisplacedProblem()),
     _component(getParam<unsigned int>("component")),
     _model(getParam<MooseEnum>("model").getEnum<ExplicitDynamicsContactModel>()),
     _normalize_penalty(getParam<bool>("normalize_penalty")),
-    _tension_release(getParam<Real>("tension_release")),
-    _capture_tolerance(getParam<Real>("capture_tolerance")),
     _update_stateful_data(true),
     _mesh_dimension(_mesh.dimension()),
     _vars(3, libMesh::invalid_uint),
@@ -114,11 +100,10 @@ ExplicitDynamicsContactConstraint::ExplicitDynamicsContactConstraint(
     _aux_system(_nodal_area_var->sys()),
     _aux_solution(_aux_system.currentSolution()),
     _print_contact_nodes(getParam<bool>("print_contact_nodes")),
+    _residual_copy(_sys.residualGhosted()),
     _neighbor_density(getNeighborMaterialPropertyByName<Real>("density")),
     _neighbor_wave_speed(getNeighborMaterialPropertyByName<Real>("wave_speed")),
-    _vel_x(&writableVariable("vel_x")),
[…]
