+++ framework/include/systems/NonlinearSystemBase.h
@@ -983,4 +983,7 @@ private:
 
   /// The number of scaling groups
   std::size_t _num_scaling_groups;
+
+  /// Mutex for auxiliary variables
+  static Threads::spin_mutex writable_variable_mutex;
 };
