+++ modules/contact/src/userobjects/NodalWaveSpeed.C
@@ -80,21 +80,12 @@ NodalWaveSpeed::execute()
 void
 NodalWaveSpeed::finalize()
 {
-
   const std::map<const Node *, Real>::iterator it_end = _node_wave_speeds.end();
   for (std::map<const Node *, Real>::iterator it = _node_wave_speeds.begin(); it != it_end; ++it)
   {
     const Node * const node = it->first;
     dof_id_type dof = node->dof_number(_system.number(), _variable->number(), 0);
-    _aux_solution.set(dof, 0);
-  }
-  _aux_solution.close();
-
-  for (std::map<const Node *, Real>::iterator it = _node_wave_speeds.begin(); it != it_end; ++it)
-  {
-    const Node * const node = it->first;
-    dof_id_type dof = node->dof_number(_system.number(), _variable->number(), 0);
-    _aux_solution.add(dof, it->second);
+    _aux_solution.set(dof, it->second);
   }
   _aux_solution.close();
 }
