### 2023-11-29 cec79f4
Revert VTK bump and just don't build it with clang 16
