+++ conda/libmesh-vtk/meta.yaml
@@ -3,10 +3,10 @@
 #   libmesh/*
 #
 # As well as any directions pertaining to modifying those files.
-{% set build = 0 %}
-{% set vtk_version = "9.3.0" %}
-{% set vtk_friendly_version = "9.3" %}
-{% set sha256 = "fdc7b9295225b34e4fdddc49cd06e66e94260cb00efee456e0f66568c9681be9" %}
+{% set build = 8 %}
+{% set vtk_version = "9.2.6" %}
+{% set vtk_friendly_version = "9.2" %}
+{% set sha256 = "06fc8d49c4e56f498c40fcb38a563ed8d4ec31358d0101e8988f0bb4d539dd12" %}
 
 # permanent
 {% set strbuild = "build_" + build|string %}
