+++ modules/doc/content/newsletter/2023/2023_11.md
@@ -59,7 +59,7 @@ when transferring between applications.
 - Minor fixes to documentation, comments, and unit tests, including
   VexCL-enabled unit tests of MetaPhysicL
 
-### `2023.11.28` Update
+### `2023.11.29` Update
 
 - `enum_to_string` and `string_to_enum` now initialize their maps via
   static constructors, so they (and functions which depend on them,
