+++ conda/pyhit/conda_build_config.yaml
@@ -1,5 +1,5 @@
 moose_wasp:
-  moose-wasp 2023.11.28
+  moose-wasp 2023.11.29
 
 moose_python:
   - python 3.10
