+++ conda/moose-dev/meta.yaml
@@ -7,7 +7,7 @@
 #   framework/doc/packages_config.yml
 #
 # As well as any directions pertaining to modifying those files.
-{% set version = "2023.11.28" %}
+{% set version = "2023.11.29" %}
 
 package:
   name: moose-dev
