+++ conda/wasp/meta.yaml
@@ -5,7 +5,7 @@
 #
 # As well as any directions pertaining to modifying those files.
 # ALSO: Follow the directions in scripts/tests/versioner_hashes.yaml
-{% set version = "2023.11.28" %}
+{% set version = "2023.11.29" %}
 
 package:
   name: moose-wasp
