+++ conda/peacock/meta.yaml
@@ -3,7 +3,7 @@
 #   moose-dev/*
 #
 # As well as any directions pertaining to modifying those files.
-{% set version = "2023.11.28" %}
+{% set version = "2023.11.29" %}
 
 package:
   name: moose-peacock
