+++ framework/doc/packages_config.yml
@@ -6,4 +6,4 @@ minimum_python: 3.7
 mpich: 4.0.2
 petsc_alt: 3.11.4
 petsc: 3.20.1
-moose_dev: 2023.11.28
+moose_dev: 2023.11.29
