### 2023-11-29 366557f
Resolve warning and actually use the result
