### 2023-11-29 642dd1e
Add missing undefined template variable
