+++ framework/src/parser/CommandLine.C
@@ -253,7 +253,7 @@ CommandLine::getExecutableName() const
 {
   // Grab the first item out of argv
   std::string command(_args[0]);
-  return command.substr(command.find_last_of("/\\") + 1);;
+  return command.substr(command.find_last_of("/\\") + 1);
 }
 
 void
