### 2023-11-29 f842fc9
Adding documentation, testing and renaming object to be more descriptive.  This completes the implementation for the vectorOfVectorRowSum ref #26198
