### 2023-11-29 a7b2769
Limit openssl pinning to OSX platforms
