+++ conda/tools/meta.yaml
@@ -24,7 +24,7 @@ requirements:
     - setuptools
   host:
     # TODO: Try to remove when this version settles down among other conda-forge packages
-    - openssl <3.2.0a
+    - openssl <3.2.0a # [osx]
   run:
     - beautifulsoup4
     - {{ clang_format }}
@@ -57,7 +57,7 @@ requirements:
     - xmltodict
     - yaml
     # TODO: Try to remove when this version settles down among other conda-forge packages
-    - {{ pin_compatible('openssl', upper_bound='3.2') }}
+    - {{ pin_compatible('openssl', upper_bound='3.2') }} # [osx]
 test:
   commands:
     - clang-format --version
