+++ modules/navier_stokes/src/fvbcs/INSFVTurbulentViscosityWallFunction.C
@@ -10,6 +10,7 @@
 #include "INSFVTurbulentViscosityWallFunction.h"
 #include "Function.h"
 #include "NavierStokesMethods.h"
+#include "NS.h"
 
 registerMooseObject("NavierStokesApp", INSFVTurbulentViscosityWallFunction);
 
