### 2023-11-29 7c9abc4
Avoided calling build_side_list_from_node_list() in THM
Closes #26200
