+++ conda/libmesh/meta.yaml
@@ -6,7 +6,7 @@
 # ALSO: Follow the directions in scripts/tests/versioner_hashes.yaml
 {% set build = 0 %}
 {% set version = "2023.11.29" %}
-{% set vtk_friendly_version = "9.3" %}
+{% set vtk_friendly_version = "9.2" %}
 
 # permanent
 {% set strbuild = "build_" + build|string %}
