### 2023-11-29 c68f6cd
Now allow execute on FINAL.
