### 2023-11-29 3973e9a
Change Parserother to Builder, clean up Main.C and move to mooseMain.C, add docs
Co-Authored-By: Cody Permann <3036960+permcody@users.noreply.github.com>
Co-Authored-By: Logan Harbour <loganhharbour@gmail.com>
