+++ framework/src/actions/MeshOnlyAction.C
@@ -57,7 +57,7 @@ MeshOnlyAction::act()
 
   if (should_generate)
   {
-    mesh_file = _app.parserOther().getPrimaryFileName();
+    mesh_file = _app.builder().getPrimaryFileName();
     size_t pos = mesh_file.find_last_of('.');
 
     // Default to writing out an ExodusII mesh base on the input filename.
