+++ examples/ex04_bcs/src/main.C
@@ -17,10 +17,8 @@
 #include "ExampleApp.h"
 
 // Moose Includes
-#include "MooseInit.h"
-#include "MooseApp.h"
-#include "AppFactory.h"
-#include "MooseCreate.h"
+#include "Moose.h"
+#include "MooseMain.h"
 
 // Create a performance log
 PerfLog Moose::perf_log("Example");
@@ -29,20 +27,7 @@ PerfLog Moose::perf_log("Example");
 int
 main(int argc, char * argv[])
 {
-  // Initialize MPI, solvers and MOOSE
-  MooseInit init(argc, argv);
-
-  // Register this application's MooseApp and any it depends on
-  ExampleApp::registerApps();
-
-  // Instanitiate Moose App
-  MooseCreate create("ExampleApp", argc, argv);
-
-  // Get the created application;
-  std::shared_ptr<MooseApp> app = create.getApp();
-
-  // Execute the application
-  app->run();
+  moose::main<ExampleApp>(argc, argv);
 
   return 0;
 }
