+++ framework/src/base/AppFactory.C
@@ -38,7 +38,7 @@ MooseAppPtr
 AppFactory::createAppShared(const std::string & default_app_type,
                             int argc,
                             char ** argv,
-                            std::shared_ptr<Parser> parser,
+                            std::unique_ptr<Parser> parser,
                             MPI_Comm comm_world_in)
 {
   auto command_line = std::make_shared<CommandLine>(argc, argv);
@@ -55,14 +55,15 @@ AppFactory::createAppShared(const std::string & default_app_type,
   app_params.set<char **>("_argv") = argv;
   app_params.set<std::shared_ptr<CommandLine>>("_command_line") = command_line;
 
-  return AppFactory::instance().createShared(app_type, "main", app_params, parser, comm_world_in);
+  return AppFactory::instance().createShared(
+      app_type, "main", app_params, std::move(parser), comm_world_in);
 }
 
 MooseAppPtr
 AppFactory::createShared(const std::string & app_type,
                          const std::string & name,
                          InputParameters parameters,
-                         std::shared_ptr<Parser> parser,
+                         std::unique_ptr<Parser> parser,
                          MPI_Comm comm_world_in)
 {
   // Error if the application type is not located
@@ -76,7 +77,7 @@ AppFactory::createShared(const std::string & app_type,
 
   // Take the front parser and add it to the parameters so that it can be retrieved in the
   // Application
-  parameters.set<std::shared_ptr<Parser>>("_parser") = parser;
+  parameters.set<std::shared_ptr<Parser>>("_parser") = std::move(parser);
 
   // Check to make sure that all required parameters are supplied
   parameters.checkParams("");
