### 2023-11-29 e47d560
Clean format for precheck
Add missing header files

Add MooseApp header in MooseMain.h

Use new main.C in ex13 and ex15
