+++ framework/include/base/MooseMain.h
@@ -11,6 +11,7 @@
 
 #include "Moose.h"
 #include "MooseInit.h"
+#include "MooseApp.h"
 
 namespace moose
 {
