+++ framework/src/parser/Parser.C
@@ -16,6 +16,7 @@
 #include "CommandLine.h"
 #include "SystemInfo.h"
 #include "MooseUtils.h"
+#include "Parser.h"
 
 #include "libmesh/parallel.h"
 #include "libmesh/fparser.hh"
