+++ examples/ex13_functions/src/main.C
@@ -17,7 +17,7 @@
 #include "MooseInit.h"
 #include "MooseApp.h"
 #include "AppFactory.h"
-#include "MooseCreate.h"
+#include "MooseMain.h"
 
 // Create a performance log
 PerfLog Moose::perf_log("Example");
@@ -26,20 +26,7 @@ PerfLog Moose::perf_log("Example");
 int
 main(int argc, char * argv[])
 {
-  // Initialize MPI, solvers and MOOSE
-  MooseInit init(argc, argv);
-
-  // Register this application's MooseApp and any it depends on
-  ExampleApp::registerApps();
-
-  // Instanitiate Moose App
-  MooseCreate create("ExampleApp", argc, argv);
-
-  // Get the created application;
-  std::shared_ptr<MooseApp> app = create.getApp();
-
-  // Execute the application
-  app->run();
+  moose::main<ExampleApp>(argc, argv);
 
   return 0;
 }
