+++ examples/ex15_actions/src/main.C
@@ -28,7 +28,7 @@ PerfLog Moose::perf_log("Example");
 int
 main(int argc, char * argv[])
 {
-  moose::main<MooseTestApp>(argc, argv);
+  moose::main<ExampleApp>(argc, argv);
 
   return 0;
 }
