### 2023-11-30 bf79957
reporter for vector dot product closes #26198
