+++ framework/src/transfers/MultiAppGeneralFieldTransfer.C
@@ -260,6 +260,9 @@ MultiAppGeneralFieldTransfer::initialSetup()
                          " while there are " + std::to_string(getFromMultiApp()->numGlobalApps()) +
                          " subapps");
     }
+    else if (isParamSetByUser("from_mesh_division_usage"))
+      paramError("from_mesh_division_usage",
+                 "'from_mesh_division' must be specified if the usage method is specified");
   }
 
   // Loop over all target problems
@@ -320,6 +323,9 @@ MultiAppGeneralFieldTransfer::initialSetup()
                          " while there are " + std::to_string(getToMultiApp()->numGlobalApps()) +
                          " subapps");
     }
+    else if (isParamSetByUser("to_mesh_division_usage"))
+      paramError("to_mesh_division_usage",
+                 "'to_mesh_division' must be specified if the usage method is specified");
   }
 
   // Check if components are set correctly if using an array variable
@@ -625,8 +631,10 @@ MultiAppGeneralFieldTransfer::extractOutgoingPoints(const unsigned int var_index
         if (!_to_boundaries.empty() && !onBoundaries(_to_boundaries, to_moose_mesh, node))
           continue;
 
-        // Skip if the node is not indexed within the mesh division
-        if (!_to_mesh_divisions.empty() && !inMeshDivision(_to_mesh_divisions[i_to], *node))
+        // Skip if the node does not meet the target mesh division behavior
+        // We cannot know from which app the data will come from so we cannot know
+        // the source mesh division index and the source app global index
+        if (!_to_mesh_divisions.empty() && !inMeshDivision(*node, i_to, i_to, false))
           continue;
 
         // Cache point information
@@ -652,7 +660,7 @@ MultiAppGeneralFieldTransfer::extractOutgoingPoints(const unsigned int var_index
           continue;
 
         // Skip if the element is not indexed within the mesh division
-        if (!_to_mesh_divisions.empty() && !inMeshDivision(_to_mesh_divisions[i_to], *elem))
+        if (!_to_mesh_divisions.empty() && !inMeshDivision(*elem, i_to, i_to, false))
           continue;
 
         // Cache point information
@@ -716,7 +724,8 @@ MultiAppGeneralFieldTransfer::cacheIncomingInterpVals(
             "Higher order discontinuous elemental variables are not supported for target-boundary "
             "restricted transfers");
       // Not implemented. Unless we contain the entire mesh!
-      if (_to_mesh_divisions.size())
+      if (_to_mesh_divisions.size() &&
+          _to_mesh_division_behavior == MeshDivisionTransferUse::RESTRICTION)
         for (const auto mesh_div : _to_mesh_divisions)
           if (!mesh_div->coversEntireMesh())
             mooseError("Higher order variable support not implemented for target mesh division "
@@ -1325,24 +1334,8 @@ MultiAppGeneralFieldTransfer::acceptPointInOriginMesh(unsigned int i_from,
       return false;
 
     // Check point against the source mesh division
-    if (!_from_mesh_divisions.empty())
-    {
-      const auto source_division_index =
-          _from_mesh_divisions[i_from]->divisionIndex(transformed_pt);
-      // If the point is not indexed in the division and the division is used as a restriction
-      if (_from_mesh_division_behavior == MeshDivisionTransferUse::RESTRICTION &&
-          source_division_index == MooseMeshDivision::INVALID_DIVISION_INDEX)
-        return false;
-      // If the point is not the at the same index in the target and the origin meshes, reject
-      else if (_from_mesh_division_behavior == MeshDivisionTransferUse::MATCH_DIVISION_INDEX &&
-               mesh_div != source_division_index)
-        return false;
-      // If the point is at a certain division index that is not the same as the index of the subapp
[…]
