### 2023-11-30 6f4ce8a
Add more options for source app determination algorithm - bbox is existing option - greedy search more explicitly greedy (only in debug mode? it would loose too much scalability otherwise?) - source mesh divisions, various options refs #25883
