### 2023-11-30 a921b80
Fix typos, add docstring to conservative transfers
