### 2023-11-30 #6045 @kyleniemeyer
Hello @andrewryh and @atismer, could you edit this submission to JOSS on the thermal hydraulics module in MOOSE?
