### 2023-11-30 601f2f6
Add utility routine for global to local This should almost never be used Use new utilities in legacy code
