### 2023-11-30 6c492e8
Add more options for the algorithm to select which processes should send data refs #25883
