+++ framework/src/transfers/MultiAppGeneralFieldTransfer.C
@@ -207,6 +207,12 @@ MultiAppGeneralFieldTransfer::MultiAppGeneralFieldTransfer(const InputParameters
       mooseWarning(
           "Extrapolation (nearest-source options, outside-app source) parameters have "
           "been passed, but no subapp bounding box expansion parameters have been passed.");
+
+  if (!_use_bounding_boxes &&
+      (isParamValid("fixed_bounding_box_size") || isParamSetByUser("bbox_factor")))
+    paramError("use_bounding_boxes",
+               "Cannot pass additional bounding box parameters (sizes, expansion, etc) if we are "
+               "not using bounding boxes");
 }
 
 void
@@ -255,6 +261,7 @@ MultiAppGeneralFieldTransfer::initialSetup()
                    "Source mesh division cannot match target mesh division if no target mesh "
                    "division is specified");
       else if (_from_mesh_division_behavior == MeshDivisionTransferUse::MATCH_SUBAPP_INDEX)
+      {
         if (!hasFromMultiApp())
           paramError("from_mesh_division_usage",
                      "Cannot match source mesh division index to source subapp index if there is "
@@ -266,6 +273,7 @@ MultiAppGeneralFieldTransfer::initialSetup()
                          std::to_string(_from_mesh_divisions[i_from]->getNumDivisions()) +
                          " while there are " + std::to_string(getFromMultiApp()->numGlobalApps()) +
                          " subapps");
+      }
     }
     else if (isParamSetByUser("from_mesh_division_usage"))
       paramError("from_mesh_division_usage",
@@ -319,6 +327,7 @@ MultiAppGeneralFieldTransfer::initialSetup()
                        "ensure appropriate use");
       }
       else if (_to_mesh_division_behavior == MeshDivisionTransferUse::MATCH_SUBAPP_INDEX)
+      {
         if (!hasToMultiApp())
           paramError("to_mesh_division_usage",
                      "Cannot match target division to target subapp index if there is only one "
@@ -329,6 +338,7 @@ MultiAppGeneralFieldTransfer::initialSetup()
                          std::to_string(_to_mesh_divisions[i_to]->getNumDivisions()) +
                          " while there are " + std::to_string(getToMultiApp()->numGlobalApps()) +
                          " subapps");
+      }
     }
     else if (isParamSetByUser("to_mesh_division_usage"))
       paramError("to_mesh_division_usage",
@@ -413,15 +423,15 @@ MultiAppGeneralFieldTransfer::transferVariable(unsigned int i)
   mooseAssert(i < _var_size, "The variable of index " << i << " does not exist");
 
   // Get the bounding boxes for the "from" domains.
-  // Clean up _bboxes
-  _bboxes.clear();
+  // Clean up _from_bboxes
+  _from_bboxes.clear();
 
   // NOTE: This ignores the app's bounding box inflation and padding
-  _bboxes = getRestrictedFromBoundingBoxes();
+  _from_bboxes = getRestrictedFromBoundingBoxes();
 
   // Expand bounding boxes. Some desired points might be excluded
   // without an expansion
-  extendBoundingBoxes(_bbox_factor, _bboxes);
+  extendBoundingBoxes(_bbox_factor, _from_bboxes);
 
   // Figure out how many "from" domains each processor owns.
   _froms_per_proc.clear();
@@ -498,32 +508,88 @@ MultiAppGeneralFieldTransfer::locatePointReceivers(const Point point,
   // A point may be close enough to several problems, hosted on several processes
   bool found = false;
 
-  // We examine all bounding boxes and find the maximum distance within a bounding box
-  // from the point. This creates a sphere around the point of interest. Any app with a bounding
[…]
