+++ framework/src/transfers/MultiAppGeneralFieldUserObjectTransfer.C
@@ -71,7 +71,8 @@ MultiAppGeneralFieldUserObjectTransfer::prepareEvaluationOfInterpValues(
     const unsigned int /* var_index */)
 {
   _local_bboxes.clear();
-  extractLocalFromBoundingBoxes(_local_bboxes);
+  if (_use_bounding_boxes)
+    extractLocalFromBoundingBoxes(_local_bboxes);
 }
 
 void
