### 2023-11-30 6b7502f
More locations which need WASP
Adding two more locations where I think we need to supply instruction
for building WASP.
