### 2023-11-30 f36b5b1
Add missing header, update test file
Remove  #pragma once
