+++ framework/src/base/MooseMain.C
@@ -7,7 +7,6 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#pragma once
 #include "MooseMain.h"
 #include "ParallelUniqueId.h"
 #include "Parser.h"
