### 2023-11-30 cda7dbd
Update include/problems/SubChannel1PhaseProblem.h
Co-authored-by: Mauricio Tano Retamales <Mauricio.TanoRetamales@inl.gov>
