### 2023-11-30 15a17b4
Update src/base/SubChannelApp.C
Co-authored-by: Mauricio Tano Retamales <Mauricio.TanoRetamales@inl.gov>
