### 2023-11-30 e17f1d8
Update src/problems/LiquidMetalSubChannel1PhaseProblem.C
Co-authored-by: Mauricio Tano Retamales <Mauricio.TanoRetamales@inl.gov>
