### 2023-11-30 a8259f6
Update src/problems/LiquidMetalSubChannel1PhaseProblem.C
Co-authored-by: Mauricio Tano Retamales <Mauricio.TanoRetamales@inl.gov>
