### 2023-11-30 6d2eb90
Update src/problems/SubChannel1PhaseProblem.C
Co-authored-by: Mauricio Tano Retamales <Mauricio.TanoRetamales@inl.gov>
