### 2023-11-30 05fb4ab
adding reporter to sum components of a vector
