### 2023-11-30 99df9fe
Add alternate function to help MOOSE application adapting to main function changes in framework
Clean format for precheck
