### 2023-12-01 7ee4c68
Update createAppShared in module unit test
