### 2023-12-01 dcc5b98
Fixup exception checks Add additional behavior on inMeshDivision to handle having passed a subapp/division index for the target and comparing it in the source refs #25883
