+++ framework/src/transfers/MultiAppGeneralFieldTransfer.C
@@ -79,15 +79,15 @@ MultiAppGeneralFieldTransfer::validParams()
                                     "Mesh division object on the origin application");
   params.addParam<MeshDivisionName>("to_mesh_division",
                                     "Mesh division object on the target application");
-  MooseEnum mesh_division_uses("spatial_restriction matching_divisions matching_subapp_index",
-                               "matching_subapp_index");
+  MooseEnum mesh_division_uses("spatial_restriction matching_division matching_subapp_index none",
+                               "none");
   params.addParam<MooseEnum>("from_mesh_division_usage",
                              mesh_division_uses,
                              "How to use the source mesh division in the transfer. See object "
                              "documentation for description of each option");
   params.addParam<MooseEnum>("to_mesh_division_usage",
                              mesh_division_uses,
-                             "How to use the source mesh division in the transfer. See object "
+                             "How to use the target mesh division in the transfer. See object "
                              "documentation for description of each option");
 
   // Array and vector variables
@@ -160,7 +160,7 @@ MultiAppGeneralFieldTransfer::MultiAppGeneralFieldTransfer(const InputParameters
             : nullptr),
     _source_app_must_contain_point(getParam<bool>("from_app_must_contain_point")),
     _from_mesh_division_behavior(getParam<MooseEnum>("from_mesh_division_usage")),
-    _to_mesh_division_behavior(getParam<MooseEnum>("from_mesh_division_usage")),
+    _to_mesh_division_behavior(getParam<MooseEnum>("to_mesh_division_usage")),
     _elemental_boundary_restriction_on_sides(
         getParam<MooseEnum>("elemental_boundary_restriction") == "sides"),
     _greedy_search(getParam<bool>("greedy_search")),
@@ -248,13 +248,15 @@ MultiAppGeneralFieldTransfer::initialSetup()
       const auto & mesh_div_name = getParam<MeshDivisionName>("from_mesh_division");
       _from_mesh_divisions.push_back(&_from_problems[i_from]->getMeshDivision(mesh_div_name));
       // Check that the behavior set makes sense
-      if (_from_mesh_division_behavior == MeshDivisionTransferUse::RESTRICTION &&
-          _from_mesh_divisions[i_from]->coversEntireMesh())
-        mooseInfo("'from_mesh_division_usage' is set to use a spatial restriction but the "
-                  "'from_mesh_division' for source app of global index " +
-                  std::to_string(getGlobalSourceAppIndex(i_from)) +
-                  " covers the entire mesh. Do not expect any restriction from a mesh "
-                  "division that covers the entire mesh");
+      if (_from_mesh_division_behavior == MeshDivisionTransferUse::RESTRICTION)
+      {
+        if (_from_mesh_divisions[i_from]->coversEntireMesh())
+          mooseInfo("'from_mesh_division_usage' is set to use a spatial restriction but the "
+                    "'from_mesh_division' for source app of global index " +
+                    std::to_string(getGlobalSourceAppIndex(i_from)) +
+                    " covers the entire mesh. Do not expect any restriction from a mesh "
+                    "division that covers the entire mesh");
+      }
       else if (_from_mesh_division_behavior == MeshDivisionTransferUse::MATCH_DIVISION_INDEX &&
                !isParamValid("to_mesh_division"))
         paramError("to_mesh_division_usage",
@@ -262,21 +264,29 @@ MultiAppGeneralFieldTransfer::initialSetup()
                    "division is specified");
       else if (_from_mesh_division_behavior == MeshDivisionTransferUse::MATCH_SUBAPP_INDEX)
       {
-        if (!hasFromMultiApp())
[…]
