+++ framework/doc/content/source/transfers/MultiAppGeneralFieldTransfer.md
@@ -29,6 +29,7 @@ All transfers derived from this base class should be able to support:
 
 - block restriction in both the source and target application
 - boundary restriction in both the source and target application
+- [mesh division](syntax/MeshDivisions/index.md) restriction in both the source and target application
 - arbitrary number of parallel processes for both the source and target application
 - support for replicated and distributed meshes in both applications involved
 - transfers between parent and child applications
@@ -42,6 +43,9 @@ All transfers derived from this base class should be able to support:
 - interpolation and extrapolation transfers, as defined by the derived class
 - detection of indetermination due to source points equidistant to a target point
 - limitation of transfer source to the nearest position (see [Positions](syntax/Positions/index.md)) of target point
+- limitation of transfer source to the matching [mesh division](syntax/MeshDivisions/index.md) index between the source and target mesh divisions
+- limitation of transfer source to subapps at the same index as the target mesh division
+- limitation of transfer target to subapps at the same index as the source mesh division
 - general coordinate transformations. Coordinate system changes (RZ to XYZ, for example) are not
   fully supported for the "floating point precision indetermination" detection.
 
@@ -118,6 +122,56 @@ The "nearest position" criterion for the source of a transfer is obeyed strictly
 an invalid value (triggering the use of the [!param](/Transfers/MultiAppGeneralFieldUserObjectTransfer/extrapolation_constant))
 should and will be preferred over a valid value.
 
+!alert note
+The origin and target locations for transferred data are both considered in the reference
+domain (the one obtained by applying the coordinate transformation) when using the nearest positions
+transfer options.
+
+## Using mesh divisions in transfers
+
+[Mesh divisions](syntax/MeshDivisions/index.md)
+may be leveraged in several ways by specifying the [!param](/Transfers/MultiAppGeneralFieldNearestLocationTransfer/from_mesh_division_usage)
+and the [!param](/Transfers/MultiAppGeneralFieldNearestLocationTransfer/to_mesh_division_usage)
+parameters.
+
+- Mesh divisions for spatial transfer restriction (usage = 'spatial_restriction')
+
+Source and target mesh divisions, specified using the [!param](/Transfers/MultiAppGeneralFieldNearestLocationTransfer/from_mesh_division) and [!param](/Transfers/MultiAppGeneralFieldNearestLocationTransfer/to_mesh_division) parameters respectively, can be used
+to limit the spatial domain that will provide data for the transfer (origin spatial restriction)
+and the target spatial domain. The exclusion of a region is simply the region that is indexed with an invalid
+index in the mesh division.
+
+!alert note
+The spatial restriction effect of the mesh divisions is active for all usages!
+If some data lies outside the source mesh division, it will not be transferred.
+
+- Matching regions using a source and target mesh divisions (usage = 'matching_division')
+
+The source domain for the values to be transferred and the target domain can also be matched on
+a one-to-one basis using the [!param](/Transfers/MultiAppGeneralFieldNearestLocationTransfer/from_mesh_division) and [!param](/Transfers/MultiAppGeneralFieldNearestLocationTransfer/to_mesh_division) parameters.
+Each region of a given index into a division is matched to the region of the same index into the other
+division.
+
+It is advised to keep the number of mesh divisions the same in the two mesh divisions for simplicity.
+
+- Matching source applications and target regions using a mesh division
+
+Set [!param](/Transfers/MultiAppGeneralFieldNearestLocationTransfer/to_mesh_division_usage) to 'matching_subapp_index'
+
+The application providing values to transfer to a target region can be restricted to the
+index of the points in the target region given by a mesh division specified by the [!param](/Transfers/MultiAppGeneralFieldNearestLocationTransfer/to_mesh_division) parameter.
+
[…]
