+++ modules/heat_transfer/src/base/HeatTransferApp.C
@@ -19,6 +19,7 @@ HeatTransferApp::validParams()
   InputParameters params = MooseApp::validParams();
 
   params.set<bool>("use_legacy_material_output") = false;
+  params.set<bool>("use_legacy_initial_residual_evaluation_bahavior") = false;
 
   return params;
 }
