+++ modules/combined/src/base/CombinedApp.C
@@ -47,6 +47,7 @@ CombinedApp::validParams()
 
   params.set<bool>("automatic_automatic_scaling") = false;
   params.set<bool>("use_legacy_material_output") = false;
+  params.set<bool>("use_legacy_initial_residual_evaluation_bahavior") = false;
 
   return params;
 }
