+++ modules/contact/src/base/ContactApp.C
@@ -20,6 +20,7 @@ ContactApp::validParams()
 
   params.set<bool>("automatic_automatic_scaling") = false;
   params.set<bool>("use_legacy_material_output") = false;
+  params.set<bool>("use_legacy_initial_residual_evaluation_bahavior") = false;
 
   return params;
 }
