+++ modules/peridynamics/src/base/PeridynamicsApp.C
@@ -21,6 +21,7 @@ PeridynamicsApp::validParams()
   params.set<bool>("automatic_automatic_scaling") = false;
 
   params.set<bool>("use_legacy_material_output") = false;
+  params.set<bool>("use_legacy_initial_residual_evaluation_bahavior") = false;
 
   return params;
 }
