+++ modules/fsi/src/base/FsiApp.C
@@ -11,6 +11,7 @@ FsiApp::validParams()
 {
   InputParameters params = MooseApp::validParams();
 
+  params.set<bool>("use_legacy_initial_residual_evaluation_bahavior") = false;
   return params;
 }
 
