+++ modules/doc/content/newsletter/2023/2023_11.md
@@ -125,6 +125,31 @@ in the future.
 
 ## PETSc-level Changes
 
+### PETSc updated to 3.20.1
+
+The `moose-petsc` package has been updated alongside the PETSc submodule to version 3.20.1. To update
+your older MOOSE environment to include this change, please activate your MOOSE conda environment and
+perform the command:
+
+```
+mamba install moose-dev=2023.11.30
+```
+
+The MOOSE package versions and builds included in this environment version are the following:
+
+```
+Package                  Version           Build
+===================================================
+moose-dev                2023.11.30        build_0
+moose-libmesh            2023.11.29        build_0
+moose-libmesh-vtk        9.2.6             build_8
+moose-mpich              4.0.2             build_15
+moose-peacock            2023.11.29        hb4c3223_0
+moose-petsc              3.20.1            build_0
+moose-tools              2023.11.28        hd1db04d_0
+moose-wasp               2023.11.29        build_0
+```
+
 ## Bug Fixes and Minor Enhancements
 
 - A DirichletBC leveraging functors was added for finite element variables: [FunctorDirichletBC.md].
