### 2023-12-01 eb323ef
Always do the redundant residual evaluation in legacy mode
