+++ framework/src/outputs/ConsoleUtils.C
@@ -404,6 +404,21 @@ outputLegacyInformation(MooseApp & app)
         << COLOR_DEFAULT << std::endl;
   }
 
+  if (app.parameters().get<bool>("use_legacy_initial_residual_evaluation_bahavior"))
+  {
+    oss << COLOR_RED << "LEGACY MODES ENABLED:" << COLOR_DEFAULT << '\n';
+    oss << " This application uses the legacy initial residual evaluation behavior. The legacy "
+           "behavior performs an often times redundant residual evaluation before the solution "
+           "modifying objects are executed prior to the initial (0th nonlinear iteration) residual "
+           "evaluation. The new behavior skips that redundant residual evaluation unless the "
+           "parameter Executioner/use_fnorm0_before_smo is set to true. To remove this message and "
+           "enable the new behavior, set the parameter "
+           "'use_legacy_initial_residual_evaluation_bahavior' to false in *App.C. Some tests that "
+           "rely on the side effects of the legacy behavior may fail/diff and should be "
+           "re-golded.\n"
+        << COLOR_DEFAULT << std::endl;
+  }
+
   return oss.str();
 }
 
