### 2023-12-01 1aa4c40
Now the legacy mode should be fully working
