### 2023-12-01 a9f01b8
Use -1 for the invalid division index when outputting refs #25883
