+++ modules/doc/content/newsletter/2023/2023_11.md
@@ -5,7 +5,7 @@
 ### MeshDivisions system to split the mesh in arbitrary ways
 
 The [MeshDivisions](syntax/MeshDivisions/index.md) system has been added to MOOSE.
-It was released with a number of simple examples, including [SubdomainDivision.md] for a block ID-based
+It was released with a number of simple examples, including [SubdomainsDivision.md] for a block ID-based
 division of the mesh and the [CartesianGridDivision.md] for a Cartesian grid-numbering overlaid on top
 of the mesh.
 It is intended to be able to specify arbitrary mappings in transfers between applications.
