### 2023-12-02 8bb7820
Correctly use initial positions for the number of positions refs #25883
