### 2023-12-02 fe4a84b
Fix support for nearest location transfer + mesh divisions: the problem was that we want to match mesh divisions index, but we lose track of where the data stored in the KDTree came from We could only look at the target point and see if it was in the right source division. But if that target point was outside the source mesh, just like in many uses of the nearest-location transfer, then we dont get a useful index for the source division. So we HAVE TO use the source mesh division on the data we feed into the KDTree.
Now the problem is that if we only have N_division KDTree (a parallel to have 
[…]
