+++ framework/src/transfers/MultiAppGeneralFieldNearestLocationTransfer.C
@@ -52,6 +52,13 @@ MultiAppGeneralFieldNearestLocationTransfer::validParams()
   params.addParam<std::vector<MooseEnum>>(
       "source_type", source_types, "Where to get the source values from for each source variable");
 
+  // choose whether to include data from multiple apps when performing nearest-position/
+  // mesh-divisions based transfers
+  params.addParam<bool>("group_subapps",
+                        false,
+                        "Whether to group data from multiple subapps "
+                        "when considering a nearest-position algorithm");
+
   return params;
 }
 
@@ -59,13 +66,26 @@ MultiAppGeneralFieldNearestLocationTransfer::MultiAppGeneralFieldNearestLocation
     const InputParameters & parameters)
   : MultiAppGeneralFieldTransfer(parameters),
     SolutionInvalidInterface(this),
-    _num_nearest_points(getParam<unsigned int>("num_nearest_points"))
+    _num_nearest_points(getParam<unsigned int>("num_nearest_points")),
+    _group_subapps(getParam<bool>("group_subapps"))
 {
   if (_source_app_must_contain_point && _nearest_positions_obj)
     paramError("use_nearest_position",
                "We do not support using both nearest positions matching and checking if target "
                "points are within an app domain because the KDTrees for nearest-positions matching "
                "are (currently) built with data from multiple applications.");
+
+  // Parameter checks on grouping subapp values
+  if (_group_subapps && _from_mesh_divisions.empty() && !_nearest_positions_obj)
+    paramError(
+        "group_subapps",
+        "This option is only available for using mesh divisions or nearest positions regions");
+  else if (_group_subapps &&
+           (_from_mesh_division_behavior == MeshDivisionTransferUse::MATCH_DIVISION_INDEX ||
+            _from_mesh_division_behavior == MeshDivisionTransferUse::MATCH_SUBAPP_INDEX))
+    paramError("group_subapps",
+               "Cannot group subapps when considering nearest-location data as we would loose "
+               "track of the division index of the source locations");
 }
 
 void
@@ -158,6 +178,14 @@ MultiAppGeneralFieldNearestLocationTransfer::initialSetup()
           "'nearest' to a node. Please consider using a ProjectionAux to build a nodal source "
           "variable (for example linear Lagrange) before transferring");
   }
+
+  // We need to improve the indexing if we are to allow this
+  if (!_from_mesh_divisions.empty())
+    for (const auto mesh_div : _from_mesh_divisions)
+      if (mesh_div->getNumDivisions() != _from_mesh_divisions[0]->getNumDivisions())
+        paramError("from_mesh_division",
+                   "This transfer has only been implemented with a uniform number of source mesh "
+                   "divisions across all source applications");
 }
 
 void
@@ -173,27 +201,24 @@ MultiAppGeneralFieldNearestLocationTransfer::prepareEvaluationOfInterpValues(
 void
 MultiAppGeneralFieldNearestLocationTransfer::buildKDTrees(const unsigned int var_index)
 {
-  const unsigned int num_sources =
-      _nearest_positions_obj ? _nearest_positions_obj->getPositions(/*initial=*/false).size()
-                             : _from_problems.size();
-
+  const auto num_sources = getNumSources();
+  const auto num_apps_per_tree = getNumAppsPerTree();
   _local_kdtrees.resize(num_sources);
   _local_points.resize(num_sources);
   _local_values.resize(num_sources);
   unsigned int max_leaf_size = 0;
 
-  // Construct a local KDTree for each source (subapp or positions)
+  // Construct a local KDTree for each source. A source can be a single app or multiple apps
+  // combined (option for nearest-position / mesh-divisions)
[…]
