+++ framework/include/transfers/MultiAppGeneralFieldNearestLocationTransfer.h
@@ -60,6 +60,30 @@ private:
       const std::vector<std::pair<Point, unsigned int>> & incoming_points,
       std::vector<std::pair<Real, Real>> & outgoing_vals);
 
+  /// Number of KDTrees used to hold the locations & variable value data
+  unsigned int getNumSources() const;
+
+  /// Transform a point towards the local frame
+  Point getPointInLocalSourceFrame(unsigned int i_from, const Point & pt) const;
+
+  /// Number of applications which contributed nearest-locations to each KD-tree
+  unsigned int getNumAppsPerTree() const;
+
+  /// Number of divisions (nearest-positions or source mesh divisions) used when building KD-Trees
+  unsigned int getNumDivisions() const;
+
+  /**
+   * @brief Examine all spatial restrictions that could preclude this source from being
+   *        a valid source for this point
+   * @param pt point of interest
+   * @param valid_mesh_div if using source mesh divisions in a 'matching' mode, the point can only
+   * be used if coming from the relevant match
+   * @param i_from index of the source (KDTree+values) of interest
+   */
+  bool checkRestrictionsForSource(const Point & pt,
+                                  const unsigned int valid_mesh_div,
+                                  const unsigned int i_from) const;
+
   /// KD-Trees for all the local source apps
   std::vector<std::shared_ptr<KDTree>> _local_kdtrees;
 
@@ -74,6 +98,9 @@ private:
 
   /// Number of points to consider
   unsigned int _num_nearest_points;
+
+  /// Whether to group data when creating the nearest-point regions
+  const bool _group_subapps;
 
   /// Whether the source of the values is at nodes (true) or centroids (false) for each variable
   std::vector<bool> _source_is_nodes;
