### 2023-12-02 30971b2
Remove a case we likely will never use for shape-evaluation Shape evaluation are local, we dont need to match indexes refs #25883
