### 2023-12-02 e3a8049
Default to new behavior to see what apps fail
