+++ framework/src/base/MooseApp.C
@@ -338,7 +338,7 @@ MooseApp::validParams()
       "Set false to allow material properties to be output on INITIAL, not just TIMESTEP_END.");
   params.addParam<bool>(
       "use_legacy_initial_residual_evaluation_bahavior",
-      true,
+      false,
       "The legacy behavior performs an often times redundant residual evaluation "
       "before the solution modifying objects are executed prior to the initial "
       "(0th nonlinear iteration) residual evaluation. The new behavior skips that redundant "
