+++ framework/src/meshdivisions/SphericalGridDivision.C
@@ -9,6 +9,8 @@
 
 #include "SphericalGridDivision.h"
 #include "MooseMesh.h"
+#include "FEProblemBase.h"
+#include "Positions.h"
 
 #include "libmesh/elem.h"
 #include <math.h>
@@ -21,8 +23,9 @@ SphericalGridDivision::validParams()
   InputParameters params = MeshDivision::validParams();
   params.addClassDescription("Divide the mesh along a spherical grid.");
 
-  // Definition of the sphere
-  params.addRequiredParam<Point>("center", "Center of the sphere");
+  // Definition of the sphere(s)
+  params.addParam<Point>("center", "Center of the sphere");
+  params.addParam<PositionsName>("center_positions", "Positions of the centers of the spheres");
 
   // Spatial bounds of the sphere
   params.addRangeCheckedParam<Real>(
@@ -43,13 +46,21 @@ SphericalGridDivision::validParams()
 
 SphericalGridDivision::SphericalGridDivision(const InputParameters & parameters)
   : MeshDivision(parameters),
-    _center(getParam<Point>("center")),
+    _center(isParamValid("center") ? &getParam<Point>("center") : nullptr),
+    _center_positions(
+        isParamValid("center_positions")
+            ? &_fe_problem->getPositionsObject(getParam<PositionsName>("center_positions"))
+            : nullptr),
     _min_r(getParam<Real>("r_min")),
     _max_r(getParam<Real>("r_max")),
     _n_radial(getParam<unsigned int>("n_radial")),
     _outside_grid_counts_as_border(getParam<bool>("assign_domain_outside_grid_to_border"))
 {
   SphericalGridDivision::initialize();
+
+  // Check that we know the centers
+  if (!_center && !_center_positions)
+    paramError("center", "You must pass a parameter for the center of the spherical frame");
 
   // Check non-negative size
   if (_max_r < _min_r)
@@ -77,7 +88,17 @@ SphericalGridDivision::divisionIndex(const Point & pt) const
 {
   // Compute coordinates of the point in the spherical coordinates
   Point pc;
-  pc(0) = (pt - _center).norm();
+  unsigned int offset = 0;
+  if (_center)
+    pc(0) = (pt - *_center).norm();
+  else
+  {
+    // If distributing using positions, find the closest position
+    bool initial = _fe_problem->getCurrentExecuteOnFlag() == EXEC_INITIAL;
+    const auto nearest_center_index = _center_positions->getNearestPositionIndex(pt, initial);
+    offset = nearest_center_index * getNumDivisions();
+    pc(0) = (pt - _center_positions->getPosition(nearest_center_index, initial)).norm();
+  }
 
   if (!_outside_grid_counts_as_border)
   {
@@ -114,5 +135,5 @@ SphericalGridDivision::divisionIndex(const Point & pt) const
     ir = 0;
 
   mooseAssert(ir != not_found, "We should have found a mesh division bin radially");
-  return ir;
+  return offset + ir;
 }

[…]
