### 2023-12-02 4548fbb
Add unit tests to the reactor module to prepare for hexlattice tests, refs #26225
