### 2023-12-02 aca0030
Move unit test for Cardinal hex lattice utility refs #26225
