+++ modules/reactor/unit/src/HexagonalLatticeTest.C
@@ -1,0 +1,1848 @@
+/********************************************************************/
+/*                  SOFTWARE COPYRIGHT NOTIFICATION                 */
+/*                             Cardinal                             */
+/*                                                                  */
+/*                  (c) 2021 UChicago Argonne, LLC                  */
+/*                        ALL RIGHTS RESERVED                       */
+/*                                                                  */
+/*                 Prepared by UChicago Argonne, LLC                */
+/*               Under Contract No. DE-AC02-06CH11357               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*             Prepared by Battelle Energy Alliance, LLC            */
+/*               Under Contract No. DE-AC07-05ID14517               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*                 See LICENSE for full restrictions                */
+/********************************************************************/
+
+#include "HexagonalLatticeTest.h"
+
+TEST_F(HexagonalLatticeTest, rings_and_pins)
+{
+  Real bp = 10.0, pp = 0.1, pd = 0.04, wd = 0.01, wp = 50.0;
+  unsigned int nr = 1, a = 2;
+  HexagonalLatticeUtils hl1(bp, pp, pd, wd, wp, nr, a);
+
+  // Test number of rings given number of pins
+  EXPECT_EQ(hl1.rings(1), 1);
+  EXPECT_EQ(hl1.rings(7), 2);
+  EXPECT_EQ(hl1.rings(19), 3);
+  EXPECT_EQ(hl1.rings(37), 4);
+
+  try
+  {
+    hl1.rings(100);
+    FAIL() << "missing expected error";
+  }
+  catch (const std::exception & e)
+  {
+    std::string msg(e.what());
+    ASSERT_NE(msg.find("Number of pins 100 not evenly divisible in a hexagonal lattice!"),
+              std::string::npos)
+        << "failed with unexpected error: " << msg;
+  }
+
+  // Test number of pins in given ring
+  EXPECT_EQ(hl1.pins(1), 1);
+  EXPECT_EQ(hl1.pins(2), 6);
+  EXPECT_EQ(hl1.pins(3), 12);
+
+  // Test first and last pins in given ring
+  EXPECT_EQ(hl1.firstPinInRing(1), 0);
+  EXPECT_EQ(hl1.lastPinInRing(1), 0);
+
+  EXPECT_EQ(hl1.firstPinInRing(2), 1);
+  EXPECT_EQ(hl1.lastPinInRing(2), 6);
+
[…]
