### 2023-12-02 0e5d629
adding remaining mixture model capability Refs #22637
