+++ modules/navier_stokes/src/fvkernels/INSFVMomentumDriftFlux.C
@@ -1,0 +1,144 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "INSFVMomentumDriftFlux.h"
+#include "INSFVRhieChowInterpolator.h"
+#include "NS.h"
+#include "SystemBase.h"
+#include "RelationshipManager.h"
+#include "Factory.h"
+
+registerMooseObject("NavierStokesApp", INSFVMomentumDriftFlux);
+
+InputParameters
+INSFVMomentumDriftFlux::validParams()
+{
+  auto params = INSFVFluxKernel::validParams();
+  params.addClassDescription(
+      "Implements the drift momentum flux source.");
+  params.addRequiredParam<MooseFunctorName>("u_slip", "The velocity in the x direction.");
+  params.addParam<MooseFunctorName>("v_slip", "The velocity in the y direction.");
+  params.addParam<MooseFunctorName>("w_slip", "The velocity in the z direction.");
+  params.addRequiredParam<MooseFunctorName>("rho_d", "Dispersed phase density.");
+  params.addParam<MooseFunctorName>("fd", 0.0, "Fraction dispersed phase.");
+
+  MooseEnum coeff_interp_method("average harmonic", "harmonic");
+  params.addParam<MooseEnum>("density_interp_method",
+                             coeff_interp_method,
+                             "Switch that can select face interpolation method for the density.");
+
+  params.set<unsigned short>("ghost_layers") = 2;
+  return params;
+}
+
+INSFVMomentumDriftFlux::INSFVMomentumDriftFlux(const InputParameters & params)
+  : INSFVFluxKernel(params),
+    _dim(_subproblem.mesh().dimension()),
+    _rho_d(getFunctor<ADReal>("rho_d")),
+    _f_d(getFunctor<ADReal>("fd")),
+    _u_slip(getFunctor<ADReal>("u_slip")),
+    _v_slip(isParamValid("v_slip") ? &getFunctor<ADReal>("v_slip") : nullptr),
+    _w_slip(isParamValid("w_slip") ? &getFunctor<ADReal>("w_slip") : nullptr),
+    _density_interp_method(Moose::FV::selectInterpolationMethod(getParam<MooseEnum>("density_interp_method")))
+{
+  if (_dim >= 2 && !_v_slip)
+    mooseError("In two or more dimensions, the v_slip velocity must be supplied using the 'v_slip' "
+               "parameter");
+  if (_dim >= 3 && !_w_slip)
+    mooseError(
+        "In three dimensions, the w_slip velocity must be supplied using the 'w_slip' parameter");
+}
+
+ADReal
+INSFVMomentumDriftFlux::computeStrongResidual(const bool populate_a_coeffs)
+{
+  _normal = _face_info->normal();
+  const auto state = determineState();
+  Moose::FaceArg face_arg;
+  if (onBoundary(*_face_info))
+    face_arg = singleSidedFaceArg();
+  else
+    face_arg = Moose::FaceArg{_face_info, Moose::FV::LimiterType::CentralDifference, true, false, nullptr};
+
+  ADRealVectorValue u_slip_vel_vec;
+  if (_dim == 1)
[…]
