### 2023-12-02 b12c04b
Move hexagonal lattice utility from Cardinal with a slight rename to avoid conflict
