+++ modules/reactor/include/utils/HexagonalLatticeUtils.h
@@ -1,0 +1,813 @@
+/********************************************************************/
+/*                  SOFTWARE COPYRIGHT NOTIFICATION                 */
+/*                             Cardinal                             */
+/*                                                                  */
+/*                  (c) 2021 UChicago Argonne, LLC                  */
+/*                        ALL RIGHTS RESERVED                       */
+/*                                                                  */
+/*                 Prepared by UChicago Argonne, LLC                */
+/*               Under Contract No. DE-AC02-06CH11357               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*             Prepared by Battelle Energy Alliance, LLC            */
+/*               Under Contract No. DE-AC07-05ID14517               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*                 See LICENSE for full restrictions                */
+/********************************************************************/
+
+#pragma once
+
+#include "Moose.h"
+#include "MooseTypes.h"
+
+#ifndef CARDINAL_ENUM_H
+namespace channel_type
+{
+/// Type of subchannel
+enum ChannelTypeEnum
+{
+  interior,
+  edge,
+  corner
+};
+} // namespace channel_type
+#endif
+
+/**
+ * Class providing various utility functions related to triangular lattices of pins
+ * enclosed in a hexagonal duct.
+ */
+class HexagonalLatticeUtils
+{
+public:
+  HexagonalLatticeUtils(const Real & bundle_inner_flat_to_flat,
+                        const Real & pin_pitch,
+                        const Real & pin_diameter,
+                        const Real & wire_diameter,
+                        const Real & wire_pitch,
+                        const unsigned int & n_rings,
+                        const unsigned int & axis);
+
+  /**
+   * Distance from a point and a gap
+   * @param[in] pt point
+   * @param[in] gap_index gap index
+   * @ return distance from gap
+   */
+  Real distanceFromGap(const Point & pt, const unsigned int & gap_index) const;
+
+  /**
+   * Get the index for the gap closest to the point
+   * @param[in] point point
+   * @return index of closest gap
+   */
+  unsigned int gapIndex(const Point & point) const;
+
+  /**
+   * Get the gap index and distance to that gap for a given point
+   * @param[in] point point
+   * @param[out] index index of closest gap
+   * @param[out] distance distance to closest gap
+   */
+  void gapIndexAndDistance(const Point & point, unsigned int & index, Real & distance) const;
+
+  /**
+   * Get the unit vector translation to move a center point to a duct wall
+   * @param[in] side duct side
+   * @return x-component translation
+   */
+  Real sideTranslationX(const unsigned int & side) const { return _translation_x[side]; }
+
+  /**
+   * Get the unit vector translation to move a center point to a duct wall
+   * @param[in] side duct side
+   * @return y-component translation
+   */
+  Real sideTranslationY(const unsigned int & side) const { return _translation_y[side]; }
+
+  /**
+   * Get the index of the "first" pin in a ring
+   * @param[in] ring ring index
+   * @return first pin
+   */
+  unsigned int firstPinInRing(const unsigned int & ring) const;
+
+  /**
+   * Get the index of the "last" pin in a ring
+   * @param[in] ring ring index
+   * @return last pin
+   */
[…]
