+++ modules/reactor/src/utils/HexagonalLatticeUtils.C
@@ -1,0 +1,897 @@
+/********************************************************************/
+/*                  SOFTWARE COPYRIGHT NOTIFICATION                 */
+/*                             Cardinal                             */
+/*                                                                  */
+/*                  (c) 2021 UChicago Argonne, LLC                  */
+/*                        ALL RIGHTS RESERVED                       */
+/*                                                                  */
+/*                 Prepared by UChicago Argonne, LLC                */
+/*               Under Contract No. DE-AC02-06CH11357               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*             Prepared by Battelle Energy Alliance, LLC            */
+/*               Under Contract No. DE-AC07-05ID14517               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*                 See LICENSE for full restrictions                */
+/********************************************************************/
+
+#include "HexagonalLatticeUtils.h"
+#include "MooseUtils.h"
+#include "GeometryUtils.h"
+
+const Real HexagonalLatticeUtils::COS60 = 0.5;
+const Real HexagonalLatticeUtils::SIN60 = std::sqrt(3.0) / 2.0;
+const unsigned int HexagonalLatticeUtils::NUM_SIDES = 6;
+
+HexagonalLatticeUtils::HexagonalLatticeUtils(const Real & bundle_inner_flat_to_flat,
+                                             const Real & pin_pitch,
+                                             const Real & pin_diameter,
+                                             const Real & wire_diameter,
+                                             const Real & wire_pitch,
+                                             const unsigned int & n_rings,
+                                             const unsigned int & axis)
+  : _bundle_pitch(bundle_inner_flat_to_flat),
+    _pin_pitch(pin_pitch),
+    _pin_diameter(pin_diameter),
+    _wire_diameter(wire_diameter),
+    _wire_pitch(wire_pitch),
+    _n_rings(n_rings),
+    _axis(axis),
+    _bundle_side_length(hexagonSide(_bundle_pitch)),
+    _pin_area(M_PI * _pin_diameter * _pin_diameter / 4.0),
+    _pin_circumference(M_PI * _pin_diameter),
+    _wire_area(M_PI * _wire_diameter * _wire_diameter / 4.0),
+    _wire_circumference(M_PI * _wire_diameter),
+    _pin_surface_area_per_pitch(M_PI * _pin_diameter * _wire_pitch),
+    _pin_volume_per_pitch(_pin_area * _wire_pitch),
+    _P_over_D(_pin_pitch / _pin_diameter),
+    _L_over_D(_wire_pitch / _pin_diameter)
+{
+  auto idx = geom_utils::projectedIndices(_axis);
+  _ix = idx.first;
+  _iy = idx.second;
+
+  // object is not tested and probably won't work if axis != 2
+  if (_axis != 2)
+    mooseError("The HexagonalLatticeUtils is currently limited to 'axis = 2'!");
+
[…]
