### 2023-12-02 d254287
Add geometry utilities to support Hex utilites
