+++ framework/include/utils/GeometryUtils.h
@@ -1,0 +1,165 @@
+/********************************************************************/
+/*                  SOFTWARE COPYRIGHT NOTIFICATION                 */
+/*                             Cardinal                             */
+/*                                                                  */
+/*                  (c) 2021 UChicago Argonne, LLC                  */
+/*                        ALL RIGHTS RESERVED                       */
+/*                                                                  */
+/*                 Prepared by UChicago Argonne, LLC                */
+/*               Under Contract No. DE-AC02-06CH11357               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*             Prepared by Battelle Energy Alliance, LLC            */
+/*               Under Contract No. DE-AC07-05ID14517               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*                 See LICENSE for full restrictions                */
+/********************************************************************/
+
+#pragma once
+
+#include "Moose.h"
+#include "libmesh/point.h"
+#include "libmesh/mesh_tools.h"
+
+namespace geom_utils
+{
+/**
+ * Check whether a point is equal to zero
+ * @param[in] pt point
+ * @return whether point is equal to the zero point
+ */
+bool isPointZero(const Point & pt);
+
+/**
+ * Get the unit vector for a point parameter
+ * @param[in] pt point
+ * @param[in] name name of the parameter
+ */
+Point unitVector(const Point & pt, const std::string & name);
+
+/**
+ * Rotate point about an axis
+ * @param[in] p point
+ * @param[in] angle angle to rotate (radians)
+ * @param[in] axis axis expressed as vector
+ * @return rotated point
+ */
+Point rotatePointAboutAxis(const Point & p, const Real & angle, const Point & axis);
+
+/**
+ * Get the minimum distance from a point to another set of points, in the plane
+ * perpendicular to the specified axis
+ * @param[in] pt point
+ * @param[in] candidates set of points we will find the nearest of
+ * @param[in] axis axis perpendicular to the plane of the polygon
+ * @return minimum distance to the provided points
+ */
+Real minDistanceToPoints(const Point & pt,
+                         const std::vector<Point> & candidates,
+                         const unsigned int & axis);
+
+/**
+ * Get the corner coordinates of a regular 2-D polygon, assuming a face of the polygon
+ * is parallel to the x0 axis
+ * @param[in] num_sides number of sides to polygon
+ * @param[in] radius distance from polygon center to a corner
+ * @param[in] axis axis perpendicular to the plane of the polygon
+ * @return corner coordinates
+ */
+std::vector<Point>
+polygonCorners(const unsigned int & num_sides, const Real & radius, const unsigned int & axis);
+
+/**
+ * Get the indices of the plane perpendicular to the specified axis.
+ * For example, if the axis is the y-axis (1), then this will return
+ * (0, 2), indicating that the coordinates of a general 3-D point once
+ * projected onto the x-z plane can be obtained with the 0 and 2 indices.
+ * @param[in] axis axis perpendicular to the projection plane
+ * @return indices of coordinates on plane
+ */
+std::pair<unsigned int, unsigned int> projectedIndices(const unsigned int & axis);
+
+/**
+ * Given two coordinates, construct a point in the 2-D plane perpendicular to the
+ * specified axis.
+ * @param[in] x0 first coordinate
+ * @param[in] x1 second coordinate
+ * @param[in] axis axis perpendicular to the projection plane
+ * @return point
+ */
+Point projectPoint(const Real & x0, const Real & x1, const unsigned int & axis);
+
+/**
[…]
