+++ framework/src/utils/GeometryUtils.C
@@ -1,0 +1,294 @@
+/********************************************************************/
+/*                  SOFTWARE COPYRIGHT NOTIFICATION                 */
+/*                             Cardinal                             */
+/*                                                                  */
+/*                  (c) 2021 UChicago Argonne, LLC                  */
+/*                        ALL RIGHTS RESERVED                       */
+/*                                                                  */
+/*                 Prepared by UChicago Argonne, LLC                */
+/*               Under Contract No. DE-AC02-06CH11357               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*             Prepared by Battelle Energy Alliance, LLC            */
+/*               Under Contract No. DE-AC07-05ID14517               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*                 See LICENSE for full restrictions                */
+/********************************************************************/
+// Cardinal copyright as this was moved over from Cardinal
+
+#include "GeometryUtils.h"
+#include "MooseUtils.h"
+
+namespace geom_utils
+{
+
+bool
+isPointZero(const Point & pt)
+{
+  Point zero(0.0, 0.0, 0.0);
+  return pt.absolute_fuzzy_equals(zero);
+}
+
+Point
+unitVector(const Point & pt, const std::string & name)
+{
+  if (isPointZero(pt))
+    mooseError("'" + name + "' cannot have zero norm!");
+
+  return pt.unit();
+}
+
+Real
+minDistanceToPoints(const Point & pt,
+                    const std::vector<Point> & candidates,
+                    const unsigned int & axis)
+{
+  const auto idx = projectedIndices(axis);
+
+  Real distance = std::numeric_limits<Real>::max();
+  for (const auto & c : candidates)
+  {
+    Real dx = c(idx.first) - pt(idx.first);
+    Real dy = c(idx.second) - pt(idx.second);
+    Real d = std::sqrt(dx * dx + dy * dy);
+    distance = std::min(d, distance);
+  }
+
+  return distance;
+}
+
+Point
+projectPoint(const Real & x0, const Real & x1, const unsigned int & axis)
+{
+  auto i = projectedIndices(axis);
+  Point point;
+  point(i.first) = x0;
+  point(i.second) = x1;
+  point(axis) = 0.0;
+
+  return point;
+}
+
+Real
+projectedLineHalfSpace(Point pt1, Point pt2, Point pt3, const unsigned int & axis)
+{
+  // project points onto plane perpendicular to axis
+  pt1(axis) = 0.0;
+  pt2(axis) = 0.0;
+  pt3(axis) = 0.0;
+
+  auto i = projectedIndices(axis);
+
+  return (pt1(i.first) - pt3(i.first)) * (pt2(i.second) - pt3(i.second)) -
+         (pt2(i.first) - pt3(i.first)) * (pt1(i.second) - pt3(i.second));
+}
+
+bool
+pointInPolygon(const Point & point, const std::vector<Point> & corners, const unsigned int & axis)
+{
+  auto n_pts = corners.size();
+
+  std::vector<bool> negative_half_space;
+  std::vector<bool> positive_half_space;
+  for (unsigned int i = 0; i < corners.size(); ++i)
+  {
+    int next = (i == n_pts - 1) ? 0 : i + 1;
[…]
