+++ framework/doc/content/source/auxkernels/ProjectedMaterialPropertyNodalPatchRecoveryAux.md
@@ -1,0 +1,15 @@
+# ProjectedMaterialPropertyNodalPatchRecoveryAux
+
+This AuxKernel performs nodal patch recovery of a material property component with data prepared by a [ProjectedStatefulMaterialNodalPatchRecovery](ProjectedStatefulMaterialNodalPatchRecovery.md) object with the purpose of projecting the property onto a nodal basis function (e.g. first order Lagrange).
+
+This object is set up by the [ProjectedStatefulMaterialStorageAction](ProjectedStatefulMaterialStorageAction.md).
+
+See also [ProjectedStatefulMaterialAux](ProjectedStatefulMaterialAux.md) for the AuxKernel that projects onto elemental basis functions.
+
+!syntax description /AuxKernels/ProjectedMaterialPropertyNodalPatchRecoveryAux
+
+!syntax parameters /AuxKernels/ProjectedMaterialPropertyNodalPatchRecoveryAux
+
+!syntax inputs /AuxKernels/ProjectedMaterialPropertyNodalPatchRecoveryAux
+
+!syntax children /AuxKernels/ProjectedMaterialPropertyNodalPatchRecoveryAux
