+++ framework/doc/content/source/materials/InterpolatedStatefulMaterial.md
@@ -1,17 +1,23 @@
-# InterpolatedStatefulMaterial
+# InterpolatedStatefulMaterial...
 
-!syntax description /Materials/InterpolatedStatefulMaterial
+!syntax description /Materials/InterpolatedStatefulMaterialReal
 
 ## Description
 
 `InterpolatedStatefulMaterial` reconstitutes the old state of a material property from the old state of the AuxVariables holding the projected components of the material property.
 
+Variants include:
+- `InterpolatedStatefulMaterialReal`
+- `InterpolatedStatefulMaterialRealVectorValue`
+- `InterpolatedStatefulMaterialRankTwoTensor`
+- `InterpolatedStatefulMaterialRankFourTensor`
+
 This object is set up by the [ProjectedStatefulMaterialStorageAction](ProjectedStatefulMaterialStorageAction.md).
 
-!syntax parameters /Materials/InterpolatedStatefulMaterial
+!syntax parameters /Materials/InterpolatedStatefulMaterialReal
 
-!syntax inputs /Materials/InterpolatedStatefulMaterial
+!syntax inputs /Materials/InterpolatedStatefulMaterialReal
 
-!syntax children /Materials/InterpolatedStatefulMaterial
+!syntax children /Materials/InterpolatedStatefulMaterialReal
 
 !bibtex bibliography
