### 2023-12-02 d13b921
Verify implementation for two ducts
