### 2023-12-02 4366f3d
correcting clang Refs #22637
