### 2023-12-02 0977687
Apply suggestions from code review from grmnptr Refs #9007
Co-authored-by: Peter German <31662443+grmnptr@users.noreply.github.com>
