### 2023-12-02 5d5e92f
removing extra examples Refs #9007
