### 2023-12-02 8b60256
Allow for averaging user-defined variables.
Not just displacements.
