+++ modules/tensor_mechanics/src/vectorpostprocessors/AverageSectionValueSampler.C
@@ -7,23 +7,23 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "SectionDisplacementAverage.h"
+#include "AverageSectionValueSampler.h"
 #include "MooseMesh.h"
 #include "SystemBase.h"
 #include <limits>
 
-registerMooseObject("TensorMechanicsApp", SectionDisplacementAverage);
+registerMooseObject("TensorMechanicsApp", AverageSectionValueSampler);
 
 InputParameters
-SectionDisplacementAverage::validParams()
+AverageSectionValueSampler::validParams()
 {
   InputParameters params = GeneralVectorPostprocessor::validParams();
-  params.addClassDescription("Compute the section displacement vector average in three-dimensions "
+  params.addClassDescription("Compute the section's variable average in three-dimensions "
                              "given a user-defined definition of the cross section.");
   params.addParam<std::vector<SubdomainName>>(
       "block",
-      "The list of blocks in which to search for cross sectional nodes to compute the displacement "
-      "vector average.");
+      "The list of blocks in which to search for cross sectional nodes to compute the variable "
+      "average.");
   params.addRequiredParam<Point>("axis_direction", "Direction of the structural component's axis");
   params.addRequiredParam<Point>("reference_point",
                                  "Structural component reference starting point from which the "
@@ -32,6 +32,9 @@ SectionDisplacementAverage::validParams()
                         std::numeric_limits<double>::max(),
                         "Optional parameter to disambiguate cross sections of different structural "
                         "components when they share the same mesh block.");
+
+  params.addRequiredParam<std::vector<VariableName>>(
+      "variables", "Variables for the cross section output. These variables must be nodal.");
   params.addRequiredParam<std::vector<Real>>(
       "lengths", "Distance(s) to cross section from the global origin.");
   params.addParam<Real>("tolerance",
@@ -40,13 +43,11 @@ SectionDisplacementAverage::validParams()
   return params;
 }
 
-SectionDisplacementAverage::SectionDisplacementAverage(const InputParameters & parameters)
+AverageSectionValueSampler::AverageSectionValueSampler(const InputParameters & parameters)
   : GeneralVectorPostprocessor(parameters),
     _displaced_mesh(_app.actionWarehouse().displacedMesh()),
     _mesh(_app.actionWarehouse().mesh()),
-    _section_displacements_x(declareVector("section_displacements_x")),
-    _section_displacements_y(declareVector("section_displacements_y")),
-    _section_displacements_z(declareVector("section_displacements_z")),
+    _variables(getParam<std::vector<VariableName>>("variables")),
     _direction(getParam<Point>("axis_direction")),
     _reference_point(getParam<Point>("reference_point")),
     _lengths(getParam<std::vector<Real>>("lengths")),
@@ -54,57 +55,68 @@ SectionDisplacementAverage::SectionDisplacementAverage(const InputParameters & p
     _number_of_nodes(_lengths.size()),
     _cross_section_maximum_radius(getParam<Real>("cross_section_maximum_radius"))
 {
+  if (_mesh->dimension() != 3)
+    mooseError("The AverageSectionValueSampler postprocessor can only be used with three "
+               "dimensional meshes.");
+
   if (!MooseUtils::absoluteFuzzyEqual(_direction.norm_sq(), 1.0))
     paramError("axis_direction",
                "Axis diretion must have a norm of one and define the direction along which to "
                "locate cross sectional nodes.");
+
+  _output_vector.resize(_variables.size());
[…]
