+++ modules/tensor_mechanics/doc/content/source/vectorpostprocessors/AverageSectionValueSampler.md
@@ -1,11 +1,13 @@
-# SectionDisplacementAverage
+# AverageSectionValueSampler
 
-!syntax description /VectorPostprocessors/SectionDisplacementAverage
+!syntax description /VectorPostprocessors/AverageSectionValueSampler
 
 ## Description
 
 This vector postprocessor locates a number of nodes in a structural component and computes their average
-displacement along the $X$, $Y$, and $Z$ axes on multiple cross sections.
+variables along the $X$, $Y$, and $Z$ axes on multiple cross sections. This feature can be used to compute,
+for example, average displacements.
+
 The location of the nodes is given by a direction and an
 associated length. This postprocessor is particularly indicated for structural components that deform in
 a way analogous to beams, thereby freeing the user from performing additional postprocessing tasks on a
@@ -28,9 +30,16 @@ See below an input file excerpt which locates the cross section along the $Z$ di
 
 !listing modules/tensor_mechanics/test/tests/cross_section_deflection/test_one_step.i block=VectorPostprocessors
 
+Additionally, if multiple structural components share the same mesh block, one can set the input parameters
+`reference_point` and `cross_section_maximum_radius` to point to the structural component's local frame
+and the maximum in-cross-section-plane distance within. These parameters disambiguate which structural
+component the nodes belong. An example of computing average section variables for multiple strucutral
+components sharing the same mesh block is given below:
 
-!syntax parameters /VectorPostprocessors/SectionDisplacementAverage
+!listing modules/tensor_mechanics/test/tests/cross_section_deflection/test_one_step_two_ducts.i block=VectorPostprocessors
 
-!syntax inputs /VectorPostprocessors/SectionDisplacementAverage
+!syntax parameters /VectorPostprocessors/AverageSectionValueSampler
 
-!syntax children /VectorPostprocessors/SectionDisplacementAverage
+!syntax inputs /VectorPostprocessors/AverageSectionValueSampler
+
+!syntax children /VectorPostprocessors/AverageSectionValueSampler
