### 2023-12-03 c804ca8
More clean up to pass tests: - use legacy app-grouping mode for nearest-positions tests - be more careful with the index of the position before looking at whether it is the closest - add more context to error message in Positions - ban nearest positions + mesh_divisions in nearest-locations as it s not tested - revert trying to pre-allocate the buffers. There s something I dont understand, it causes diffs
