+++ framework/src/transfers/MultiAppGeneralFieldNearestLocationTransfer.C
@@ -16,6 +16,7 @@
 #include "MooseVariableFE.h"
 #include "SystemBase.h"
 #include "Positions.h"
+#include "MooseAppCoordTransform.h"
 
 #include "libmesh/system.h"
 
@@ -74,6 +75,9 @@ MultiAppGeneralFieldNearestLocationTransfer::MultiAppGeneralFieldNearestLocation
                "We do not support using both nearest positions matching and checking if target "
                "points are within an app domain because the KDTrees for nearest-positions matching "
                "are (currently) built with data from multiple applications.");
+  if (_nearest_positions_obj &&
+      (isParamValid("from_mesh_divisions") || isParamValid("to_mesh_divisions")))
+    paramError("use_nearest_position", "Cannot use nearest positions with mesh divisions");
 
   // Parameter checks on grouping subapp values
   if (_group_subapps && _from_mesh_divisions.empty() && !_nearest_positions_obj)
@@ -210,13 +214,15 @@ MultiAppGeneralFieldNearestLocationTransfer::buildKDTrees(const unsigned int var
 
   // Construct a local KDTree for each source. A source can be a single app or multiple apps
   // combined (option for nearest-position / mesh-divisions)
-  for (unsigned int i_source = 0; i_source < num_sources; ++i_source)
+  for (const auto i_source : make_range(num_sources))
   {
-    // Nest a loop on apps in case multiple apps contribute to the same source
+    // Nest a loop on apps in case multiple apps contribute to the same KD-Tree source
     for (const auto app_i : make_range(num_apps_per_tree))
     {
       // Get the current app index
-      const unsigned int i_from = _group_subapps ? app_i : (i_source / getNumDivisions());
+      const auto i_from = _group_subapps ? app_i : (i_source / getNumDivisions());
+      // Current position index, if using nearest positions
+      const auto i_pos = _group_subapps ? i_source : (i_source % getNumDivisions());
 
       // Get access to the variable and some variable information
       FEProblemBase & from_problem = *_from_problems[i_from];
@@ -270,11 +276,11 @@ MultiAppGeneralFieldNearestLocationTransfer::buildKDTrees(const unsigned int var
           // Only add to the KDTree nodes that are closest to the 'position'
           // When querying values at a target point, the KDTree associated to the closest
           // position to the target point is queried
-          if (_nearest_positions_obj && !closestToPosition(i_source, transformed_node))
+          if (_nearest_positions_obj && !closestToPosition(i_pos, transformed_node))
             continue;
 
           _local_points[i_source].push_back(transformed_node);
-          auto dof = node->dof_number(from_sys.number(), from_var_num, 0);
+          const auto dof = node->dof_number(from_sys.number(), from_var_num, 0);
           _local_values[i_source].push_back((*from_sys.solution)(dof));
           if (!_use_zero_dof_for_value[var_index])
             flagInvalidSolution(
@@ -322,7 +328,7 @@ MultiAppGeneralFieldNearestLocationTransfer::buildKDTrees(const unsigned int var
           const auto transformed_vertex_average =
               (*_from_transforms[getGlobalSourceAppIndex(i_from)])(vertex_average);
 
-          if (_nearest_positions_obj && !closestToPosition(i_source, transformed_vertex_average))
+          if (_nearest_positions_obj && !closestToPosition(i_pos, transformed_vertex_average))
             continue;
 
           _local_points[i_source].push_back(transformed_vertex_average);
[…]
