### 2023-12-03 1fa0d3a
For loop formatting and more constness
