+++ framework/src/transfers/MultiAppGeneralFieldTransfer.C
@@ -222,7 +222,7 @@ MultiAppGeneralFieldTransfer::initialSetup()
 
   // Use IDs for block and boundary restriction
   // Loop over all source problems
-  for (unsigned int i_from = 0; i_from < _from_problems.size(); ++i_from)
+  for (const auto i_from : index_range(_from_problems))
   {
     const auto & from_moose_mesh = _from_problems[i_from]->mesh(_displaced_source_mesh);
     if (isParamValid("from_blocks"))
@@ -291,7 +291,7 @@ MultiAppGeneralFieldTransfer::initialSetup()
   }
 
   // Loop over all target problems
-  for (unsigned int i_to = 0; i_to < _to_problems.size(); ++i_to)
+  for (const auto i_to : index_range(_to_problems))
   {
     const auto & to_moose_mesh = _to_problems[i_to]->mesh(_displaced_target_mesh);
     if (isParamValid("to_blocks"))
@@ -367,7 +367,7 @@ MultiAppGeneralFieldTransfer::initialSetup()
   }
 
   // Check if components are set correctly if using an array variable
-  for (unsigned int i_from = 0; i_from < _from_problems.size(); ++i_from)
+  for (const auto i_from : index_range(_from_problems))
   {
     for (const auto var_index : make_range(_from_var_names.size()))
     {
@@ -383,7 +383,7 @@ MultiAppGeneralFieldTransfer::initialSetup()
                    "Component passed is larger than size of variable");
     }
   }
-  for (unsigned int i_to = 0; i_to < _to_problems.size(); ++i_to)
+  for (const auto i_to : index_range(_to_problems))
   {
     for (const auto var_index : make_range(_to_var_names.size()))
     {
@@ -404,9 +404,9 @@ MultiAppGeneralFieldTransfer::initialSetup()
   if (_to_problems.size())
   {
     _to_variables.resize(_to_var_names.size());
-    for (unsigned int i_to = 0; i_to < _to_var_names.size(); ++i_to)
-      _to_variables[i_to] = &_to_problems[0]->getVariable(
-          0, _to_var_names[i_to], Moose::VarKindType::VAR_ANY, Moose::VarFieldType::VAR_FIELD_ANY);
+    for (const auto i_var : index_range(_to_var_names))
+      _to_variables[i_var] = &_to_problems[0]->getVariable(
+          0, _to_var_names[i_var], Moose::VarKindType::VAR_ANY, Moose::VarFieldType::VAR_FIELD_ANY);
   }
 }
 
@@ -417,7 +417,7 @@ MultiAppGeneralFieldTransfer::getAppInfo()
 
   // Create the point locators to locate evaluation points in the origin mesh(es)
   _from_point_locators.resize(_from_problems.size());
-  for (unsigned int i_from = 0; i_from < _from_problems.size(); ++i_from)
+  for (const auto i_from : index_range(_from_problems))
   {
     const auto & from_moose_mesh = _from_problems[i_from]->mesh(_displaced_source_mesh);
     _from_point_locators[i_from] =
@@ -432,7 +432,7 @@ MultiAppGeneralFieldTransfer::execute()
   getAppInfo();
 
   // loop over the vector of variables and make the transfer one by one
-  for (unsigned int i = 0; i < _var_size; ++i)
+  for (const auto i : make_range(_var_size))
     transferVariable(i);
 
   postExecute();
@@ -627,7 +627,7 @@ MultiAppGeneralFieldTransfer::cacheOutgoingPointInfo(const Point point,
   locatePointReceivers(point, processors);
 
   // We need to send this location data to these processors so they can send back values
-  for (auto pid : processors)
+  for (const auto pid : processors)
   {
[…]
