### 2023-12-03 b49e0a4
Move unit tests for geometry uilities, refs #26225
