+++ unit/src/GeometryUtilityTest.C
@@ -1,0 +1,313 @@
+/********************************************************************/
+/*                  SOFTWARE COPYRIGHT NOTIFICATION                 */
+/*                             Cardinal                             */
+/*                                                                  */
+/*                  (c) 2021 UChicago Argonne, LLC                  */
+/*                        ALL RIGHTS RESERVED                       */
+/*                                                                  */
+/*                 Prepared by UChicago Argonne, LLC                */
+/*               Under Contract No. DE-AC02-06CH11357               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*             Prepared by Battelle Energy Alliance, LLC            */
+/*               Under Contract No. DE-AC07-05ID14517               */
+/*                With the U. S. Department of Energy               */
+/*                                                                  */
+/*                 See LICENSE for full restrictions                */
+/********************************************************************/
+
+#include "GeometryUtilityTest.h"
+
+TEST_F(GeometryUtilityTest, norm)
+{
+  Point n(0.0, 0.0, 0.0);
+  Point pt;
+  try
+  {
+    pt = geom_utils::unitVector(n, "n");
+    FAIL() << "missing expected error";
+  }
+  catch (const std::exception & e)
+  {
+    std::string msg(e.what());
+    ASSERT_NE(msg.find("'n' cannot have zero norm!"), std::string::npos)
+        << "failed with unexpected error: " << msg;
+  }
+
+  Point m(5.0, 0.0, 0.0);
+  pt = geom_utils::unitVector(m, "m");
+  EXPECT_DOUBLE_EQ(pt(0), 1.0);
+  EXPECT_DOUBLE_EQ(pt(1), 0.0);
+  EXPECT_DOUBLE_EQ(pt(2), 0.0);
+}
+
+TEST_F(GeometryUtilityTest, rotate_about_y)
+{
+  Point n(0.0, 0.0, -1.0);
+  Point a(0.0, 1.0, 0.0);
+
+  Point pt(0.0, 0.0, -1.0);
+  Point ptr = geom_utils::rotatePointAboutAxis(pt, M_PI / 4.0, a);
+  EXPECT_DOUBLE_EQ(ptr(0), -sqrt(2.0) / 2.0);
+  EXPECT_DOUBLE_EQ(ptr(1), 0.0);
+  EXPECT_DOUBLE_EQ(ptr(2), -sqrt(2.0) / 2.0);
+}
+
[…]
