+++ framework/doc/content/source/meshgenerators/FlipSidesetGenerator.md
@@ -1,0 +1,14 @@
+# FlipSidesetGenerator
+
+!syntax description /Mesh/FlipSidesetGenerator
+
+## Overview
+
+The `FlipSidesetGenerator` object flips the normal of a sideset by changing the element the sideset is attached to. 
+The sideset can't be flipped if the original element it is attached to doesn't have a neighbor through the sideset. 
+
+!syntax parameters /Mesh/FlipSidesetGenerator
+
+!syntax inputs /Mesh/FlipSidesetGenerator
+
+!syntax children /Mesh/FlipSidesetGenerator
