### 2023-12-04 cf54e93
Example of porous flow natural convection
Refs #26238
