+++ modules/porous_flow/doc/content/modules/porous_flow/tests/relperm/relperm.py
@@ -1,0 +1,276 @@
+#!/usr/bin/env python3
+#* This file is part of the MOOSE framework
+#* https://www.mooseframework.org
+#*
+#* All rights reserved, see COPYRIGHT for full restrictions
+#* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+#*
+#* Licensed under LGPL 2.1, please see LICENSE for details
+#* https://www.gnu.org/licenses/lgpl-2.1.html
+
+import matplotlib.pyplot as plt
+import numpy as np
+
+## creating the analytical function
+
+def effS(S_res,sum_S_res,S): #this function calculate the effective saturation
+
+  eff_S=[]
+
+  for i in range(len(S)):
+    num=S[i]-S_res #numerator
+    den=1-sum_S_res #denominator
+    a=num/den
+    if a>1:
+      a=1
+    elif a<0:
+      a=0
+
+    eff_S.append(a)
+  return eff_S
+
+
+def brook_corey_func(lamda,S_res_w=0,sum_S_res_w=0,S_res_nw=0,sum_S_res_nw=0): #This function return a list of 2 listes, the first one is for wet phase and second is for non wet
+  k=[]
+  Seff_w=effS(S_res_w,sum_S_res_w,np.linspace(0,1,100))
+
+
+  k_w=[]
+  for i in range(len(Seff_w)):
+    a=Seff_w[i]**((2+3*lamda)/lamda)
+
+    k_w.append(a)
+
+  k.append(k_w)
+
+  Seff_nw=effS(S_res_nw,sum_S_res_nw,np.linspace(0,1,100))
+  k_nw=[]
+  for i in range(len(Seff_nw)):
+    a=(1-Seff_nw[i])**2
+    b=(1-Seff_nw[i]**((2+lamda)/lamda))
+    k_nw.append(a*b)
+
+  k.append(k_nw)
+  return k
+
+
+def corey_func(n,S_res_w=0,sum_S_res_w=0,S_res_nw=0,sum_S_res_nw=0,scalew=1,scalenw=1): #This function return a list of 2 listes, the first one is for wet phase and second is for non wet
+  k=[]
+  Seff_w=effS(S_res_w,sum_S_res_w,np.linspace(0,1,100))
+
+
+  k_w=[]
+  for i in range(len(Seff_w)):
+    a=((Seff_w[i])**(n))*scalew
+
+    k_w.append(a)
+
+  k.append(k_w)
+
+  Seff_nw=effS(S_res_nw,sum_S_res_nw,np.linspace(1,0,100))
+  k_nw=[]
+  for i in range(len(Seff_nw)):
+    a=((Seff_nw[i])**(n))*scalenw
+
+    k_nw.append(a)
+
+  k.append(k_nw)
+  return k
+
+
+
+def vanGen_func(m,S_res_w=0,sum_S_res_w=0,S_res_nw=0,sum_S_res_nw=0,scalew=1,scalenw=1): #This function return a list of 2 listes, the first one is for wet phase and second is for non wet
+  k=[]
+  Seff_w=effS(S_res_w,sum_S_res_w,np.linspace(0,1,100))
+
+
+  k_w=[]
+  for i in range(len(Seff_w)):
+    a=(Seff_w[i])**0.5
+    b=(1-(1-Seff_w[i]**(1/m))**m)**2
+
+    k_w.append(a*b)
+
+  k.append(k_w)
+
+  Seff_nw=effS(S_res_nw,sum_S_res_nw,np.linspace(1,0,100))
+  k_nw=[]
[…]
