+++ modules/porous_flow/doc/content/modules/porous_flow/tests/capillary_pressure/capillary_pressure_tests.md
@@ -1,3 +1,35 @@
 # Capillary pressure test descriptions
 
-TODO
+#### The detail documentation of the implementation of this code packet can be founded in [here](porous_flow/capillary_pressure.md).
+
+## Brooks-Corey
+
+!media capillary_pressure/brook_corey.png style=width:100%;margin-left:10px; caption=Brooks-Corey Capillary Pressure test cases id=brookcoreyPc
+
+The input file for log-extension off test:
+
+!listing modules/porous_flow/test/tests/capillary_pressure/brooks_corey1.i
+
+The input file for log-extension on test:
+
+!listing modules/porous_flow/test/tests/capillary_pressure/brooks_corey2.i
+
+
+
+## van Genuchten
+
+
+
+!media capillary_pressure/vangen.png style=width:100%;margin-left:10px; caption=van Genuchten Capillary Pressure test cases id=vangenPc
+
+The input file for log-extension off test:
+
+!listing modules/porous_flow/test/tests/capillary_pressure/vangenuchten1.i
+
+The input file for log-extension on test:
+
+!listing modules/porous_flow/test/tests/capillary_pressure/vangenuchten2.i
+
+The input file for applied-scaling test:
+
+!listing modules/porous_flow/test/tests/capillary_pressure/vangenuchten3.i

[…]
