### 2023-12-04 0315baa
Fixup typo in gallery extension error message
