### 2023-12-04 a784920
Add Helper functions to bypass new parser changes for Apps
Modify helper functions
