+++ framework/src/base/MooseServer.C
@@ -54,6 +54,9 @@ MooseServer::parseDocumentForDiagnostics(wasp::DataArray & diagnosticsList)
   std::string parse_file_path = document_path;
   pcrecpp::RE("(.*://)(.*)").Replace("\\2", &parse_file_path);
 
+  // create new parser tree for the application
+  auto front_parser = std::make_unique<Parser>();
+
   // copy parent application parameters and modify to set up input check
   InputParameters app_params = _moose_app.parameters();
   app_params.set<std::vector<std::string>>("input_file") = {parse_file_path};
@@ -64,20 +67,16 @@ MooseServer::parseDocumentForDiagnostics(wasp::DataArray & diagnosticsList)
   app_params.set<bool>("error_deprecated") = true;
   app_params.set<std::string>("color") = "off";
   app_params.set<bool>("disable_perf_graph_live") = true;
+  app_params.set<std::shared_ptr<Parser>>("_parser") = std::move(front_parser);
+
   app_params.remove("language_server");
 
   // turn output off so input check application does not affect messages
   std::streambuf * cached_output_buffer = Moose::out.rdbuf(nullptr);
 
-  // create new parser tree for the application
-  auto front_parser = std::make_unique<Parser>();
-
   // create new application with parameters modified for input check run
-  _check_app = AppFactory::instance().createShared(_moose_app.type(),
-                                                   _moose_app.name(),
-                                                   app_params,
-                                                   std::move(front_parser),
-                                                   _moose_app.getCommunicator()->get());
+  _check_app = AppFactory::instance().createShared(
+      _moose_app.type(), _moose_app.name(), app_params, _moose_app.getCommunicator()->get());
 
   // disable logs and enable error exceptions with initial values cached
   bool cached_logging_enabled = Moose::perf_log.logging_enabled();
