+++ modules/phase_field/unit/include/EBSDMeshErrorTest.h
@@ -25,7 +25,7 @@ protected:
   void SetUp()
   {
     const char * argv[2] = {"foo", "\0"};
-    _app = AppFactory::createAppShared("PhaseFieldApp", 1, (char **)argv);
+    _app = moose::createMooseApp("PhaseFieldApp", 1, (char **)argv);
     _factory = &_app->getFactory();
   }
 
