### 2023-12-04 73f43a1
Remove details page, add 2024 Oct training
Remove the need to click on a separate link to list details for a
giving training. Instead, when clicked, you are brought to the
details page, and there is a link to register from that.

Added the upcoming 2024 Oct training I saw on the drop down list
at the registration site.

Closes #26249
