+++ modules/doc/content/training/more_detail/MOOSE_2024_10_15-17_UIUC.md
@@ -1,0 +1,9 @@
+#### MOOSE Framework (2024 Oct 15-17, University of Illinois Urbana-Champaign)
+
+!style! halign=left
+!include training/more_detail/registration_link.md
+!style-end!
+
+!include training/more_detail/generic_moose_training.md
+
+[Back](training/index.md)
