+++ modules/doc/content/training/more_detail/registration_link.md
@@ -1,0 +1,1 @@
+[Register for this training](https://ncrcaims.inl.gov/Identity/Account/TrainingRegistration).
