### 2023-12-04 4692153
Address Alex's review - constness - clearer code split - test wording Co-authored-by: Alex Lindsay <alexander.lindsay@inl.gov>
