### 2023-12-04 7960a46
Protect EigenProblem constructor against missing SLEPc
Refs #26113
