+++ framework/src/userobjects/ElementSubdomainModifierBase.C
@@ -1,0 +1,530 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "ElementSubdomainModifierBase.h"
+#include "DisplacedProblem.h"
+
+#include "libmesh/parallel_algebra.h"
+#include "libmesh/parallel.h"
+#include "libmesh/dof_map.h"
+#include "libmesh/remote_elem.h"
+#include "libmesh/parallel_ghost_sync.h"
+#include "libmesh/petsc_vector.h"
+
+InputParameters
+ElementSubdomainModifierBase::validParams()
+{
+  InputParameters params = ElementUserObject::validParams();
+
+  // ESMs only operate on the undisplaced mesh to gather subdomain and sideset information. This
+  // information is used to modify both the undisplaced and the displaced meshes. It is the
+  // developer's responsibility to make sure the element IDs and sideset info are consistent across
+  // both meshes.
+  params.set<bool>("use_displaced_mesh") = false;
+  params.suppressParameter<bool>("use_displaced_mesh");
+
+  params.addParam<std::vector<BoundaryName>>(
+      "moving_boundaries",
+      {},
+      "Moving boundaries between subdomains. These boundaries (both sidesets and nodesets) will be "
+      "updated as elements change their subdomain. The corresponding subdomains of each moving "
+      "boundary shall be specified using the parameter 'moving_boundary_subdomain_pairs'. A "
+      "boundary will be created on the mesh if it does not already exist.");
+  params.addParam<std::vector<std::vector<SubdomainName>>>(
+      "moving_boundary_subdomain_pairs",
+      {},
+      "The subdomain pairs associated with each moving boundary. For each pair of subdomains, only "
+      "the element side from the first subdomain will be added to the moving boundary, i.e., the "
+      "side normal is pointing from the first subdomain to the second subdomain. The pairs shall "
+      "be delimited by ';'. If a pair only has one subdomain, the moving boundary is associated "
+      "with the subdomain's external boundary, i.e., when the elements have no neighboring "
+      "elements.");
+
+  params.addParam<std::vector<SubdomainName>>(
+      "active_subdomains",
+      {},
+      "The variables specified by the parameter 'initialize_variables' are initialized to their "
+      "initial conditions when elements are moved from 'inactive' subdomains to 'active' "
+      "subdomains, or from an 'active' subdomain to a different 'active' subdomain. The 'active' "
+      "subdomains are usually where the simulation is performed, i.e. the PDE's computational "
+      "domain. If this parameter is left empty, then the entire mesh is treated as the active "
+      "subdomain.");
+  params.addParam<bool>(
+      "amorphous_activation",
+      false,
+      "If set to false (default), only elements moving from inactive subdomains to active "
+      "subdomains are considered as 'activated'. When set to true, elements moving from an active "
+      "subdomain to another are _also_ considered to be 'activated'.");
+
+  return params;
+}
+
[…]
