+++ framework/src/meshmodifiers/ElementSubdomainModifier.C
@@ -8,138 +8,39 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ElementSubdomainModifier.h"
-#include "DisplacedProblem.h"
-
-#include "libmesh/parallel_algebra.h"
-#include "libmesh/parallel.h"
-#include "libmesh/dof_map.h"
-#include "libmesh/remote_elem.h"
-#include "libmesh/parallel_ghost_sync.h"
 
 InputParameters
 ElementSubdomainModifier::validParams()
 {
-  InputParameters params = ElementUserObject::validParams();
-  params.addClassDescription(
-      "Modify element subdomain ID. This userobject only runs on the undisplaced mesh, and it will "
-      "modify both the undisplaced and the displaced mesh.");
-  params.addParam<bool>("apply_initial_conditions",
-                        true,
-                        "Whether to apply initial conditions on the moved nodes and elements");
-  params.addParam<BoundaryName>(
-      "moving_boundary_name",
-      "Boundary to modify when an element is moved. A boundary with the provided name will be "
-      "created if not already exists on the mesh.");
-  params.addParam<BoundaryName>("complement_moving_boundary_name",
-                                "Boundary that associated with the unmoved domain when neighbor "
-                                "element(s) is moved. A boundary with the provided name will be "
-                                "created if not already exists on the mesh.");
-  params.set<bool>("use_displaced_mesh") = false;
-  params.suppressParameter<bool>("use_displaced_mesh");
-  params.registerBase("MeshModifier");
+  InputParameters params = ElementSubdomainModifierBase::validParams();
   return params;
 }
 
 ElementSubdomainModifier::ElementSubdomainModifier(const InputParameters & parameters)
-  : ElementUserObject(parameters),
-    _displaced_problem(_fe_problem.getDisplacedProblem().get()),
-    _apply_ic(getParam<bool>("apply_initial_conditions")),
-    _moving_boundary_specified(isParamValid("moving_boundary_name")),
-    _complement_moving_boundary_specified(isParamValid("complement_moving_boundary_name")),
-    _moving_boundary_id(-1)
+  : ElementSubdomainModifierBase(parameters)
 {
-}
-
-void
-ElementSubdomainModifier::initialSetup()
-{
-  if (_moving_boundary_specified)
-  {
-    _moving_boundary_name = getParam<BoundaryName>("moving_boundary_name");
-    setMovingBoundaryName(_mesh);
-    if (_displaced_problem)
-      setMovingBoundaryName(_displaced_problem->mesh());
-  }
-
-  if (_complement_moving_boundary_specified)
-  {
-    _complement_moving_boundary_name = getParam<BoundaryName>("complement_moving_boundary_name");
-    setComplementMovingBoundaryName(_mesh);
-    if (_displaced_problem)
-      setComplementMovingBoundaryName(_displaced_problem->mesh());
-  }
-}
-
-void
-ElementSubdomainModifier::setMovingBoundaryName(MooseMesh & mesh)
-{
-  // We only need one boundary to modify. Create a dummy vector just to use the API.
-  const std::vector<BoundaryID> boundary_ids = mesh.getBoundaryIDs({{_moving_boundary_name}}, true);
-  mooseAssert(boundary_ids.size() == 1, "Expect exactly one boundary ID.");
-  _moving_boundary_id = boundary_ids[0];
-  mesh.setBoundaryName(_moving_boundary_id, _moving_boundary_name);
-  mesh.getMesh().get_boundary_info().sideset_name(_moving_boundary_id) = _moving_boundary_name;
[…]
