### 2023-12-04 740e0e3
Improve `ElementSubdomainModifier`s
There have been several bug reports and feature requests since the introduction of `ElementSubdomainModifier`.
This PR fixes most, if not all, of the bugs and introduces a new feature.

Notable changes:
- `meshChanged()` is now only called when at least one element changes its subdomain. The old behavior led to an adaptivity file (*-s0*.e) file at every time step.
- A new input parameter "active_subdomain" asks for _active_ subdomains on which the equation system exists. This enables the
  detection of newly activated dofs so that variable initialization and
[…]
