+++ framework/src/userobjects/ElementSubdomainModifier.C
@@ -15,6 +15,7 @@
 #include "libmesh/dof_map.h"
 #include "libmesh/remote_elem.h"
 #include "libmesh/parallel_ghost_sync.h"
+#include "libmesh/petsc_vector.h"
 
 InputParameters
 ElementSubdomainModifier::validParams()
@@ -23,9 +24,6 @@ ElementSubdomainModifier::validParams()
   params.addClassDescription(
       "Modify element subdomain ID. This userobject only runs on the undisplaced mesh, and it will "
       "modify both the undisplaced and the displaced mesh.");
-  params.addParam<bool>("apply_initial_conditions",
-                        true,
-                        "Whether to apply initial conditions on the moved nodes and elements");
   params.addParam<BoundaryName>(
       "moving_boundary_name",
       "Boundary to modify when an element is moved. A boundary with the provided name will be "
@@ -34,6 +32,33 @@ ElementSubdomainModifier::validParams()
                                 "Boundary that associated with the unmoved domain when neighbor "
                                 "element(s) is moved. A boundary with the provided name will be "
                                 "created if not already exists on the mesh.");
+
+  params.addParam<std::vector<VariableName>>(
+      "initialize_variables",
+      {},
+      "Which variables to initialize when an element is moved into the active subdomains.");
+  MooseEnum init_strategy("IC NEAREST CONSTANT", "IC");
+  params.addParam<std::vector<MooseEnum>>(
+      "initialization_strategy",
+      {init_strategy},
+      "Which strategy to use when initializing variables on newly activated elements and nodes. "
+      "IC applies the variable initial condition to the newly activated dofs. NEAREST applies the "
+      "old value from the nearest element/node to the newly activated dofs. CONSTANT initializes "
+      "the newly activated dofs with a constant. This parameter should either have size of 1 or "
+      "size equal to that of 'initialize_variables'. If the size of this parameter is 1, then the "
+      "same strategy will be used for all variables listed in 'initialize_variables'.");
+  params.addParam<std::vector<Real>>(
+      "initialization_constant",
+      "The constants to use when initialization_strategy = CONSTANT. This parameter should either "
+      "have size of 1 or size equal to the number of variables with "
+      "initialization_strategy==CONSTANT.");
+  params.addParam<std::vector<SubdomainName>>(
+      "active_subdomains",
+      {},
+      "The 'active' subdomains on which the simulation is performed, i.e. the PDE "
+      "computational domain. If this parameter is left empty, then the entire mesh is treated as "
+      "the activate subdomain.");
+
   params.set<bool>("use_displaced_mesh") = false;
   params.suppressParameter<bool>("use_displaced_mesh");
   return params;
@@ -42,11 +67,59 @@ ElementSubdomainModifier::validParams()
 ElementSubdomainModifier::ElementSubdomainModifier(const InputParameters & parameters)
   : ElementUserObject(parameters),
     _displaced_problem(_fe_problem.getDisplacedProblem().get()),
-    _apply_ic(getParam<bool>("apply_initial_conditions")),
+    _nl_sys(_fe_problem.getNonlinearSystemBase(systemNumber())),
+    _aux_sys(_fe_problem.getAuxiliarySystem()),
+    _init_vars(getParam<std::vector<VariableName>>("initialize_variables")),
+    _active_subdomains(
[…]
