+++ framework/include/userobjects/ElementSubdomainModifier.h
@@ -21,10 +21,12 @@ public:
   ElementSubdomainModifier(const InputParameters & parameters);
 
   virtual void initialSetup() override;
+  virtual void timestepSetup() override;
   virtual void initialize() override;
   virtual void execute() override;
   virtual void threadJoin(const UserObject & /*uo*/) override;
   virtual void finalize() override;
+  virtual void meshChanged() override;
 
 protected:
   /// Compute the subdomain ID of the current element
@@ -62,16 +64,30 @@ protected:
     return _complement_moving_boundary_name;
   }
 
-  /// Range of the elements who changed their subdomain ID
-  ConstElemRange & movedElemsRange() const { return *_moved_elems_range; }
+  /// Range of activated elements
+  ConstElemRange & activatedElemsRange() const { return *_activated_elems_range; }
 
-  /// Range of the boundary nodes on moved elements
-  ConstBndNodeRange & movedBndNodesRange() const { return *_moved_bnd_nodes_range; }
+  /// Range of activated nodes
+  ConstNodeRange & activatedNodesRange() const { return *_activated_nodes_range; }
+
+  /// Range of activated boundary nodes
+  ConstBndNodeRange & activatedBndNodesRange() const { return *_activated_bnd_nodes_range; }
 
   /// Pointer to the displaced problem
   DisplacedProblem * _displaced_problem;
 
+  /// Nonlinear system
+  NonlinearSystemBase & _nl_sys;
+
+  /// Auxiliary system
+  AuxiliarySystem & _aux_sys;
+
 private:
+  /// Serialize the old solution
+  void serializeSolutionOld(dof_id_type ndof,
+                            SystemBase & sys,
+                            std::unique_ptr<NumericVector<Real>> & sol);
+
   /// Set the name of the moving boundary. Create the nodeset/sideset if not exist.
   void setMovingBoundaryName(MooseMesh & mesh);
 
@@ -97,14 +113,24 @@ private:
   /// Change the subdomain ID of all ancestor elements
   void setAncestorsSubdomainIDs(const SubdomainID & subdomain_id, const dof_id_type & elem_id);
 
-  /// Helper function to build the range of moved elements
-  void buildMovedElemsRange();
+  /// Helper function to build the range of activated elements
+  void buildActivatedElemsRange();
 
-  /// Helper function to build the range of boundary nodes on moved elements
-  void buildMovedBndNodesRange();
+  /// Helper function to build the range of activated nodes
+  void buildActivatedNodesRange();
 
-  /// Set old and older solutions to be the same as the current solution
-  void setOldAndOlderSolutionsForMovedNodes(SystemBase & sys);
+  /// Helper function to build the range of activated boundary nodes
+  void buildActivatedBndNodesRange();
+
+  /// Find nearest dofs for each variable
+  void findNearestDofs(SystemBase &, const VariableName &);
+
+  /// Set current solution to be the same as the nearest dof's old value
+  void
+  setNearestSolutionForActivatedDofs(SystemBase &, const VariableName &, NumericVector<Real> &);
+
+  /// Set activated dofs to constant
+  void setConstantForActivatedDofs(SystemBase & sys, const VariableName & var_name);
 
   /// Elements on the undisplaced mesh whose subdomain IDs have changed
   std::vector<const Elem *> _moved_elems;
@@ -112,17 +138,35 @@ private:
   /// Elements on the displaced mesh whose subdomain IDs have changed
   std::vector<const Elem *> _moved_displaced_elems;
 
[…]
