### 2023-12-05 72dd549
Fix clang error
