### 2023-12-05 45f475a
Add LinearVariableName and start transitioning towards functors in BCs. (#25722)
