### 2023-12-05 792f266
Fixed some porous tests following suggestions Refs #26240
