+++ modules/porous_flow/doc/content/modules/porous_flow/tests/capillary_pressure/capillary_pressure_tests.md
@@ -1,10 +1,12 @@
 # Capillary pressure test descriptions
 
-#### The detail documentation of the implementation of this code packet can be founded in [here](porous_flow/capillary_pressure.md).
+The details of the capillary pressure curve implementations can be found in [here](porous_flow/capillary_pressure.md).
 
 ## Brooks-Corey
 
-!media capillary_pressure/brook_corey.png style=width:100%;margin-left:10px; caption=Brooks-Corey Capillary Pressure test cases id=brookcoreyPc
+!media capillary_pressure/brook_corey.png style=width:100%;margin-left:10px; caption=Brooks-Corey Capillary Pressure test cases. id=brookcoreyPc
+
+There are two test cases for Brooks-Corey including the log-extension turning on, and log-extension turning off at low saturation.
 
 The input file for log-extension off test:
 
@@ -20,7 +22,9 @@ The input file for log-extension on test:
 
 
 
-!media capillary_pressure/vangen.png style=width:100%;margin-left:10px; caption=van Genuchten Capillary Pressure test cases id=vangenPc
+!media capillary_pressure/vangen.png style=width:100%;margin-left:10px; caption=van Genuchten Capillary Pressure test cases. id=vangenPc
+
+There are three test cases for van Genuchten including the log-extension turning on, log-extension turning off at low saturation, and changing the scale factor.
 
 The input file for log-extension off test:
 

[…]
