+++ framework/src/utils/NestedSolve.C
@@ -25,7 +25,7 @@ NestedSolveTempl<is_ad>::validParams()
                         absoluteToleranceDefault(),
                         "Absolute convergence tolerance for Newton iteration");
   params.addParam<Real>(
-      "x_tolerance", xToleranceDefault(), "Threshold for minimum step size of linear iterations");
+      "step_size_tolerance", xToleranceDefault(), "Minimum step size of linear iterations relative to value of the solution");
   params.addParam<unsigned int>(
       "min_iterations",
       minIterationsDefault(),
@@ -44,7 +44,7 @@ template <bool is_ad>
 NestedSolveTempl<is_ad>::NestedSolveTempl()
   : _relative_tolerance_square(Utility::pow<2>(relativeToleranceDefault())),
     _absolute_tolerance_square(Utility::pow<2>(absoluteToleranceDefault())),
-    _x_tolerance_square(Utility::pow<2>(xToleranceDefault())),
+    _delta_thresh(xToleranceDefault()),
     _min_iterations(minIterationsDefault()),
     _max_iterations(maxIterationsDefault()),
     _acceptable_multiplier(acceptableMultiplierDefault()),
@@ -57,7 +57,7 @@ template <bool is_ad>
 NestedSolveTempl<is_ad>::NestedSolveTempl(const InputParameters & params)
   : _relative_tolerance_square(Utility::pow<2>(params.get<Real>("relative_tolerance"))),
     _absolute_tolerance_square(Utility::pow<2>(params.get<Real>("absolute_tolerance"))),
-    _x_tolerance_square(Utility::pow<2>(params.get<Real>("x_tolerance"))),
+    _delta_thresh(params.get<Real>("step_size_tolerance")),
     _min_iterations(params.get<unsigned int>("min_iterations")),
     _max_iterations(params.get<unsigned int>("max_iterations")),
     _acceptable_multiplier(params.get<Real>("acceptable_multiplier")),
