+++ framework/include/linearfvkernel/FluxLinearFVKernel.h
@@ -13,20 +13,24 @@
 #include "ElemInfo.h"
 #include "MooseVariableFV.h"
 #include "MooseVariableInterface.h"
+#include "FaceArgInterface.h"
 
-class FluxLinearFVKernel : public LinearFVKernel
+class FluxLinearFVKernel : public LinearFVKernel, public FaceArgProducerInterface
 {
 public:
   static InputParameters validParams();
   FluxLinearFVKernel(const InputParameters & params);
 
-  /// Compute this object's contribution to the residual
+  /// Compute this object's contribution to the system matrix
   virtual void addMatrixContribution() override;
 
-  /// Compute this object's contribution to the diagonal Jacobian entries
+  /// Compute this object's contribution to the right hand side
   virtual void addRightHandSideContribution() override;
 
-  /// Set current element
+  /// If the current element is defined on the elem side or neighbor side
+  bool hasFaceSide(const FaceInfo & fi, bool fi_elem_side) const override;
+
+  /// Set current face info
   void setCurrentFaceInfo(const FaceInfo * face_info)
   {
     _current_face_info = face_info;
@@ -34,21 +38,42 @@ public:
         _current_face_info->faceType(std::make_pair(_var->number(), _var->sys().number()));
   }
 
+  /// Computes the system matrix contribution from an element side on an internal face
   virtual Real computeElemMatrixContribution() = 0;
 
+  /// Computes the system matrix contribution from the neighbor side on an internal face
   virtual Real computeNeighborMatrixContribution() = 0;
 
+  /// Computes the right hand side contribution from the element side on an internal face
   virtual Real computeElemRightHandSideContribution() = 0;
 
+  /// Computes the right hand side contribution from the neighbor side on an internal face
   virtual Real computeNeighborRightHandSideContribution() = 0;
 
+  /// Computes the system matrix contribution from a boundary face
   virtual Real computeBoundaryMatrixContribution() = 0;
 
+  /// Computes the righth and side contribution from a boundary face
   virtual Real computeBoundaryRHSContribution() = 0;
 
 protected:
+  /**
+   * Determine the single sided face argument when evaluating a functor on a face.
+   * This is used to perform evaluations of material properties with the actual face values of
+   * their dependences, rather than interpolate the material property to the boundary.
+   * @param fi the FaceInfo for this face
+   * @param limiter_type the limiter type, to be specified if more than the default average
+   *        interpolation is required for the parameters of the functor
+   * @param correct_skewness whether to perform skew correction at the face
+   */
+  Moose::FaceArg singleSidedFaceArg(
+      const FaceInfo * fi = nullptr,
+      Moose::FV::LimiterType limiter_type = Moose::FV::LimiterType::CentralDifference,
+      bool correct_skewness = false) const;
+
+  /// Pointer to the face info we are operating on right now
   const FaceInfo * _current_face_info;
 
-  /// The face type
+  /// Face ownership information for the current face
   FaceInfo::VarFaceNeighbors _current_face_type;
 };
