### 2023-12-05 3de3937
Propagate face information to boundary conditions, polish documentation, consolidatesource kernel. (#25722)
