+++ framework/src/linearfvkernels/LinearFVFluxKernel.C
@@ -7,19 +7,19 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "FluxLinearFVKernel.h"
+#include "LinearFVFluxKernel.h"
 #include "Assembly.h"
 #include "SubProblem.h"
 
 InputParameters
-FluxLinearFVKernel::validParams()
+LinearFVFluxKernel::validParams()
 {
   InputParameters params = LinearFVKernel::validParams();
-  params.registerSystemAttributeName("FluxLinearFVKernel");
+  params.registerSystemAttributeName("LinearFVFluxKernel");
   return params;
 }
 
-FluxLinearFVKernel::FluxLinearFVKernel(const InputParameters & params)
+LinearFVFluxKernel::LinearFVFluxKernel(const InputParameters & params)
   : LinearFVKernel(params),
     FaceArgProducerInterface(),
     _current_face_info(nullptr),
@@ -28,7 +28,7 @@ FluxLinearFVKernel::FluxLinearFVKernel(const InputParameters & params)
 }
 
 void
-FluxLinearFVKernel::addMatrixContribution()
+LinearFVFluxKernel::addMatrixContribution()
 {
   if (_current_face_type == FaceInfo::VarFaceNeighbors::BOTH)
   {
@@ -45,29 +45,31 @@ FluxLinearFVKernel::addMatrixContribution()
     (*_linear_system.matrix).add(dof_id_neighbor, dof_id_elem, -elem_matrix_contribution);
     (*_linear_system.matrix).add(dof_id_neighbor, dof_id_neighbor, -neighbor_matrix_contribution);
   }
-  else if (_var->getBoundaryCondition(*_current_face_info->boundaryIDs().begin()))
+  else if (auto * bc_pointer =
+               _var->getBoundaryCondition(*_current_face_info->boundaryIDs().begin()))
   {
     mooseAssert(_current_face_info->boundaryIDs().size() == 1,
                 "We should only have one boundary on every face.");
+    bc_pointer->setCurrentFaceInfo(_current_face_info, _current_face_type);
     if (_current_face_type == FaceInfo::VarFaceNeighbors::ELEM)
     {
       const auto dof_id_elem =
           _current_face_info->elemInfo()->dofIndices()[_var->sys().number()][_var->number()];
-      const auto matrix_contribution = computeBoundaryMatrixContribution();
+      const auto matrix_contribution = computeBoundaryMatrixContribution(bc_pointer);
       (*_linear_system.matrix).add(dof_id_elem, dof_id_elem, matrix_contribution);
     }
     else if (_current_face_type == FaceInfo::VarFaceNeighbors::NEIGHBOR)
     {
       const auto dof_id_neighbor =
           _current_face_info->neighborInfo()->dofIndices()[_var->sys().number()][_var->number()];
-      const auto matrix_contribution = computeBoundaryMatrixContribution();
+      const auto matrix_contribution = computeBoundaryMatrixContribution(bc_pointer);
       (*_linear_system.matrix).add(dof_id_neighbor, dof_id_neighbor, matrix_contribution);
     }
   }
 }
 
 void
-FluxLinearFVKernel::addRightHandSideContribution()
+LinearFVFluxKernel::addRightHandSideContribution()
 {
   if (_current_face_type == FaceInfo::VarFaceNeighbors::BOTH)
   {
@@ -82,27 +84,29 @@ FluxLinearFVKernel::addRightHandSideContribution()
     (*_linear_system.rhs).add(dof_id_elem, elem_rhs_contribution);
     (*_linear_system.rhs).add(dof_id_neighbor, neighbor_rhs_contribution);
   }
[…]
