### 2023-12-05 95e1356
Add signal handler for clean exit (#25969)
