### 2023-12-05 ddef998
Apply suggestions from code review (#26053)
Co-authored-by: Alex Lindsay <alexander.lindsay@inl.gov>

address_more_comments
