### 2023-12-06 eb7e763
Simplify MooseLinearFVVariable. (#25722)
