+++ framework/include/variables/MooseLinearVariableFV.h
@@ -65,6 +65,8 @@ public:
   using FieldVariablePhiGradient =
       typename MooseVariableField<OutputType>::FieldVariablePhiGradient;
   using FieldVariablePhiSecond = typename MooseVariableField<OutputType>::FieldVariablePhiSecond;
+  using FieldVariablePhiDivergence =
+      typename MooseVariableField<OutputType>::FieldVariablePhiDivergence;
   using ElemQpArg = Moose::ElemQpArg;
   using ElemSideQpArg = Moose::ElemSideQpArg;
   using ElemArg = Moose::ElemArg;
@@ -77,131 +79,13 @@ public:
   using typename MooseVariableField<OutputType>::GradientType;
 
   static InputParameters validParams();
-
   MooseLinearVariableFV(const InputParameters & parameters);
 
   virtual bool isFV() const override { return true; }
 
-  // TODO: many of these functions are not relevant to FV variables but are
-  // still called at various points from existing moose codepaths.  Ideally we
-  // would figure out how to remove calls to these functions and then allow
-  // throwing mooseError's from them instead of silently doing nothing (e.g.
-  // reinitNodes, reinitAux, prepareLowerD, etc.).
-
-  virtual void prepare() override final {}
-  virtual void prepareNeighbor() override final {}
-  virtual void prepareAux() override final {}
-  virtual void reinitNode() override final {}
-  virtual void reinitNodes(const std::vector<dof_id_type> & /*nodes*/) override final {}
-  virtual void reinitNodesNeighbor(const std::vector<dof_id_type> & /*nodes*/) override final {}
-  virtual void reinitAux() override final {}
-  virtual void reinitAuxNeighbor() override final {}
-  virtual void prepareLowerD() override final {}
-
-  virtual const dof_id_type & nodalDofIndex() const override final
-  {
-    mooseError("nodalDofIndex not supported by MooseLinearVariableFV");
-  }
-  virtual const dof_id_type & nodalDofIndexNeighbor() const override final
-  {
-    mooseError("nodalDofIndexNeighbor not supported by MooseLinearVariableFV");
-  }
-  virtual std::size_t phiSize() const override final
-  {
-    mooseError("phiSize not supported by MooseLinearVariableFVBase");
-  }
-  virtual std::size_t phiFaceSize() const override final
-  {
-    mooseError("phiFaceSize not supported by MooseLinearVariableFVBase");
-  }
-  virtual std::size_t phiNeighborSize() const override final
-  {
-    mooseError("phiNeighborSize not supported by MooseLinearVariableFVBase");
-  }
-  virtual std::size_t phiFaceNeighborSize() const override final
-  {
-    mooseError("phiFaceNeighborSize not supported by MooseLinearVariableFVBase");
-  }
-  virtual std::size_t phiLowerSize() const override final
-  {
-    mooseError("phiLowerSize not supported by MooseLinearVariableFVBase");
-  }
-
-  virtual void computeElemValuesFace() override {}
-  virtual void computeNeighborValuesFace() override {}
-  virtual void computeNeighborValues() override {}
-  virtual void computeLowerDValues() override final {}
-  virtual void computeNodalNeighborValues() override final {}
-  virtual void computeNodalValues() override final {}
-  virtual const std::vector<dof_id_type> & dofIndicesLower() const override final;
-
-  unsigned int numberOfDofs() const override final
-  {
-    mooseError("numberOfDofsNeighbor not supported by MooseLinearVariableFVBase");
-  }
-
-  virtual unsigned int numberOfDofsNeighbor() override final
-  {
-    mooseError("numberOfDofs not supported by MooseLinearVariableFVBase");
-  }
-
[…]
