+++ framework/src/variables/MooseLinearVariableFV.C
@@ -38,20 +38,6 @@ MooseLinearVariableFV<OutputType>::validParams()
   params.set<bool>("fv") = true;
   params.set<MooseEnum>("family") = "MONOMIAL";
   params.set<MooseEnum>("order") = "CONSTANT";
-  params.template addParam<bool>(
-      "two_term_boundary_expansion",
-      true,
-      "Whether to use a two-term Taylor expansion to calculate boundary face values. "
-      "If the two-term expansion is used, then the boundary face value depends on the "
-      "adjoining cell center gradient, which itself depends on the boundary face value. "
-      "Consequently an implicit solve is used to simultaneously solve for the adjoining cell "
-      "center gradient and boundary face value(s).");
-  MooseEnum face_interp_method("average skewness-corrected", "average");
-  params.template addParam<MooseEnum>("face_interp_method",
-                                      face_interp_method,
-                                      "Switch that can select between face interpoaltion methods.");
-  params.template addParam<bool>(
-      "cache_cell_gradients", true, "Whether to cache cell gradients or re-compute them.");
   return params;
 }
 
@@ -59,24 +45,8 @@ template <typename OutputType>
 MooseLinearVariableFV<OutputType>::MooseLinearVariableFV(const InputParameters & parameters)
   : MooseVariableField<OutputType>(parameters),
     _solution(this->_sys.currentSolution()),
-    _prev_elem(nullptr),
-    _two_term_boundary_expansion(this->isParamValid("two_term_boundary_expansion")
-                                     ? this->template getParam<bool>("two_term_boundary_expansion")
-                                     : true),
-    _cache_cell_gradients(this->isParamValid("cache_cell_gradients")
-                              ? this->template getParam<bool>("cache_cell_gradients")
-                              : true)
+    _prev_elem(nullptr)
 {
-  if (this->isParamValid("face_interp_method"))
-  {
-    const auto & interp_method = this->template getParam<MooseEnum>("face_interp_method");
-    if (interp_method == "average")
-      _face_interp_method = Moose::FV::InterpMethod::Average;
-    else if (interp_method == "skewness-corrected")
-      _face_interp_method = Moose::FV::InterpMethod::SkewCorrectedAverage;
-  }
-  else
-    _face_interp_method = Moose::FV::InterpMethod::Average;
 }
 
 template <typename OutputType>
@@ -94,23 +64,6 @@ MooseLinearVariableFV<OutputType>::fieldType() const
 }
 
 template <typename OutputType>
-OutputType
-MooseLinearVariableFV<OutputType>::getValue(const Elem * elem) const
-{
-  Moose::initDofIndices(const_cast<MooseLinearVariableFV<OutputType> &>(*this), *elem);
-  mooseAssert(this->_dof_indices.size() == 1, "Wrong size for dof indices");
-  OutputType value = (*this->_sys.currentSolution())(this->_dof_indices[0]);
-  return value;
-}
-
-template <typename OutputType>
-typename OutputTools<OutputType>::OutputGradient
-MooseLinearVariableFV<OutputType>::getGradient(const Elem * /*elem*/) const
-{
-  return {};
-}
-
-template <typename OutputType>
 bool
 MooseLinearVariableFV<OutputType>::isArray() const
 {
@@ -125,292 +78,47 @@ MooseLinearVariableFV<OutputType>::isVector() const
 }
 
 template <typename OutputType>
-std::pair<bool, const FVDirichletBCBase *>
-MooseLinearVariableFV<OutputType>::getDirichletBC(const FaceInfo & /*fi*/) const
+Real
+MooseLinearVariableFV<OutputType>::getElemValue(const Elem * const /*elem*/,
+                                                const StateArg & /*state*/) const
 {
[…]
