### 2023-12-06 b7b617f
Applied patch and re-added override to DerivativeParsedMaterial.h
