+++ modules/phase_field/src/kernels/NestKKSMultiACBulkC.C
@@ -1,252 +1,0 @@
-//* This file is part of the MOOSE framework
-//* https://www.mooseframework.org
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
-#include "NestKKSMultiACBulkC.h"
-
-registerMooseObject("PhaseFieldApp", NestKKSMultiACBulkC);
-
-InputParameters
-NestKKSMultiACBulkC::validParams()
-{
-  InputParameters params = KKSMultiACBulkBase::validParams();
-  params.addClassDescription("Multi-phase KKS model kernel (part 2 of 2) for the Bulk Allen-Cahn. "
-                             "This includes all terms dependent on chemical potential.");
-  params.addRequiredCoupledVar("global_cs", "Global concentrations, for example, c, b.");
-  params.addRequiredCoupledVar("all_etas", "Order parameters.");
-  params.addRequiredParam<std::vector<MaterialPropertyName>>(
-      "ci_names",
-      "Phase concentrations. They must have the same order as Fj_names and global_cs, for "
-      "example, c1, c2, b1, b2.");
-  return params;
-}
-
-NestKKSMultiACBulkC::NestKKSMultiACBulkC(const InputParameters & parameters)
-  : KKSMultiACBulkBase(parameters),
-    _c_names(coupledNames("global_cs")),
-    _c_map(getParameterJvarMap("global_cs")),
-    _num_c(coupledComponents("global_cs")),
-    _eta_names(coupledNames("all_etas")),
-    _eta_map(getParameterJvarMap("all_etas")),
-    _k(-1),
-    _ci_names(getParam<std::vector<MaterialPropertyName>>("ci_names")),
-    _ci_name_matrix(_num_c),
-    _prop_ci(_num_c),
-    _dcidetaj(_num_c),
-    _dcidb(_num_c),
-    _prop_d2hjdetaidetap(_num_j),
-    _dF1dc1(_num_c),
-    _d2F1dc1db1(_num_c),
-    _d2F1dc1darg(_n_args)
-{
-  for (unsigned int i = 0; i < _num_j; ++i)
-  {
-    // get order parameter names and variable indices
-    _eta_names[i] = getVar("all_etas", i)->name();
-
-    // Set _k to the position of the nonlinear variable in the list of etaj's
-    if (coupled("all_etas", i) == _var.number())
-      _k = i;
-  }
-
-  // initialize _ci_name_matrix and _prop_ci for easy reference
-  for (unsigned int m = 0; m < _num_c; ++m)
-  {
-    _ci_name_matrix[m].resize(_num_j);
-    _prop_ci[m].resize(_num_j);
-
-    for (unsigned int n = 0; n < _num_j; ++n)
-    {
-      _ci_name_matrix[m][n] = _ci_names[m * _num_j + n];
-      _prop_ci[m][n] = &getMaterialPropertyByName<Real>(_ci_name_matrix[m][n]);
-    }
-  }
-
-  // _dcidetaj and _dcidb are computed in KKSPhaseConcentrationMultiPhaseDerivatives
-  for (unsigned int m = 0; m < _num_c; ++m)
-  {
[…]
