+++ modules/phase_field/src/kernels/NestKKSMultiACBulkF.C
@@ -1,207 +1,0 @@
-//* This file is part of the MOOSE framework
-//* https://www.mooseframework.org
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
-#include "NestKKSMultiACBulkF.h"
-
-registerMooseObject("PhaseFieldApp", NestKKSMultiACBulkF);
-
-InputParameters
-NestKKSMultiACBulkF::validParams()
-{
-  InputParameters params = KKSMultiACBulkBase::validParams();
-  params.addClassDescription("KKS model kernel (part 1 of 2) for the Bulk Allen-Cahn. This "
-                             "includes all terms NOT dependent on chemical potential.");
-  params.addRequiredCoupledVar("global_cs", "The interpolated concentrations c, b, etc.");
-  params.addRequiredCoupledVar("all_etas", "Order parameters.");
-  params.addRequiredParam<std::vector<MaterialPropertyName>>(
-      "ci_names",
-      "Phase concentrations. They must have the same order as Fj_names and global_cs, for "
-      "example, c1, c2, b1, b2.");
-  params.addRequiredParam<Real>("wi", "Double well height parameter.");
-  params.addRequiredParam<MaterialPropertyName>(
-      "gi_name", "Base name for the double well function g_i(eta_i).");
-  return params;
-}
-
-NestKKSMultiACBulkF::NestKKSMultiACBulkF(const InputParameters & parameters)
-  : KKSMultiACBulkBase(parameters),
-    _c_names(coupledNames("global_cs")),
-    _c_map(getParameterJvarMap("global_cs")),
-    _num_c(coupledComponents("global_cs")),
-    _eta_names(coupledNames("all_etas")),
-    _eta_map(getParameterJvarMap("all_etas")),
-    _k(-1),
-    _ci_names(getParam<std::vector<MaterialPropertyName>>("ci_names")),
-    _dcidetaj(_num_c),
-    _dcidb(_num_c),
-    _wi(getParam<Real>("wi")),
-    _gi_name(getParam<MaterialPropertyName>("gi_name")),
-    _dgi(getMaterialPropertyDerivative<Real>("gi_name", _etai_name)),
-    _d2gi(getMaterialPropertyDerivative<Real>("gi_name", _etai_name, _etai_name)),
-    _d2hjdetaidetap(_num_j),
-    _dF1dc1(_num_c),
-    _dFidarg(_num_j)
-{
-  for (unsigned int i = 0; i < _num_j; ++i)
-  {
-    // get order parameter names and variable indices
-    _eta_names[i] = getVar("all_etas", i)->name();
-
-    // Set _k to the position of the nonlinear variable in the list of etaj's
-    if (coupled("all_etas", i) == _var.number())
-      _k = i;
-  }
-
-  // _dcideta and _dcidb are computed in KKSPhaseConcentrationDerivatives
-  for (unsigned int m = 0; m < _num_c; ++m)
-  {
-    _dcidetaj[m].resize(_num_j);
-    _dcidb[m].resize(_num_j);
-
-    for (unsigned int n = 0; n < _num_j; ++n)
-    {
-      _dcidetaj[m][n].resize(_num_j);
-      _dcidb[m][n].resize(_num_c);
[…]

[…]
