### 2023-12-06 43e7d3c
Apply suggestions from code review
Applying suggested changes from review.

Co-authored-by: Daniel Schwen <daniel@schwen.de>
