### 2023-12-06 de0f92f
Added gold exodus files for damped KKS solve and added fixed kernels
