### 2023-12-06 4f3be00
Address review: use Moose namespace, improve doco, add deprecated message
Co-Authored-By: Logan Harbour <loganhharbour@gmail.com>

Use Moose namespace instead of moose in modules
