+++ unit/src/LimitersTest.C
@@ -32,7 +32,7 @@ TEST(LimitersTest, limitVector)
   TensorValue<Real> grad(0);
 
   const char * argv[2] = {"foo", "\0"};
-  auto app = moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
+  auto app = Moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
   ReplicatedMesh mesh(app->comm(), /*dim=*/2);
   MeshTools::Generation::build_square(mesh, 2, 2);
   ElemInfo ei(mesh.elem_ptr(0));
