+++ unit/src/PositionsTest.C
@@ -18,7 +18,7 @@ TEST(Positions, getUninitialized)
 {
   // Create a minimal app that can create objects
   const char * argv[2] = {"foo", "\0"};
-  const auto & app = moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
+  const auto & app = Moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
   const auto & factory = &app->getFactory();
   app->parameters().set<bool>("minimal") = true;
   app->run();
@@ -85,7 +85,7 @@ TEST(Positions, getters)
 {
   // Create a minimal app that can create objects
   const char * argv[2] = {"foo", "\0"};
-  const auto & app = moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
+  const auto & app = Moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
   const auto & factory = &app->getFactory();
   app->parameters().set<bool>("minimal") = true;
   app->run();
