+++ unit/src/MooseCoordTest.C
@@ -194,7 +194,7 @@ TEST(MooseCoordTest, testLengthUnit)
 
   const auto nx = 2;
 
-  std::shared_ptr<MooseApp> app = moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
+  std::shared_ptr<MooseApp> app = Moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
 
   auto * factory = &app->getFactory();
   std::string mesh_type = "MeshGeneratorMesh";
