+++ unit/src/RestartableDataIOTest.C
@@ -25,7 +25,7 @@ void
 RestartableDataIOTest::SetUp()
 {
   const char * argv[2] = {"foo", "\0"};
-  _app = moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
+  _app = Moose::createMooseApp("MooseUnitApp", 1, (char **)argv);
 }
 
 std::map<std::string, RestartableDataIOTest::DataInfo>
