### 2023-12-06 e817aad
Allow different normals in neighbor face material
