### 2023-12-06 6691700
Implement clone method for ArbitraryQuadrature
