### 2023-12-06 46edf2a
Use quadrature clone method when creating new qrules in MooseVariableFE
