### 2023-12-06 31e5035
Copy q-points, JxW, and normals in Assembly::reinitFEFaceNeighbor
Also make a comment about the incorrectness of JxW neighbor
