### 2023-12-06 f4005d4
Add test objects to detect Qp calculations refs #26244
