+++ modules/navier_stokes/test/src/variables/INSFVPressureNoQpComputation.C
@@ -1,0 +1,17 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "INSFVPressureNoQpComputation.h"
+
+registerMooseObject("NavierStokesTestApp", INSFVPressureNoQpComputation);
+
+INSFVPressureNoQpComputation::INSFVPressureNoQpComputation(const InputParameters & params)
+  : INSFVPressureVariable(params)
+{
+}
