### 2023-12-06 9250a57
Add two tests to make sure: - we can detect when Qp calculations are requested - we do not request them in your average INSFV caclulation refs #26244
