### 2023-12-06 f8474a9
Add block restrictable SubdomainDivisions to allow a selection of blocks
