### 2023-12-06 87a8fb6
Use Moose namespace in modules and tutorials
