### 2023-12-06 c6399ca
Remove _neighbor_normal computation from reinitFEFaceNeighbor
We are reiniting with only an elem and no side so there is no
normal information to get
