### 2023-12-06 cc17f16
Revert "Allow different normals in neighbor face material"
This reverts commit 87914fe3048d59e84d2df66c96a3a9b07d2af29a.
