### 2023-12-06 a3b4dc3
Burn neighborNormals to the ground
