### 2023-12-07 3747eb2
Add initial data only capability
refs #26295
