+++ framework/src/mesh/MeshGeneratorMesh.C
@@ -22,6 +22,9 @@ MeshGeneratorMesh::validParams()
 
   params.addParam<std::string>("final_generator",
                                "The name of the mesh generator output to use for the final Mesh");
+  params.addParam<bool>("data_only",
+                        "Set to use data only generation for all generators except the final");
+
   params.addClassDescription("Mesh generated using mesh generators");
   return params;
 }
