### 2023-12-07 f5e95e7
Use neighbor data when _neighbor in Material
Refs #25666
