+++ framework/src/loops/ComputeLinearFVElementalThread.C
@@ -12,18 +12,22 @@
 #include "LinearFVKernel.h"
 #include "LinearFVElementalKernel.h"
 
-ComputeLinearFVElementalThread::ComputeLinearFVElementalThread(FEProblemBase & fe_problem,
-                                                               const unsigned int linear_system_num,
-                                                               const std::set<TagID> & tags)
-  : _fe_problem(fe_problem), _linear_system_number(linear_system_num), _tags(tags)
+ComputeLinearFVElementalThread::ComputeLinearFVElementalThread(
+    FEProblemBase & fe_problem,
+    const unsigned int linear_system_num,
+    const Moose::FV::LinearFVComputationMode mode,
+    const std::set<TagID> & tags)
+  : _fe_problem(fe_problem), _linear_system_number(linear_system_num), _mode(mode), _tags(tags)
 {
 }
 
 // Splitting Constructor
 ComputeLinearFVElementalThread::ComputeLinearFVElementalThread(ComputeLinearFVElementalThread & x,
                                                                Threads::split /*split*/)
-  : _fe_problem(x._fe_problem), _linear_system_number(x._linear_system_number), _tags(x._tags)
-
+  : _fe_problem(x._fe_problem),
+    _linear_system_number(x._linear_system_number),
+    _mode(x._mode),
+    _tags(x._tags)
 {
 }
 
@@ -34,19 +38,29 @@ ComputeLinearFVElementalThread::operator()(const ElemInfoRange & range)
   _tid = puid.id;
 
   std::vector<LinearFVElementalKernel *> kernels;
-  _fe_problem.theWarehouse()
-      .query()
-      .template condition<AttribSysNum>(_fe_problem.getLinearSystem(_linear_system_number).number())
-      .template condition<AttribSystem>("LinearFVElementalKernel")
-      .queryInto(kernels);
+  auto base_query =
+      _fe_problem.theWarehouse()
+          .query()
+          .condition<AttribSysNum>(_fe_problem.getLinearSystem(_linear_system_number).number())
+          .condition<AttribSystem>("LinearFVElementalKernel")
+          .condition<AttribThread>(_tid);
+
+  if (_mode == Moose::FV::LinearFVComputationMode::Matrix)
+    base_query.condition<AttribMatrixTags>(_tags).queryInto(kernels);
+  else
+    base_query.condition<AttribVectorTags>(_tags).queryInto(kernels);
 
   // Iterate over all the elements in the range
   for (const auto & elem_info : range)
     for (auto kernel : kernels)
     {
       kernel->setCurrentElemInfo(elem_info);
-      kernel->addMatrixContribution();
-      kernel->addRightHandSideContribution();
+      if (_mode == Moose::FV::LinearFVComputationMode::Matrix ||
+          _mode == Moose::FV::LinearFVComputationMode::FullSystem)
+        kernel->addMatrixContribution();
+      if (_mode == Moose::FV::LinearFVComputationMode::RHS ||
+          _mode == Moose::FV::LinearFVComputationMode::FullSystem)
+        kernel->addRightHandSideContribution();
     }
 }
 
