### 2023-12-07 e48844c
Create Assembly::_current_neighbor_ref_points data member
