### 2023-12-07 a869ba3
Stop using qRuleNeighbor
