+++ framework/src/materials/Material.C
@@ -37,8 +37,7 @@ Material::Material(const InputParameters & parameters)
     _neighbor(_material_data_type == Moose::NEIGHBOR_MATERIAL_DATA),
     _q_point(_bnd ? (_neighbor ? _assembly.qPointsFaceNeighbor() : _assembly.qPointsFace())
                   : _assembly.qPoints()),
-    _qrule(_bnd ? (_neighbor ? _assembly.qRuleNeighbor() : _assembly.qRuleFace())
-                : _assembly.qRule()),
+    _qrule(_bnd ? _assembly.qRuleFace() : _assembly.qRule()),
     _JxW(_bnd ? _assembly.JxWFace() : _assembly.JxW()),
     _current_elem(_neighbor ? _assembly.neighbor() : _assembly.elem()),
     _current_subdomain_id(_neighbor ? _assembly.currentNeighborSubdomainID()
