+++ framework/src/problems/FEProblemBase.C
@@ -2072,7 +2072,7 @@ FEProblemBase::reinitNeighbor(const Elem * elem, unsigned int side, const THREAD
     // because the physical points coming from the element don't actually lie on the neighbor.
     // Moreover, what's the point of doing another physical point inversion in other cases? We only
     // care about the reference points which we can just take from the undisplaced computation
-    const auto & displaced_ref_pts = _assembly[tid][0]->qRuleNeighbor()->get_points();
+    const auto & displaced_ref_pts = _assembly[tid][0]->currentNeighborRefPoints();
 
     _displaced_problem->reinitNeighbor(
         _displaced_mesh->elemPtr(elem->id()), side, tid, &displaced_ref_pts);
@@ -3580,7 +3580,7 @@ FEProblemBase::reinitMaterialsNeighbor(const SubdomainID blk_id,
                 "The provided blk_id " << blk_id << " and neighbor subdomain ID "
                                        << neighbor->subdomain_id() << " do not match.");
 
-    unsigned int n_points = _assembly[tid][0]->qRuleNeighbor()->n_points();
+    unsigned int n_points = _assembly[tid][0]->qRuleFace()->n_points();
 
     auto & neighbor_material_data = _neighbor_material_props.getMaterialData(tid);
     neighbor_material_data.resize(n_points);
