### 2023-12-07 fa21a7d
Revert "Stop using qRuleNeighbor"
This reverts commit 15323b63bd94f2a393fe28b9eeee65cdd94df528.
