+++ framework/src/variables/MooseVariableDataFV.C
@@ -100,6 +100,7 @@ MooseVariableDataFV<OutputType>::uDot() const
 {
   if (_sys.solutionUDot())
   {
+    static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
     _need_u_dot = true;
     return _u_dot;
   }
@@ -114,6 +115,7 @@ MooseVariableDataFV<OutputType>::uDotDot() const
 {
   if (_sys.solutionUDotDot())
   {
+    static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
     _need_u_dotdot = true;
     return _u_dotdot;
   }
@@ -129,6 +131,7 @@ MooseVariableDataFV<OutputType>::uDotOld() const
 {
   if (_sys.solutionUDotOld())
   {
+    static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
     _need_u_dot_old = true;
     return _u_dot_old;
   }
@@ -144,6 +147,7 @@ MooseVariableDataFV<OutputType>::uDotDotOld() const
 {
   if (_sys.solutionUDotDotOld())
   {
+    static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
     _need_u_dotdot_old = true;
     return _u_dotdot_old;
   }
@@ -154,11 +158,20 @@ MooseVariableDataFV<OutputType>::uDotDotOld() const
 }
 
 template <typename OutputType>
+const typename MooseVariableDataFV<OutputType>::FieldVariableValue &
+MooseVariableDataFV<OutputType>::sln(Moose::SolutionState state) const
+{
+  static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
+  return MooseVariableDataBase<OutputType>::sln(state);
+}
+
+template <typename OutputType>
 const typename MooseVariableDataFV<OutputType>::FieldVariableGradient &
 MooseVariableDataFV<OutputType>::gradSlnDot() const
 {
   if (_sys.solutionUDot())
   {
+    static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
     _need_grad_dot = true;
     return _grad_u_dot;
   }
@@ -173,6 +186,7 @@ MooseVariableDataFV<OutputType>::gradSlnDotDot() const
 {
   if (_sys.solutionUDotDot())
   {
+    static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
     _need_grad_dotdot = true;
     return _grad_u_dotdot;
   }
@@ -186,6 +200,7 @@ template <typename OutputType>
 const typename MooseVariableDataFV<OutputType>::FieldVariableSecond &
 MooseVariableDataFV<OutputType>::secondSln(Moose::SolutionState state) const
 {
+  static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
   switch (state)
   {
     case Moose::Current:
@@ -223,6 +238,7 @@ template <typename OutputType>
 const typename MooseVariableDataFV<OutputType>::FieldVariableCurl &
 MooseVariableDataFV<OutputType>::curlSln(Moose::SolutionState state) const
 {
+  static_cast<const MooseVariableFV<OutputType> *>(&_var)->requireQpComputations();
   switch (state)
   {
     case Moose::Current:
