+++ modules/navier_stokes/include/variables/INSFVVariable.h
@@ -26,7 +26,7 @@ public:
   void computeElemValuesFace() override;
   void computeNeighborValuesFace() override;
   void computeNeighborValues() override;
-  void requireQpComputations() override { _qp_calculations = true; }
+  void requireQpComputations() const override { _qp_calculations = true; }
 
 protected:
   /**
@@ -36,8 +36,9 @@ protected:
 
 private:
   /// Whether to pre-initialize variable data for use in traditional MOOSE quadrature point based
-  /// objects
-  bool _qp_calculations;
+  /// objects. This can be changed to true if objects request variable values in a way that is
+  /// generated by quadrature point based calculations
+  mutable bool _qp_calculations;
 };
 
 inline void
