### 2023-12-07 c62be53
Added a templated function isRelSmall to check for step size tolerance. Fixes bug that caused failed framework test
