### 2023-12-07 9c9f699
Address Alex review on block restriction feature
