+++ python/TestHarness/tests/test_Ignore.py
@@ -97,23 +97,23 @@ class TestHarnessTester(TestHarnessTestCase):
         self.assertRegex(output.decode('utf-8'), 'test_harness\.ignore_multi_prereq_dependency.*?OK')
 
         # Run a multiple caveat prereq test by manually supplying each caveat
-        output = self.runTests('-i', 'ignore_multiple_prereq', '--ignore', 'prereq skip heavy compiler platform')
+        output = self.runTests('-i', 'ignore_multiple_prereq', '--ignore', 'prereq skip compiler platform')
         self.assertRegex(output.decode('utf-8'), 'test_harness\.always_skipped.*?OK')
         self.assertRegex(output.decode('utf-8'), 'test_harness\.ignore_multi_prereq_dependency.*?OK')
 
         # Skip a multiple caveat prereq test by not supplying enough caveats to ignore
-        output = self.runTests('--no-color', '-i', 'ignore_multiple_prereq', '--ignore', 'prereq skip heavy compiler')
+        output = self.runTests('--no-color', '-i', 'ignore_multiple_prereq', '--ignore', 'prereq skip compiler')
         self.assertRegex(output.decode('utf-8'), 'test_harness\.always_skipped.*?OK')
         self.assertRegex(output.decode('utf-8'), 'test_harness\.ignore_multi_prereq_dependency.*? \[PLATFORM!=NON_EXISTENT\] SKIP')
 
         # Check that a multiple caveat dependency test runs when its prereq test is skipped
         # This test may seem redundant, but `prereq` is handled differently than the other caveats
-        output = self.runTests('--no-color', '-i', 'ignore_multiple_prereq', '--ignore', 'prereq heavy compiler platform')
+        output = self.runTests('--no-color', '-i', 'ignore_multiple_prereq', '--ignore', 'prereq compiler platform')
         self.assertRegex(output.decode('utf-8'), 'test_harness\.always_skipped.*? \[ALWAYS SKIPPED\] SKIP')
         self.assertRegex(output.decode('utf-8'), 'test_harness\.ignore_multi_prereq_dependency.*?OK')
 
         # Check that by supplying a very specific set of ignored paramaters, we
         # can properly trigger a skipped dependency scenario
-        output = self.runTests('--no-color', '-i', 'ignore_multiple_prereq', '--ignore', 'heavy compiler platform')
+        output = self.runTests('--no-color', '-i', 'ignore_multiple_prereq', '--ignore', 'compiler platform')
         self.assertRegex(output.decode('utf-8'), 'test_harness\.always_skipped.*? \[ALWAYS SKIPPED\] SKIP')
         self.assertRegex(output.decode('utf-8'), 'test_harness\.ignore_multi_prereq_dependency.*? \[SKIPPED DEPENDENCY\] SKIP')
