### 2023-12-07 d114430
Relax assertion for _current_neighbor_subdomain_id in reinitNeighborFaceRef
This method is used heavily by mortar and we don't call the set
neighbor subdomain ID API from there, and I don't think we need to
