### 2023-12-07 8a796f2
Add a base class for solving the diffusion equation
