### 2023-12-07 43c3f2b
Add a derived physics class to solve diffusion with CG
