+++ framework/src/physics/DiffusionCG.C
@@ -1,0 +1,230 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "DiffusionCG.h"
+#include "MooseVariableBase.h"
+
+// Register the actions for the objects actually used
+registerMooseAction("MooseApp", DiffusionCG, "add_kernel");
+registerMooseAction("MooseApp", DiffusionCG, "add_bc");
+registerMooseAction("MooseApp", DiffusionCG, "add_variable");
+
+InputParameters
+DiffusionCG::validParams()
+{
+  InputParameters params = DiffusionPhysicsBase::validParams();
+  params.addClassDescription(
+      "Discretizes a diffusion equation with the continuous Galerkin finite element method");
+  params.addParam<bool>(
+      "automatic_differentiation",
+      true,
+      "Whether to use automatic differentiation for all the terms in the equation");
+  params.transferParam<MooseEnum>(MooseVariableBase::validParams(), "order", "variable_order");
+
+  return params;
+}
+
+DiffusionCG::DiffusionCG(const InputParameters & parameters)
+  : DiffusionPhysicsBase(parameters), _use_ad(getParam<bool>("automatic_differentiation"))
+{
+}
+
+void
+DiffusionCG::addFEKernels()
+{
+  // Diffusion term
+  {
+    // Select the kernel type based on the user parameters
+    std::string kernel_type;
+    if (isParamValid("diffusivity_matprop"))
+      kernel_type = _use_ad ? "ADMatDiffusion" : "MatDiffusion";
+    else if (isParamValid("diffusivity_functor"))
+    {
+      const auto d = getParam<MooseFunctorName>("diffusivity_functor");
+      if (getProblem().hasFunction(d))
+        kernel_type = "FunctionDiffusion";
+      else
+        paramError(
+            "diffusivity_functor", "No diffusion kernel implemented for the source type of", d);
+    }
+    else
+      kernel_type = _use_ad ? "ADDiffusion" : "Diffusion";
+    InputParameters params = getFactory().getValidParams(kernel_type);
+    params.set<NonlinearVariableName>("variable") = _var_name;
+
+    // Transfer the diffusivity parameter from the Physics to the kernel
+    if (isParamValid("diffusivity_matprop"))
+      params.set<MaterialPropertyName>("diffusivity") =
+          getParam<MaterialPropertyName>("diffusivity_matprop");
+    else if (isParamValid("diffusivity_functor"))
+      params.set<MaterialPropertyName>("diffusivity") =
+          getParam<MaterialPropertyName>("diffusivity_matprop");
+
+    getProblem().addKernel(kernel_type, name() + "_" + _var_name + "_diffusion", params);
+  }
+
+  // Source term
+  if (isParamValid("source_functor"))
+  {
+    // Select the kernel type based on the user parameters
+    std::string kernel_type;
+    const auto & source = getParam<MooseFunctorName>("source_functor");
+    if (MooseUtils::parsesToReal(source) || getProblem().hasFunction(source) ||
+        getProblem().hasPostprocessorValueByName(source))
+      kernel_type = _use_ad ? "ADBodyForce" : "BodyForce";
+    else if (getProblem().hasVariable(source))
+      kernel_type = _use_ad ? "ADCoupledForce" : "CoupledForce";
+    else
+      paramError("source_functor",
+                 "No kernel defined for a source term in CG for the type of '",
+                 source,
[…]
