+++ framework/src/physics/DiffusionFV.C
@@ -1,0 +1,177 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "DiffusionFV.h"
+#include "MooseUtils.h"
+
+// Register the actions for the objects actually used
+registerMooseAction("MooseApp", DiffusionFV, "add_fv_kernel");
+registerMooseAction("MooseApp", DiffusionFV, "add_fv_bc");
+registerMooseAction("MooseApp", DiffusionFV, "add_variable");
+
+InputParameters
+DiffusionFV::validParams()
+{
+  InputParameters params = DiffusionPhysicsBase::validParams();
+  params.addClassDescription(
+      "Add the heat conduction physics discretized with cell-centered finite volume");
+  // No kernel implemented in the framework for a material property diffusivity
+  params.suppressParameter<MaterialPropertyName>("diffusivity_matprop");
+  return params;
+}
+
+DiffusionFV::DiffusionFV(const InputParameters & parameters) : DiffusionPhysicsBase(parameters) {}
+
+void
+DiffusionFV::addFVKernels()
+{
+  // Diffusion term
+  {
+    const std::string kernel_type = "FVDiffusion";
+    InputParameters params = getFactory().getValidParams(kernel_type);
+    params.set<NonlinearVariableName>("variable") = _var_name;
+    params.set<MooseFunctorName>("coeff") = isParamValid("diffusivity_functor")
+                                                ? getParam<MooseFunctorName>("diffusivity_functor")
+                                                : "1";
+    getProblem().addFVKernel(kernel_type, name() + "_" + _var_name + "_diffusion", params);
+    // This should not be necessary but it is.
+    Action::addRelationshipManagers(Moose::RelationshipManagerType::COUPLING, params);
+  }
+  // Source term
+  if (isParamValid("source_functor"))
+  {
+    // Select the kernel type based on the user parameters
+    std::string kernel_type;
+
+    const auto & source = getParam<MooseFunctorName>("source_functor");
+    if (MooseUtils::parsesToReal(source) || getProblem().hasFunction(source) ||
+        getProblem().hasPostprocessorValueByName(source))
+      kernel_type = "FVBodyForce";
+    else if (getProblem().hasVariable(source))
+      kernel_type = "FVCoupledForce";
+    else
+      paramError("source_functor",
+                 "No kernel defined for a source term in FV for the type used for '",
+                 source,
+                 "'");
+
+    InputParameters params = getFactory().getValidParams(kernel_type);
+    params.set<NonlinearVariableName>("variable") = _var_name;
+
+    // Transfer the source and coefficient parameter from the Physics to the kernel
+    const auto coef = getParam<Real>("source_coef");
+    if (MooseUtils::parsesToReal(source))
+      params.set<Real>("value") = MooseUtils::convert<Real>(source) * coef;
+    else if (getProblem().hasFunction(source))
+    {
+      params.set<Real>("value") = coef;
+      params.set<FunctionName>("function") = source;
+    }
+    else if (getProblem().hasPostprocessorValueByName(source))
+    {
+      params.set<Real>("value") = coef;
+      params.set<PostprocessorName>("postprocessor") = source;
+    }
+    else
+    {
+      params.set<Real>("coef") = coef;
[…]
