+++ framework/src/problems/FEProblemBase.C
@@ -6079,6 +6079,48 @@ FEProblemBase::resetState()
   }
 }
 
+void
+FEProblemBase::solveLinearSystem(const unsigned int linear_sys_num)
+{
+  TIME_SECTION("solve", 1, "Solving", false);
+
+  setCurrentLinearSystem(linear_sys_num);
+
+  // This prevents stale dof indices from lingering around and possibly leading to invalid reads
+  // and writes. Dof indices may be made stale through operations like mesh adaptivity
+
+#if PETSC_RELEASE_LESS_THAN(3, 12, 0)
+  Moose::PetscSupport::petscSetOptions(
+      _petsc_options, _solver_params); // Make sure the PETSc options are setup for this app
+#else
+  // Now this database will be the default
+  // Each app should have only one database
+  if (!_app.isUltimateMaster())
+    PetscOptionsPush(_petsc_option_data_base);
+  // We did not add PETSc options to database yet
+  if (!_is_petsc_options_inserted)
+  {
+    Moose::PetscSupport::petscSetOptions(_petsc_options, _solver_params);
+    _is_petsc_options_inserted = true;
+  }
+#endif
+  Moose::setSolverDefaults(*this);
+
+  // Setup the output system for printing linear/nonlinear iteration information
+  initPetscOutput();
+
+  if (_solve)
+    _current_linear_sys->solve();
+
+  if (_solve)
+    _current_linear_sys->update();
+
+#if !PETSC_RELEASE_LESS_THAN(3, 12, 0)
+  if (!_app.isUltimateMaster())
+    PetscOptionsPop();
+#endif
+}
+
 bool
 FEProblemBase::nlConverged(const unsigned int nl_sys_num)
 {
@@ -6361,6 +6403,23 @@ FEProblemBase::computeResidualAndJacobian(const NumericVector<Number> & soln,
                                           NumericVector<Number> & residual,
                                           SparseMatrix<Number> & jacobian)
 {
+  // vector tags
+  {
+    _current_nl_sys->associateVectorToTag(residual, _current_nl_sys->residualVectorTag());
+    const auto & residual_vector_tags = getVectorTags(Moose::VECTOR_TAG_RESIDUAL);
+
+    // We filter out tags which do not have associated vectors in the current nonlinear
+    // system. This is essential to be able to use system-dependent residual tags.
+    selectVectorTagsFromSystem(*_current_nl_sys, residual_vector_tags, _fe_vector_tags);
+
+    setCurrentResidualVectorTags(_fe_vector_tags);
+  }
+
+  // matrix tags
+  {
+    selectMatrixTagsFromSystem(*_current_nl_sys, getMatrixTags(), _fe_matrix_tags);
+  }
+
   try
   {
     try
@@ -7112,6 +7171,108 @@ FEProblemBase::computeLinearSystemMatrixTags(const NumericVector<Number> & soln,
   _current_execute_on_flag = EXEC_NONE;
   _current_linear_sys->disassociateMatrixFromTag(system_matrix,
                                                  _current_linear_sys->systemMatrixTag());
+}
+
+void
+FEProblemBase::computeLinearSystemSys(LinearImplicitSystem & sys,
+                                      SparseMatrix<Number> & system_matrix,
+                                      NumericVector<Number> & rhs)
+{
+  TIME_SECTION("computeLinearSystemSys", 5);
+
+  setCurrentNonlinearSystem(sys.number());
+
+  // We are using the residual tag system for right hand sides so we fetch everything
+  const auto & vector_tags = getVectorTags(Moose::VECTOR_TAG_RESIDUAL);
+
+  // We filter out tags which do not have associated vectors in the current
[…]
