### 2023-12-07 da9feee
Remove bad calls to reinite*Neighbor with lower-d elem
We are calling through to reinit methods with a lower-d elem with
reference points from the higher-d elem. This is just plain wrong
