### 2023-12-07 3448f80
Remove incorrect computations of neighor JxW in Assembly
And consequently remove our uses of that bad computation in the repo
