+++ framework/src/base/Assembly.C
@@ -1614,8 +1614,6 @@ Assembly::reinitFEFaceNeighbor(const Elem * neighbor, const std::vector<Point> &
 
   _current_q_points_face_neighbor.shallowCopy(
       const_cast<std::vector<Point> &>(_holder_fe_face_neighbor_helper[neighbor_dim]->get_xyz()));
-  _current_JxW_neighbor.shallowCopy(
-      const_cast<std::vector<Real> &>(_holder_fe_face_neighbor_helper[neighbor_dim]->get_JxW()));
 }
 
 void
@@ -2006,14 +2004,6 @@ Assembly::reinitElemAndNeighbor(const Elem * elem,
   }
 
   _current_neighbor_side_elem = &_current_neighbor_side_elem_builder(*neighbor, neighbor_side);
-
-  if (_need_JxW_neighbor)
-    // first do the side element. We need to do this to at a minimum get the correct neighbor
-    // face physical q points, and neighbor normals for the neighbor face. *Note* that while this
-    // below call is triggered by the _need_JxW_neighbor flag, the irony is that the JxW that comes
-    // out of this is *wrong* because we are providing the "quadrature" points but no weights.
-    // Consequently dummy weights (equal to 1) are used
-    reinitFEFaceNeighbor(_current_neighbor_side_elem, *reference_points_ptr);
 
   reinitFEFaceNeighbor(neighbor, *reference_points_ptr);
   reinitNeighbor(neighbor, *reference_points_ptr);
@@ -2430,10 +2420,6 @@ Assembly::reinitNeighborAtPhysical(const Elem * neighbor,
   // first do the side element
   reinitFEFaceNeighbor(_current_neighbor_side_elem, reference_points);
   reinitNeighbor(_current_neighbor_side_elem, reference_points);
-  // compute JxW on the neighbor's face
-  unsigned int neighbor_side_dim = _current_neighbor_side_elem->dim();
-  _current_JxW_neighbor.shallowCopy(const_cast<std::vector<Real> &>(
-      _holder_fe_face_neighbor_helper[neighbor_side_dim]->get_JxW()));
 
   reinitFEFaceNeighbor(neighbor, reference_points);
   reinitNeighbor(neighbor, reference_points);
