### 2023-12-07 530ba88
Remove neighbor JxW from Assembly/NodeFaceConstraint
