+++ framework/src/constraints/NodeFaceConstraint.C
@@ -80,8 +80,7 @@ NodeFaceConstraint::NodeFaceConstraint(const InputParameters & parameters)
     _dof_map(_sys.dofMap()),
     _node_to_elem_map(_mesh.nodeToElemMap()),
 
-    _overwrite_secondary_residual(true),
-    _primary_JxW(_assembly.JxWNeighbor())
+    _overwrite_secondary_residual(true)
 {
   addMooseVariableDependency(&_var);
 
