### 2023-12-07 ccf6242
Revert "Remove neighbor JxW from Assembly/NodeFaceConstraint"
This reverts commit 530ba88967bf9f22c4257ec2cf740e8757652eca.
