+++ framework/src/constraints/NodeFaceConstraint.C
@@ -80,7 +80,8 @@ NodeFaceConstraint::NodeFaceConstraint(const InputParameters & parameters)
     _dof_map(_sys.dofMap()),
     _node_to_elem_map(_mesh.nodeToElemMap()),
 
-    _overwrite_secondary_residual(true)
+    _overwrite_secondary_residual(true),
+    _primary_JxW(_assembly.JxWNeighbor())
 {
   addMooseVariableDependency(&_var);
 
