### 2023-12-07 414c68f
Add k-NN search feature in `ValueCache` (#26276)
