### 2023-12-07 cedc170
Create and use Assembly::computeNeighborJxW
