### 2023-12-07 414ee46
Revert "Create and use Assembly::computeNeighborJxW"
This reverts commit cedc170ee6acf839d9040b1e9ebb40ec314a10a3.
