+++ framework/include/utils/InputParameters.h
@@ -961,6 +961,7 @@ public:
   template <typename T>
   void transferParam(const InputParameters & source_param,
                      const std::string & name,
+                     const std::string & new_name = "",
                      const std::string & new_description = "");
 
   /**
@@ -987,7 +988,7 @@ private:
   void setParameters() {}
 
   /**
-   * Helper that uses overloading to distinguish adding commandline parameters of
+   * Helper that uses overloading to distinguish adding command-line parameters of
    * a scalar and a vector kind. Vector parameters are options that may appear multiple
    * times on the command line (like -i).
    */
@@ -1867,9 +1868,11 @@ template <typename T>
 void
 InputParameters::transferParam(const InputParameters & source_params,
                                const std::string & name_in,
+                               const std::string & new_name,
                                const std::string & new_description)
 {
   const auto name = source_params.checkForRename(std::string(name_in));
+  const auto p_name = (new_name == "") ? name_in : new_name;
   if (!source_params.have_parameter<T>(name) && !source_params.hasCoupledValue(name))
     mooseError("The '",
                name_in,
@@ -1885,18 +1888,17 @@ InputParameters::transferParam(const InputParameters & source_params,
 
   if (source_params.isParamRequired(name))
   {
-    // Use the name_in that was specified in the transfer
     // Check for a variable parameter
     if (source_params.hasCoupledValue(name))
-      addRequiredCoupledVar(name_in, description);
+      addRequiredCoupledVar(p_name, description);
     // Enums parameters have a default list of options
     else if constexpr (std::is_same_v<MooseEnum, T> || std::is_same_v<MultiMooseEnum, T>)
-      addRequiredParam<T>(name_in, source_params.get<T>(name), description);
+      addRequiredParam<T>(p_name, source_params.get<T>(name), description);
     else if (source_params.isRangeChecked(name))
       addRequiredRangeCheckedParam<T>(
-          name_in, source_params.rangeCheckedFunction(name), description);
+          p_name, source_params.rangeCheckedFunction(name), description);
     else
-      addRequiredParam<T>(name_in, description);
+      addRequiredParam<T>(p_name, description);
   }
   else
   {
@@ -1904,41 +1906,41 @@ InputParameters::transferParam(const InputParameters & source_params,
     if (source_params.hasCoupledValue(name))
     {
       if (!source_params.hasDefaultCoupledValue(name))
-        addCoupledVar(name_in, description);
+        addCoupledVar(p_name, description);
       else if (source_params.numberDefaultCoupledValues(name) == 1)
-        addCoupledVar(name_in, source_params.defaultCoupledValue(name), description);
+        addCoupledVar(p_name, source_params.defaultCoupledValue(name), description);
       else
       {
         std::vector<Real> coupled_values;
         for (const auto i : make_range(source_params.numberDefaultCoupledValues(name)))
           coupled_values.push_back(source_params.defaultCoupledValue(name, i));
-        addCoupledVar(name_in, coupled_values, description);
+        addCoupledVar(p_name, coupled_values, description);
       }
     }
     else if (source_params.isRangeChecked(name))
     {
       if (source_params.hasDefault(name))
-        addRangeCheckedParam<T>(name_in,
+        addRangeCheckedParam<T>(p_name,
                                 source_params.get<T>(name),
                                 source_params.rangeCheckedFunction(name),
                                 description);
       else
-        addRangeCheckedParam<T>(name_in, source_params.rangeCheckedFunction(name), description);
+        addRangeCheckedParam<T>(p_name, source_params.rangeCheckedFunction(name), description);
     }
[…]
