### 2023-12-07 ef05443
Set some default preconditioning options for solving diffusion
