### 2023-12-07 38abec1
Implement solve for linear solver. (#25722)
