### 2023-12-08 e827284
Set Petsc options before every execute call in optimizeSolve to make sure Automatic Adjoint Petsc Options do not overwrite this. This allows the finite difference gradient check to be output to the terminal. closes #26262
