### 2023-12-08 a44327e
AD-ify cohesive zone modeling tools.
The intent of this update is to use the same tools in the modeling of
cohesive zones when using a "weigthed" or mortar approach to the problem
from the mortar user objects.
