+++ modules/tensor_mechanics/src/materials/cohesive_zone_model/CZMComputeDisplacementJumpTotalLagrangian.C
@@ -83,7 +83,7 @@ void
 CZMComputeDisplacementJumpTotalLagrangianTempl<is_ad>::computeRotationMatrices()
 {
   _czm_reference_rotation[_qp] =
-      CohesiveZoneModelTools::computeReferenceRotation(_normals[_qp], this->_mesh.dimension());
+      CohesiveZoneModelTools::computeReferenceRotationTempl<is_ad>(_normals[_qp], this->_mesh.dimension());
   computeFandR();
   _czm_total_rotation[_qp] = _R[_qp] * _czm_reference_rotation[_qp];
 }
