+++ modules/tensor_mechanics/include/utils/CohesiveZoneModelTools.h
@@ -12,37 +12,121 @@
 #include "RankFourTensor.h"
 #include "RankTwoTensor.h"
 #include "libmesh/vector_value.h"
+#include "MooseTypes.h"
+#include "RotationMatrix.h"
 
 namespace CohesiveZoneModelTools
 {
-
-RealVectorValue computedVnormdV(const RealVectorValue & V);
+template <bool is_ad>
+GenericRealVectorValue<is_ad>
+computedVnormdVTempl(const GenericRealVectorValue<is_ad> & V)
+{
+  return V / V.norm();
+}
 
 /// compute the derivative of the rotation matrix, R=FU^-1, using Chen and Wheeler 1993
-RankFourTensor computedRdF(const RankTwoTensor & R, const RankTwoTensor & U);
+template <bool is_ad>
+GenericRankFourTensor<is_ad>
+computedRdFTempl(const GenericRankTwoTensor<is_ad> & R, const GenericRankTwoTensor<is_ad> & U)
+{
+  const auto Uhat = U.trace() * RankTwoTensor::Identity() - U;
+  unsigned int k, l, m, n, p, q;
+  const auto Uhat_det = Uhat.det();
+
+  GenericRankFourTensor<is_ad> dR_dF;
+  for (k = 0; k < 3; k++)
+    for (l = 0; l < 3; l++)
+      for (m = 0; m < 3; m++)
+        for (n = 0; n < 3; n++)
+        {
+          dR_dF(k, l, m, n) = 0.;
+          for (p = 0; p < 3; p++)
+            for (q = 0; q < 3; q++)
+              dR_dF(k, l, m, n) +=
+                  R(k, p) * (Uhat(p, q) * R(m, q) * Uhat(n, l) - Uhat(p, n) * R(m, q) * Uhat(q, l));
+
+          dR_dF(k, l, m, n) /= Uhat_det;
+        }
+
+  return dR_dF;
+}
 
 /// compute the derivative of F^-1 w.r.t. to F
-RankFourTensor computedFinversedF(const RankTwoTensor & F_inv);
+RankFourTensor
+computedFinversedF(const RankTwoTensor & F_inv);
 
 /// compute the area ratio betweeen the deformed and undeformed configuration, and its derivatives w.r.t. the deformation gradient, F
 ///@{
-Real computeAreaRatio(const RankTwoTensor & FinvT, const Real & J, const RealVectorValue & N);
-RankTwoTensor computeDAreaRatioDF(const RankTwoTensor & FinvT,
-                                  const RealVectorValue & N,
-                                  const Real & J,
-                                  const RankFourTensor & DFinv_DF);
+template <bool is_ad>
+GenericReal<is_ad>
+computeAreaRatioTempl(const RankTwoTensor & FinvT, const Real & J, const RealVectorValue & N)
+{
+  return J * (FinvT * N).norm();
+}
+
+template <bool is_ad>
+GenericRankTwoTensor<is_ad>
+computeDAreaRatioDFTempl(const GenericRankTwoTensor<is_ad> & FinvT,
+                    const GenericRealVectorValue<is_ad> & N,
+                    const GenericReal<is_ad> & J,
+                    const GenericRankFourTensor<is_ad> & DFinv_DF)
+{
+  const auto Fitr_N = FinvT * N;
+  const auto Fitr_N_norm = Fitr_N.norm();
+  GenericRankTwoTensor<is_ad> R2temp;
+  for (unsigned int l = 0; l < 3; l++)
+    for (unsigned int m = 0; m < 3; m++)
+    {
+      R2temp(l, m) = 0;
+      for (unsigned int i = 0; i < 3; i++)
[…]
