### 2023-12-08 00dea11
Avoid using a temporary object to initialize TwoMaterialPropertyInterface
