### 2023-12-08 b50a658
Attach assembly function to the linear solver. (#25722)
