+++ framework/include/utils/MooseTypes.h
@@ -1068,6 +1068,9 @@ DerivativeStringClass(ParsedFunctionExpression);
 /// System name support of multiple nonlinear systems on the same mesh
 DerivativeStringClass(NonlinearSystemName);
 
+/// System name support of multiple linear systems on the same mesh
+DerivativeStringClass(LinearSystemName);
+
 /// Command line argument, specialized to handle quotes in vector arguments
 DerivativeStringClass(CLIArgString);
 
