+++ framework/include/systems/LinearSystem.h
@@ -144,6 +144,11 @@ public:
   void computeSystemMatrix(SparseMatrix<Number> & matrix);
 
   /**
+   * Return a reference to the stored linear implicit system
+   */
+  LinearImplicitSystem & linearImplicitSystem();
+
+  /**
    * Called at the beginning of the time step
    */
   void onTimestepBegin();
