+++ framework/src/parser/Builder.C
@@ -1094,6 +1094,7 @@ Builder::extractParams(const std::string & prefix, InputParameters & p)
         setscalar(PostprocessorName, PostprocessorName);
         setscalar(ExecutorName, string);
         setscalar(NonlinearSystemName, string);
+        setscalar(LinearSystemName, string);
         setscalar(CLIArgString, string);
 
         // Moose Compound Scalars
@@ -1170,6 +1171,7 @@ Builder::extractParams(const std::string & prefix, InputParameters & p)
         setvector(ReporterValueName, string);
         setvector(ExecutorName, string);
         setvector(NonlinearSystemName, string);
+        setvector(LinearSystemName, string);
 
         // map types
         setmap(string, Real);
