### 2023-12-08 a867ff9
Add documentation to example diffusion physics, refs #25642
