+++ framework/doc/content/syntax/Physics/DiffusionFV/index.md
@@ -1,0 +1,1 @@
+../../../source/physics/DiffusionFV.md
