+++ framework/doc/content/syntax/Physics/DiffusionCG/index.md
@@ -1,0 +1,1 @@
+../../../source/physics/DiffusionCG.md
