### 2023-12-08 07fd60a
Fix error message when use scaling_factors in adjoint system #26282
