### 2023-12-13 #6146 @editorialbot
[REVIEW]: The MOOSE Thermal Hydraulics Module
**Submitting author:** @joshuahansel (<a href="http://orcid.org/0000-0001-6782-5275">Joshua Hansel</a>)
**Repository:** https://github.com/idaholab/moose
**Branch with paper.md** (empty if default branch): master
**Version:** 2024-02-12
