### 2023-12-27 idaholab/moose#26422 @atismer
JOSS Review THM - Link to documentation is not working
The link to the documentation within

https://github.com/idaholab/moose/blob/next/modules/thermal_hydraulics/README.md

does not work.

JOSS Review Thread: https://github.com/openjournals/joss-reviews/issues/6146
