### 2023-12-27 idaholab/moose#26423 @atismer
JOSS Review THM - Docker container not working
I try to run the `docker` container by executing:
```
docker run -ti idaholab/moose:latest /bin/bash -c 'cd test; ./run_tests'
```

This fails with:
```
Building and linking hit...
In file included from /opt/moose/framework/contrib/hit/hitpre311.cpp:1168,
                 from /opt/moose/framework/contrib/hit/hit.cpp:6:
/opt/moose/framework/contrib/hit/parse.h:11:10: fatal error: wasphit/HITInterpreter.h: No such file or directory
   11 | #include "wasphit/HITInterpreter.h"
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~
compilation terminated.
In file included from /opt/moose/framework/contrib/hit/braceexpr.h:3,
                 from /opt/moose/framework/contrib/hit/lex.cc:7:
/opt/moose/framework/contrib/hit/parse.h:11:10: fatal error: wasphit/HITInterpreter.h: No such file or directory
   11 | #include "wasphit/HITInterpreter.h"
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~
compilation terminated.
In file included from /opt/moose/framework/contrib/hit/parse.cc:10:
/opt/moose/framework/contrib/hit/parse.h:11:10: fatal error: wasphit/HITInterpreter.h: No such file or directory
   11 | #include "wasphit/HITInterpreter.h"
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~
compilation terminated.
In file included from /opt/moose/framework/contrib/hit/braceexpr.h:3,
                 from /opt/moose/framework/contrib/hit/braceexpr.cc:1:
/opt/moose/framework/contrib/hit/parse.h:11:10: fatal error: wasphit/HITInterpreter.h: No such file or directory
   11 | #include "wasphit/HITInterpreter.h"
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~
compilation terminated.
In file included from /opt/moose/framework/contrib/hit/braceexpr.h:3,
                 from /opt/moose/framework/contrib/hit/main.cc:10:
/opt/moose/framework/contrib/hit/parse.h:11:10: fatal error: wasphit/HITInterpreter.h: No such file or directory
   11 | #include "wasphit/HITInterpreter.h"
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~
compilation terminated.
make: *** [/opt/moose/framework/moose.mk:138: hit] Error 1
Traceback (most recent call last):
  File "/opt/moose/python/pyhit/__init__.py", line 13, in <module>
    import hit
ModuleNotFoundError: No module named 'hit'

Traceback (most recent call last):
  File "/opt/moose/python/pyhit/__init__.py", line 18, in <module>
    import hit
ModuleNotFoundError: No module named 'hit'

Traceback (most recent call last):
  File "./run_tests", line 6, in <module>
    from TestHarness import TestHarness
  File "/opt/moose/python/TestHarness/__init__.py", line 15, in <module>
    from .TestHarness import TestHarness
  File "/opt/moose/python/TestHarness/TestHarness.py", line 20, in <module>
    from FactorySystem.Factory import Factory
  File "/opt/moose/python/FactorySystem/__init__.py", line 13, in <module>
    from .Parser import Parser
  File "/opt/moose/python/FactorySystem/Parser.py", line 13, in <module>
    import pyhit
  File "/opt/moose/python/pyhit/__init__.py", line 22, in <module>
    import hit
ModuleNotFoundError: No module named 'hit'
```

JOSS Review Thread: https://github.com/openjournals/joss-reviews/issues/6146
