### 2023-12-27 #6146 @atismer
@joshuahansel
I start my review and created the following PRs

- https://github.com/idaholab/moose/issues/26422
- https://github.com/idaholab/moose/issues/26423
