### 2023-12-28 e686919
Update documentation link for THM refs #26422
