### 2024-01-02 idaholab/moose#26422 @joshuahansel
I have confirmed that the new link given in #26426 works, but I'll let @atismer confirm and close this issue.
