### 2024-01-13 #6146 @kaeley55
Upon review of the JOSS COI Policy, I am disclosing that I do have a COI as I work at the same organization and collaborate with the submitting author of this paper.
