### 2024-01-17 e81bdcd
Update docker documentation
refs #26423
