+++ framework/doc/packages_config.yml
@@ -8,3 +8,6 @@ petsc_alt: 3.11.4
 petsc: 3.20.1
 moose_dev: 2023.12.20
 moose_tools: 2023.12.20
+apptainer_rockylinux: 8.8.20230518
+apptainer_mpich: 3.4.3
+apptainer_gcc: 12.2.1
