### 2024-01-29 idaholab/moose#26665 @atismer
JOSS Review THM - Chapter 'Getting Started'
The chapter of [Getting started](https://mooseframework.inl.gov/modules/thermal_hydraulics/getting_started.html) is not really filled with information. Maybe link to another site or skip this page.

JOSS Review Thread: https://github.com/openjournals/joss-reviews/issues/6146
