### 2024-01-29 idaholab/moose#26666 @atismer
JOSS Review THM - Tutorial Step 1 Flow Through a Channel
At the end of [PostProcessors](https://mooseframework.inl.gov/modules/thermal_hydraulics/tutorials/single_phase_flow/step01.html#3481303c-7327-4fea-87eb-be582c74a52b) the sentence references to `p` as temperatire, see:

> The side is specified via the boundary parameter, and both postprocessors
> operate on the temperature variable p.

I guess `p` is the pressure and not the temperature?

Additionally, the bullet point

> `core_p_in` for monitoring core outlet pressure

is not the outlet pressure. This is the inlet pressure, isn't it?

JOSS Review Thread: https://github.com/openjournals/joss-reviews/issues/6146
