### 2024-01-29 idaholab/moose#26669 @atismer
JOSS Review THM - Custom Closure Chapter
The name of the `closures`-parameter within the chapter [Add Custom Closure Correlations](https://mooseframework.inl.gov/modules/thermal_hydraulics/tutorials/single_phase_flow/step06.html#fc5312ad-70f2-465d-92bd-8466d63acc5e) is wrong, see:

> Then the name we gave this closures object (no_closures) is passed to the closures parameter in the hx/pri component.

I think, this should be `none_closures`.
