### 2024-01-29 #6146 @atismer
@joshuahansel
I continued my review and created the following Issues

- https://github.com/idaholab/moose/issues/26669
- https://github.com/idaholab/moose/issues/26666
- https://github.com/idaholab/moose/issues/26665
