### 2024-01-29 5ce8bbf
Fixed typo in custom closures tutorial
Closes #26669
