+++ modules/thermal_hydraulics/doc/content/modules/thermal_hydraulics/tutorials/single_phase_flow/step06.md
@@ -44,7 +44,7 @@ which does not create any of its own Materials:
          block=Closures
          link=False
 
-Then the name we gave this closures object (`no_closures`) is passed to the `closures` parameter in the `hx/pri` component. This will overwrite the `closures` parameter set in the `GlobalParams` block.
+Then the name we gave this closures object (`none_closures`) is passed to the `closures` parameter in the `hx/pri` component. This will overwrite the `closures` parameter set in the `GlobalParams` block.
 
 
 !listing thermal_hydraulics/tutorials/single_phase_flow/06_custom_closures.i
