### 2024-01-29 0bc23d5
Fixed typos in flow channel tutorial
Closes #26666
