### 2024-01-29 10c19bc
Made various improvements to THM documentation
Closes #26665
