+++ modules/thermal_hydraulics/doc/content/modules/thermal_hydraulics/modeling_guide/index.md
@@ -1,6 +1,0 @@
-!config navigation breadcrumbs=False scrollspy=False
-
-# Modeling Guide
-
-- [thermal_hydraulics/modeling_guide/brayton_cycle/brayton_cycle.md]
-- [thermal_hydraulics/modeling_guide/recuperated_brayton_cycle/recuperated_brayton_cycle.md]
