+++ modules/thermal_hydraulics/doc/content/modules/thermal_hydraulics/v_and_v/verification/sod_shock_tube.md
@@ -1,1 +1,0 @@
-!template load file=stubs/validation_case.md.template name=Sod Shock Tube
