+++ modules/thermal_hydraulics/doc/content/modules/thermal_hydraulics/examples/index.md
@@ -1,0 +1,6 @@
+!config navigation breadcrumbs=False scrollspy=False
+
+# Modeling Guide
+
+- [thermal_hydraulics/examples/brayton_cycle/brayton_cycle.md]
+- [thermal_hydraulics/examples/recuperated_brayton_cycle/recuperated_brayton_cycle.md]
