+++ modules/thermal_hydraulics/doc/content/modules/thermal_hydraulics/v_and_v/verification/super_sonic_tube.md
@@ -1,1 +1,0 @@
-!template load file=stubs/validation_case.md.template name=Super Sonic Tube
