+++ modules/thermal_hydraulics/doc/content/modules/thermal_hydraulics/v_and_v/verification/hydrostatic_pressure.md
@@ -1,1 +1,0 @@
-!template load file=stubs/validation_case.md.template name=Hydrostatic Pressure
