### 2024-01-31 #6146 @joshuahansel
Thank you for the review @atismer. I've addressed your issues (and made some other THM home page improvements): https://mooseframework.inl.gov/modules/thermal_hydraulics/index.html
