### 2024-01-31 #6146 @powersrobert
@joshuahansel , I like to start off by apologizing for the time it's taken me to review you submission. Going over everything I do have some questions. Under documentation it asks for a 'statement of need', you clearly provide this in the software paper but the documentation at https://mooseframework.inl.gov/modules/thermal_hydraulics/index.html, does not provide a statement of need. Maybe @kyleniemeyer would be the better person to ask for the review but is the statement of need required for both the paper and the documentation?

Secondly, the community guidelines for those wishing to contribute, report errors, or seek support. Your paper discusses that support will be given "to comply with MOOSE's SQA standards". I know this is a module of MOOSE, but wanted to ask if the community guidelines are the same for this module as they are with MOOSE and should this be stated somewhere?

Lastly, are there other commonly used software packages this module can be compared against?
