### 2024-01-31 #6146 @joshuahansel
@powersrobert

> wanted to ask if the community guidelines are the same for this module as they are with MOOSE and should this be stated somewhere?

The community guidelines are indeed the same; the framework instructions apply to all of the modules as well: https://mooseframework.inl.gov/framework/contributing.html. The "support" we mentioned in the paper was referring to these instructions, along with instructions we give during the pull request reviews if we find the proposed changes are not in compliance. Would you like me to change either the paper or the THM documentation?

> Lastly, are there other commonly used software packages this module can be compared against?

I'm aware of export-controlled nuclear systems codes that provide similar capability. However, I'm not aware of any that provide an extendable framework for systems codes, and I'm not aware of any that are open-source.
