### 2024-02-01 #6146 @powersrobert
Thank you both for your answers, @kyleniemeyer @joshuahansel . To Joshua, this is my first review for JOSS. If it is common practice to assume the community guidelines are the same then I don't see the need to add it. I just wanted to verify before accepting it in the review. As for the other commonly used software, thank you for that knowledge as well.
