### 2024-02-02 #6146 @joshuahansel
@powersrobert Are you waiting on input from @kyleniemeyer to give the final call on

> If it is common practice to assume the community guidelines are the same then I don't see the need to add it
