### 2024-02-02 #6146 @kyleniemeyer
@powersrobert as long as the community guidelines are clear and accessible in the repository, it's not necessary to repeat the information
