### 2024-02-02 #6146 @powersrobert
@kyleniemeyer, I am happy with the contribution. The paper is very well written and the few questions I had I believe were simply my misunderstanding of where they should be located. If there anything specifically you'd like from me to continue forward?
