### 2024-02-12 #6146 @kyleniemeyer
@joshuahansel ok, thanks. Please update the version metadata in the Zenodo archive to match that (currently it just says "v1").
