### 2024-02-12 #6146 @kyleniemeyer
@joshuahansel I made some edits to the paper and citations in https://github.com/idaholab/moose/pull/26794, can you please review?

Also, it appears that the SQA plan citation is to something that is an INL internal document, that cannot be viewed publicly. I don't think that's appropriate to cite; can you remove that or replace it?
