### 2024-02-14 #6146 @kyleniemeyer
@joshuahansel can you take a final look at the [proof](https://github.com/openjournals/joss-papers/blob/eb733997f1a6ba4b0c0d9af2df69b87d9cc9a0c0/joss.06146/10.21105.joss.06146.pdf)?
